Network.AWS.OpsWorks.AssignInstance

Description

Assign a registered instance to a custom layer. You cannot use this action with instances that were created with AWS OpsWorks.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_AssignInstance.html

Request

data AssignInstance

Request constructor

assignInstance

Arguments

:: Text
aiInstanceId

-> AssignInstance

AssignInstance constructor.

The fields accessible through corresponding lenses are:

aiInstanceId :: Text

aiLayerIds :: [Text]

Request lenses

aiInstanceId :: Lens' AssignInstance Text

The instance ID.

aiLayerIds :: Lens' AssignInstance [Text]

The layer ID, which must correspond to a custom layer. You cannot assign a registered instance to a built-in layer.

Response

data AssignInstanceResponse

Response constructor

assignInstanceResponse :: AssignInstanceResponse

AssignInstanceResponse constructor.