Network.AWS.OpsWorks.AssignVolume

Description

Assigns one of the stack's registered Amazon EBS volumes to a specified instance. The volume must first be registered with the stack by calling RegisterVolume. After you register the volume, you must call UpdateVolume to specify a mount point before calling AssignVolume. For more information, see ResourceManagement.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_AssignVolume.html

Request

data AssignVolume

Request constructor

assignVolume

Arguments

:: Text
avVolumeId

-> AssignVolume

AssignVolume constructor.

The fields accessible through corresponding lenses are:

avInstanceId :: Maybe Text

avVolumeId :: Text

Request lenses

avInstanceId :: Lens' AssignVolume (Maybe Text)

The instance ID.

avVolumeId :: Lens' AssignVolume Text

The volume ID.

Response

data AssignVolumeResponse

Response constructor

assignVolumeResponse :: AssignVolumeResponse

AssignVolumeResponse constructor.