Network.AWS.OpsWorks.CloneStack

Description

Creates a clone of a specified stack. For more information, see Clone a Stack.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_CloneStack.html

Request

data CloneStack

Request constructor

cloneStack

Arguments

:: Text
cs1SourceStackId

-> Text
cs1ServiceRoleArn

-> CloneStack

CloneStack constructor.

The fields accessible through corresponding lenses are:

cs1Attributes :: HashMap StackAttributesKeys Text

cs1ChefConfiguration :: Maybe ChefConfiguration

cs1CloneAppIds :: [Text]

cs1ClonePermissions :: Maybe Bool

cs1ConfigurationManager :: Maybe StackConfigurationManager

cs1CustomCookbooksSource :: Maybe Source

cs1CustomJson :: Maybe Text

cs1DefaultAvailabilityZone :: Maybe Text

cs1DefaultInstanceProfileArn :: Maybe Text

cs1DefaultOs :: Maybe Text

cs1DefaultRootDeviceType :: Maybe RootDeviceType

cs1DefaultSshKeyName :: Maybe Text

cs1DefaultSubnetId :: Maybe Text

cs1HostnameTheme :: Maybe Text

cs1Name :: Maybe Text

cs1Region :: Maybe Text

cs1ServiceRoleArn :: Text

cs1SourceStackId :: Text

cs1UseCustomCookbooks :: Maybe Bool

cs1UseOpsworksSecurityGroups :: Maybe Bool

cs1VpcId :: Maybe Text

Request lenses

cs1Attributes :: Lens' CloneStack (HashMap StackAttributesKeys Text)

A list of stack attributes and values as key/value pairs to be added to the cloned stack.

cs1ChefConfiguration :: Lens' CloneStack (Maybe ChefConfiguration)

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version on Chef 11.10 stacks. For more information, see Create aNew Stack.

cs1CloneAppIds :: Lens' CloneStack [Text]

A list of source stack app IDs to be included in the cloned stack.

cs1ClonePermissions :: Lens' CloneStack (Maybe Bool)

Whether to clone the source stack's permissions.

cs1ConfigurationManager :: Lens' CloneStack (Maybe StackConfigurationManager)

The configuration manager. When you clone a stack we recommend that you use the configuration manager to specify the Chef version, 0.9, 11.4, or 11.10. The default value is currently 11.4.

cs1CustomCookbooksSource :: Lens' CloneStack (Maybe Source)

cs1CustomJson :: Lens' CloneStack (Maybe Text)

A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format and must escape characters such as '"'.:

'"{"key1": "value1", "key2": "value2",...}"'

For more information on custom JSON, see Use Custom JSON to Modify the StackConfiguration JSON

cs1DefaultAvailabilityZone :: Lens' CloneStack (Maybe Text)

The cloned stack's default Availability Zone, which must be in the specified region. For more information, see Regions and Endpoints. If you also specify a value for DefaultSubnetId, the subnet must be in the same zone. For more information, see the VpcId parameter description.

cs1DefaultInstanceProfileArn :: Lens' CloneStack (Maybe Text)

The ARN of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

cs1DefaultOs :: Lens' CloneStack (Maybe Text)

The stack's operating system, which must be set to one of the following.

Standard operating systems: an Amazon Linux version such as 'Amazon Linux2014.09', 'Ubuntu 12.04 LTS', or 'Ubuntu 14.04 LTS'. Custom AMIs: Custom. You specify the custom AMI you want to use when you create instances. The default option is the current Amazon Linux version.

cs1DefaultRootDeviceType :: Lens' CloneStack (Maybe RootDeviceType)

The default root device type. This value is used by default for all instances in the cloned stack, but you can override it when you create an instance. For more information, see Storage for the Root Device.

cs1DefaultSshKeyName :: Lens' CloneStack (Maybe Text)

A default Amazon EC2 key pair name. The default value is none. If you specify a key pair name, AWS OpsWorks installs the public key on the instance and you can use the private key with an SSH client to log in to the instance. For more information, see Using SSH to Communicate with an Instance and Managing SSH Access. You can override this setting by specifying a different key pair, or no key pair, when you create an instance.

cs1DefaultSubnetId :: Lens' CloneStack (Maybe Text)

The stack's default VPC subnet ID. This parameter is required if you specify a value for the VpcId parameter. All instances are launched into this subnet unless you specify otherwise when you create the instance. If you also specify a value for DefaultAvailabilityZone, the subnet must be in that zone. For information on default values and when this parameter is required, see the VpcId parameter description.

cs1HostnameTheme :: Lens' CloneStack (Maybe Text)

The stack's host name theme, with spaces are replaced by underscores. The theme is used to generate host names for the stack's instances. By default, HostnameTheme is set to Layer_Dependent, which creates host names by appending integers to the layer's short name. The other themes are:

Baked_Goods Clouds Europe_Cities Fruits Greek_Deities Legendary_creatures_from_Japan Planets_and_Moons Roman_Deities Scottish_Islands US_Cities Wild_Cats To obtain a generated host name, call GetHostNameSuggestion, which returns a host name based on the current theme.

cs1Name :: Lens' CloneStack (Maybe Text)

The cloned stack name.

cs1Region :: Lens' CloneStack (Maybe Text)

The cloned stack AWS region, such as "us-east-1". For more information about AWS regions, see Regions and Endpoints.

cs1ServiceRoleArn :: Lens' CloneStack Text

The stack AWS Identity and Access Management (IAM) role, which allows AWS OpsWorks to work with AWS resources on your behalf. You must set this parameter to the Amazon Resource Name (ARN) for an existing IAM role. If you create a stack by using the AWS OpsWorks console, it creates the role for you. You can obtain an existing stack's IAM ARN programmatically by calling DescribePermissions. For more information about IAM ARNs, see Using Identifiers.

You must set this parameter to a valid service role ARN or the action will fail; there is no default value. You can specify the source stack's service role ARN, if you prefer, but you must do so explicitly.

cs1SourceStackId :: Lens' CloneStack Text

The source stack ID.

cs1UseCustomCookbooks :: Lens' CloneStack (Maybe Bool)

Whether to use custom cookbooks.

cs1UseOpsworksSecurityGroups :: Lens' CloneStack (Maybe Bool)

Whether to associate the AWS OpsWorks built-in security groups with the stack's layers.

AWS OpsWorks provides a standard set of built-in security groups, one for each layer, which are associated with layers by default. With UseOpsworksSecurityGroups you can instead provide your own custom security groups. UseOpsworksSecurityGroups has the following settings:

True - AWS OpsWorks automatically associates the appropriate built-in security group with each layer (default setting). You can associate additional security groups with a layer after you create it but you cannot delete the built-in security group. False - AWS OpsWorks does not associate built-in security groups with layers. You must create appropriate EC2 security groups and associate a security group with each layer that you create. However, you can still manually associate a built-in security group with a layer on creation; custom security groups are required only for those layers that need custom settings. For more information, see Create a NewStack.

cs1VpcId :: Lens' CloneStack (Maybe Text)

The ID of the VPC that the cloned stack is to be launched into. It must be in the specified region. All instances are launched into this VPC, and you cannot change the ID later.

If your account supports EC2 Classic, the default value is no VPC. If your account does not support EC2 Classic, the default value is the default VPC for the specified region. If the VPC ID corresponds to a default VPC and you have specified either the DefaultAvailabilityZone or the DefaultSubnetId parameter only, AWS OpsWorks infers the value of the other parameter. If you specify neither parameter, AWS OpsWorks sets these parameters to the first valid Availability Zone for the specified region and the corresponding default VPC subnet ID, respectively.

If you specify a nondefault VPC ID, note the following:

It must belong to a VPC in your account that is in the specified region. You must specify a value for DefaultSubnetId. For more information on how to use AWS OpsWorks with a VPC, see Running a Stack in a VPC. For more information on default VPC and EC2 Classic, see Supported Platforms.

Response

data CloneStackResponse

Response constructor

cloneStackResponse :: CloneStackResponse

CloneStackResponse constructor.

The fields accessible through corresponding lenses are:

csrStackId :: Maybe Text

Response lenses

csrStackId :: Lens' CloneStackResponse (Maybe Text)

The cloned stack ID.