Network.AWS.OpsWorks.CreateApp

Description

Creates an app for a specified stack. For more information, see Creating Apps.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_CreateApp.html

Request

data CreateApp

Request constructor

createApp

Arguments

:: Text
caStackId

-> Text
caName

-> AppType
caType

-> CreateApp

CreateApp constructor.

The fields accessible through corresponding lenses are:

caAppSource :: Maybe Source

caAttributes :: HashMap AppAttributesKeys Text

caDataSources :: [DataSource]

caDescription :: Maybe Text

caDomains :: [Text]

caEnableSsl :: Maybe Bool

caName :: Text

caShortname :: Maybe Text

caSslConfiguration :: Maybe SslConfiguration

caStackId :: Text

caType :: AppType

Request lenses

caAppSource :: Lens' CreateApp (Maybe Source)

A Source object that specifies the app repository.

caAttributes :: Lens' CreateApp (HashMap AppAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

caDataSources :: Lens' CreateApp [DataSource]

The app's data source.

caDescription :: Lens' CreateApp (Maybe Text)

A description of the app.

caDomains :: Lens' CreateApp [Text]

The app virtual host settings, with multiple domains separated by commas. For example: ''www.example.com, example.com''

caEnableSsl :: Lens' CreateApp (Maybe Bool)

Whether to enable SSL for the app.

caName :: Lens' CreateApp Text

The app name.

caShortname :: Lens' CreateApp (Maybe Text)

The app's short name.

caSslConfiguration :: Lens' CreateApp (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

caStackId :: Lens' CreateApp Text

The stack ID.

caType :: Lens' CreateApp AppType

The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks deploys an application to those instances that are members of the corresponding layer.

Response

data CreateAppResponse

Response constructor

createAppResponse :: CreateAppResponse

CreateAppResponse constructor.

The fields accessible through corresponding lenses are:

carAppId :: Maybe Text

Response lenses

carAppId :: Lens' CreateAppResponse (Maybe Text)

The app ID.