Network.AWS.OpsWorks.CreateDeployment

Description

Runs deployment or stack commands. For more information, see Deploying Apps and Run Stack Commands .

Required Permissions : To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

createDeployment

Arguments

:: Text
cdStackId

-> DeploymentCommand
cdCommand

-> CreateDeployment

Creates a value of CreateDeployment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdCustomJSON - A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes .

cdAppId - The app ID. This parameter is required for app deployments, but not for other deployment commands.

cdInstanceIds - The instance IDs for the deployment targets.

cdLayerIds - The layer IDs for the deployment targets.

cdComment - A user-defined comment.

cdStackId - The stack ID.

cdCommand - A DeploymentCommand object that specifies the deployment command and any associated arguments.

data CreateDeployment

See: createDeployment smart constructor.

Request Lenses

cdCustomJSON :: Lens' CreateDeployment (Maybe Text)

A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes .

cdAppId :: Lens' CreateDeployment (Maybe Text)

The app ID. This parameter is required for app deployments, but not for other deployment commands.

cdInstanceIds :: Lens' CreateDeployment [Text]

The instance IDs for the deployment targets.

cdLayerIds :: Lens' CreateDeployment [Text]

The layer IDs for the deployment targets.

cdComment :: Lens' CreateDeployment (Maybe Text)

A user-defined comment.

cdStackId :: Lens' CreateDeployment Text

The stack ID.

cdCommand :: Lens' CreateDeployment DeploymentCommand

A DeploymentCommand object that specifies the deployment command and any associated arguments.

Destructuring the Response

createDeploymentResponse

Arguments

:: Int
cdrsResponseStatus

-> CreateDeploymentResponse

Creates a value of CreateDeploymentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdrsDeploymentId - The deployment ID, which can be used with other requests to identify the deployment.

cdrsResponseStatus - -- | The response status code.

data CreateDeploymentResponse

Contains the response to a CreateDeployment request.

See: createDeploymentResponse smart constructor.

Response Lenses

cdrsDeploymentId :: Lens' CreateDeploymentResponse (Maybe Text)

The deployment ID, which can be used with other requests to identify the deployment.

cdrsResponseStatus :: Lens' CreateDeploymentResponse Int

- | The response status code.