Network.AWS.OpsWorks.CreateInstance

Description

Creates an instance in a specified stack. For more information, see Adding anInstance to a Layer.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_CreateInstance.html

Request

data CreateInstance

Request constructor

createInstance

Arguments

:: Text
ciStackId

-> Text
ciInstanceType

-> CreateInstance

CreateInstance constructor.

The fields accessible through corresponding lenses are:

ciAmiId :: Maybe Text

ciArchitecture :: Maybe Architecture

ciAutoScalingType :: Maybe AutoScalingType

ciAvailabilityZone :: Maybe Text

ciBlockDeviceMappings :: [BlockDeviceMapping]

ciEbsOptimized :: Maybe Bool

ciHostname :: Maybe Text

ciInstallUpdatesOnBoot :: Maybe Bool

ciInstanceType :: Text

ciLayerIds :: [Text]

ciOs :: Maybe Text

ciRootDeviceType :: Maybe RootDeviceType

ciSshKeyName :: Maybe Text

ciStackId :: Text

ciSubnetId :: Maybe Text

ciVirtualizationType :: Maybe Text

Request lenses

ciAmiId :: Lens' CreateInstance (Maybe Text)

A custom AMI ID to be used to create the instance. The AMI should be based on one of the standard AWS OpsWorks AMIs: Amazon Linux, Ubuntu 12.04 LTS, or Ubuntu 14.04 LTS. For more information, see Instances.

If you specify a custom AMI, you must set Os to Custom.

ciArchitecture :: Lens' CreateInstance (Maybe Architecture)

The instance architecture. The default option is x86_64. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families andTypes.

ciAutoScalingType :: Lens' CreateInstance (Maybe AutoScalingType)

For load-based or time-based instances, the type.

ciAvailabilityZone :: Lens' CreateInstance (Maybe Text)

The instance Availability Zone. For more information, see Regions andEndpoints.

ciBlockDeviceMappings :: Lens' CreateInstance [BlockDeviceMapping]

An array of BlockDeviceMapping objects that specify the instance's block devices. For more information, see Block Device Mapping.

ciEbsOptimized :: Lens' CreateInstance (Maybe Bool)

Whether to create an Amazon EBS-optimized instance.

ciHostname :: Lens' CreateInstance (Maybe Text)

The instance host name.

ciInstallUpdatesOnBoot :: Lens' CreateInstance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or 'apt-get' (Ubuntu) on the instances.

We strongly recommend using the default value of true to ensure that your instances have the latest security updates.

ciInstanceType :: Lens' CreateInstance Text

The instance type. AWS OpsWorks supports all instance types except Cluster Compute, Cluster GPU, and High Memory Cluster. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

ciLayerIds :: Lens' CreateInstance [Text]

An array that contains the instance layer IDs.

ciOs :: Lens' CreateInstance (Maybe Text)

The instance's operating system, which must be set to one of the following.

Standard operating systems: an Amazon Linux version such as 'Amazon Linux2014.09', 'Ubuntu 12.04 LTS', or 'Ubuntu 14.04 LTS'. Custom AMIs: Custom The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. For more information on the standard operating systems, see Operating SystemsFor more information on how to use custom AMIs with OpsWorks, see Using Custom AMIs.

ciRootDeviceType :: Lens' CreateInstance (Maybe RootDeviceType)

The instance root device type. For more information, see Storage for the RootDevice.

ciSshKeyName :: Lens' CreateInstance (Maybe Text)

The instance's Amazon EC2 key pair name.

ciStackId :: Lens' CreateInstance Text

The stack ID.

ciSubnetId :: Lens' CreateInstance (Maybe Text)

The ID of the instance's subnet. If the stack is running in a VPC, you can use this parameter to override the stack's default subnet ID value and direct AWS OpsWorks to launch the instance in a different subnet.

ciVirtualizationType :: Lens' CreateInstance (Maybe Text)

The instance's virtualization type, paravirtual or hvm.

Response

data CreateInstanceResponse

Response constructor

createInstanceResponse :: CreateInstanceResponse

CreateInstanceResponse constructor.

The fields accessible through corresponding lenses are:

cirInstanceId :: Maybe Text

Response lenses

cirInstanceId :: Lens' CreateInstanceResponse (Maybe Text)

The instance ID.