Network.AWS.OpsWorks.CreateUserProfile

Description

Creates a new user profile.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_CreateUserProfile.html

Request

data CreateUserProfile

Request constructor

createUserProfile

Arguments

:: Text
cupIamUserArn

-> CreateUserProfile

CreateUserProfile constructor.

The fields accessible through corresponding lenses are:

cupAllowSelfManagement :: Maybe Bool

cupIamUserArn :: Text

cupSshPublicKey :: Maybe Text

cupSshUsername :: Maybe Text

Request lenses

cupAllowSelfManagement :: Lens' CreateUserProfile (Maybe Bool)

Whether users can specify their own SSH public key through the My Settings page. For more information, see Setting an IAM User's Public SSH Key.

cupIamUserArn :: Lens' CreateUserProfile Text

The user's IAM ARN.

cupSshPublicKey :: Lens' CreateUserProfile (Maybe Text)

The user's public SSH key.

cupSshUsername :: Lens' CreateUserProfile (Maybe Text)

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], -, and '_'. If the specified name includes other punctuation marks, AWS OpsWorks removes them. For example, 'my.name' will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks generates one from the IAM user name.

Response

data CreateUserProfileResponse

Response constructor

createUserProfileResponse :: CreateUserProfileResponse

CreateUserProfileResponse constructor.

The fields accessible through corresponding lenses are:

cuprIamUserArn :: Maybe Text

Response lenses

cuprIamUserArn :: Lens' CreateUserProfileResponse (Maybe Text)

The user's IAM ARN.