Network.AWS.OpsWorks.DeregisterEcsCluster

Description

Deregisters a specified Amazon ECS cluster from a stack. For more information, see Resource Management .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html .

Creating a Request

deregisterEcsCluster

Arguments

:: Text
decEcsClusterARN

-> DeregisterEcsCluster

Creates a value of DeregisterEcsCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

decEcsClusterARN - The cluster's ARN.

data DeregisterEcsCluster

See: deregisterEcsCluster smart constructor.

Request Lenses

decEcsClusterARN :: Lens' DeregisterEcsCluster Text

The cluster's ARN.

Destructuring the Response

deregisterEcsClusterResponse :: DeregisterEcsClusterResponse

Creates a value of DeregisterEcsClusterResponse with the minimum fields required to make a request.

data DeregisterEcsClusterResponse

See: deregisterEcsClusterResponse smart constructor.