Network.AWS.OpsWorks.DeregisterVolume

Description

Deregisters an Amazon EBS volume. The volume can then be registered by another stack. For more information, see Resource Management .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

deregisterVolume

Arguments

:: Text
dvVolumeId

-> DeregisterVolume

Creates a value of DeregisterVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvVolumeId - The AWS OpsWorks Stacks volume ID, which is the GUID that AWS OpsWorks Stacks assigned to the instance when you registered the volume with the stack, not the Amazon EC2 volume ID.

data DeregisterVolume

See: deregisterVolume smart constructor.

Request Lenses

dvVolumeId :: Lens' DeregisterVolume Text

The AWS OpsWorks Stacks volume ID, which is the GUID that AWS OpsWorks Stacks assigned to the instance when you registered the volume with the stack, not the Amazon EC2 volume ID.

Destructuring the Response

deregisterVolumeResponse :: DeregisterVolumeResponse

Creates a value of DeregisterVolumeResponse with the minimum fields required to make a request.

data DeregisterVolumeResponse

See: deregisterVolumeResponse smart constructor.