Network.AWS.OpsWorks.DescribeApps

Description

Requests a description of a specified set of apps.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeApps.html

Request

data DescribeApps

Request constructor

describeApps :: DescribeApps

DescribeApps constructor.

The fields accessible through corresponding lenses are:

daAppIds :: [Text]

daStackId :: Maybe Text

Request lenses

daAppIds :: Lens' DescribeApps [Text]

An array of app IDs for the apps to be described. If you use this parameter, DescribeApps returns a description of the specified apps. Otherwise, it returns a description of every app.

daStackId :: Lens' DescribeApps (Maybe Text)

The app stack ID. If you use this parameter, DescribeApps returns a description of the apps in the specified stack.

Response

data DescribeAppsResponse

Response constructor

describeAppsResponse :: DescribeAppsResponse

DescribeAppsResponse constructor.

The fields accessible through corresponding lenses are:

darApps :: [App]

Response lenses

darApps :: Lens' DescribeAppsResponse [App]

An array of App objects that describe the specified apps.