Network.AWS.OpsWorks.DescribeCommands

Description

Describes the results of specified commands.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeCommands.html

Request

data DescribeCommands

Request constructor

describeCommands :: DescribeCommands

DescribeCommands constructor.

The fields accessible through corresponding lenses are:

dcCommandIds :: [Text]

dcDeploymentId :: Maybe Text

dcInstanceId :: Maybe Text

Request lenses

dcCommandIds :: Lens' DescribeCommands [Text]

An array of command IDs. If you include this parameter, DescribeCommands returns a description of the specified commands. Otherwise, it returns a description of every command.

dcDeploymentId :: Lens' DescribeCommands (Maybe Text)

The deployment ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified deployment.

dcInstanceId :: Lens' DescribeCommands (Maybe Text)

The instance ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified instance.

Response

data DescribeCommandsResponse

Response constructor

describeCommandsResponse :: DescribeCommandsResponse

DescribeCommandsResponse constructor.

The fields accessible through corresponding lenses are:

dcrCommands :: [Command]

Response lenses

dcrCommands :: Lens' DescribeCommandsResponse [Command]

An array of Command objects that describe each of the specified commands.