Network.AWS.OpsWorks.DescribeDeployments

Description

Requests a description of a specified set of deployments.

Required Permissions : To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

describeDeployments :: DescribeDeployments

Creates a value of DescribeDeployments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddAppId - The app ID. If you include this parameter, DescribeDeployments returns a description of the commands associated with the specified app.

ddDeploymentIds - An array of deployment IDs to be described. If you include this parameter, DescribeDeployments returns a description of the specified deployments. Otherwise, it returns a description of every deployment.

ddStackId - The stack ID. If you include this parameter, DescribeDeployments returns a description of the commands associated with the specified stack.

data DescribeDeployments

See: describeDeployments smart constructor.

Request Lenses

ddAppId :: Lens' DescribeDeployments (Maybe Text)

The app ID. If you include this parameter, DescribeDeployments returns a description of the commands associated with the specified app.

ddDeploymentIds :: Lens' DescribeDeployments [Text]

An array of deployment IDs to be described. If you include this parameter, DescribeDeployments returns a description of the specified deployments. Otherwise, it returns a description of every deployment.

ddStackId :: Lens' DescribeDeployments (Maybe Text)

The stack ID. If you include this parameter, DescribeDeployments returns a description of the commands associated with the specified stack.

Destructuring the Response

describeDeploymentsResponse

Arguments

:: Int
ddrsResponseStatus

-> DescribeDeploymentsResponse

Creates a value of DescribeDeploymentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddrsDeployments - An array of Deployment objects that describe the deployments.

ddrsResponseStatus - -- | The response status code.

data DescribeDeploymentsResponse

Contains the response to a DescribeDeployments request.

See: describeDeploymentsResponse smart constructor.

Response Lenses

ddrsDeployments :: Lens' DescribeDeploymentsResponse [Deployment]

An array of Deployment objects that describe the deployments.

ddrsResponseStatus :: Lens' DescribeDeploymentsResponse Int

- | The response status code.