Network.AWS.OpsWorks.DescribeElasticIPs

Description

Describes Elastic IP addresses .

Required Permissions : To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

describeElasticIPs :: DescribeElasticIPs

Creates a value of DescribeElasticIPs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deiInstanceId - The instance ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses associated with the specified instance.

deiIPs - An array of Elastic IP addresses to be described. If you include this parameter, DescribeElasticIps returns a description of the specified Elastic IP addresses. Otherwise, it returns a description of every Elastic IP address.

deiStackId - A stack ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses that are registered with the specified stack.

data DescribeElasticIPs

See: describeElasticIPs smart constructor.

Request Lenses

deiInstanceId :: Lens' DescribeElasticIPs (Maybe Text)

The instance ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses associated with the specified instance.

deiIPs :: Lens' DescribeElasticIPs [Text]

An array of Elastic IP addresses to be described. If you include this parameter, DescribeElasticIps returns a description of the specified Elastic IP addresses. Otherwise, it returns a description of every Elastic IP address.

deiStackId :: Lens' DescribeElasticIPs (Maybe Text)

A stack ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses that are registered with the specified stack.

Destructuring the Response

describeElasticIPsResponse

Arguments

:: Int
deirsResponseStatus

-> DescribeElasticIPsResponse

Creates a value of DescribeElasticIPsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deirsElasticIPs - An ElasticIps object that describes the specified Elastic IP addresses.

deirsResponseStatus - -- | The response status code.

data DescribeElasticIPsResponse

Contains the response to a DescribeElasticIps request.

See: describeElasticIPsResponse smart constructor.

Response Lenses

deirsElasticIPs :: Lens' DescribeElasticIPsResponse [ElasticIP]

An ElasticIps object that describes the specified Elastic IP addresses.

deirsResponseStatus :: Lens' DescribeElasticIPsResponse Int

- | The response status code.