Network.AWS.OpsWorks.DescribeElasticIps

Description

Describes Elastic IP addresses.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeElasticIps.html

Request

data DescribeElasticIps

Request constructor

describeElasticIps :: DescribeElasticIps

DescribeElasticIps constructor.

The fields accessible through corresponding lenses are:

deiInstanceId :: Maybe Text

deiIps :: [Text]

deiStackId :: Maybe Text

Request lenses

deiInstanceId :: Lens' DescribeElasticIps (Maybe Text)

The instance ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses associated with the specified instance.

deiIps :: Lens' DescribeElasticIps [Text]

An array of Elastic IP addresses to be described. If you include this parameter, DescribeElasticIps returns a description of the specified Elastic IP addresses. Otherwise, it returns a description of every Elastic IP address.

deiStackId :: Lens' DescribeElasticIps (Maybe Text)

A stack ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses that are registered with the specified stack.

Response

data DescribeElasticIpsResponse

Response constructor

describeElasticIpsResponse :: DescribeElasticIpsResponse

DescribeElasticIpsResponse constructor.

The fields accessible through corresponding lenses are:

deirElasticIps :: [ElasticIp]

Response lenses

deirElasticIps :: Lens' DescribeElasticIpsResponse [ElasticIp]

An ElasticIps object that describes the specified Elastic IP addresses.