Network.AWS.OpsWorks.DescribeLayers

Description

Requests a description of one or more layers in a specified stack. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeLayers.html

Request

data DescribeLayers

Request constructor

describeLayers :: DescribeLayers

DescribeLayers constructor.

The fields accessible through corresponding lenses are:

dlLayerIds :: [Text]

dlStackId :: Maybe Text

Request lenses

dlLayerIds :: Lens' DescribeLayers [Text]

An array of layer IDs that specify the layers to be described. If you omit this parameter, DescribeLayers returns a description of every layer in the specified stack.

dlStackId :: Lens' DescribeLayers (Maybe Text)

The stack ID.

Response

data DescribeLayersResponse

Response constructor

describeLayersResponse :: DescribeLayersResponse

DescribeLayersResponse constructor.

The fields accessible through corresponding lenses are:

dlrLayers :: [Layer]

Response lenses

dlrLayers :: Lens' DescribeLayersResponse [Layer]

An array of Layer objects that describe the layers.