Network.AWS.OpsWorks.DescribeLoadBasedAutoScaling

Description

Describes load-based auto scaling configurations for specified layers. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data DescribeLoadBasedAutoScaling

Request constructor

describeLoadBasedAutoScaling

Arguments

:: [Text]
dlbasLayerIds

-> DescribeLoadBasedAutoScaling

Smart constructor for the minimum required parameters to construct a valid DescribeLoadBasedAutoScaling request.

The fields accessible through corresponding lenses are:

LayerIds :: [Text]

Request lenses

dlbasLayerIds :: Lens' DescribeLoadBasedAutoScaling [Text]

An array of layer IDs.

Response

data DescribeLoadBasedAutoScalingResponse

Contains the response to a DescribeLoadBasedAutoScaling request.

Response constructor

describeLoadBasedAutoScalingResponse :: DescribeLoadBasedAutoScalingResponse

Smart constructor for the minimum required parameters to construct a valid DescribeLoadBasedAutoScalingResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

LoadBasedAutoScalingConfigurations :: [LoadBasedAutoScalingConfiguration]

Response lenses

dlbasrLoadBasedAutoScalingConfigurations :: Lens' DescribeLoadBasedAutoScalingResponse [LoadBasedAutoScalingConfiguration]

An array of LoadBasedAutoScalingConfiguration objects that describe each layer's configuration.