Network.AWS.OpsWorks.DescribeMyUserProfile

Description

Describes a user's SSH information.

Required Permissions: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeMyUserProfile.html

Request

data DescribeMyUserProfile

Request constructor

describeMyUserProfile :: DescribeMyUserProfile

DescribeMyUserProfile constructor.

Response

data DescribeMyUserProfileResponse

Response constructor

describeMyUserProfileResponse :: DescribeMyUserProfileResponse

DescribeMyUserProfileResponse constructor.

The fields accessible through corresponding lenses are:

dmuprUserProfile :: Maybe SelfUserProfile

Response lenses

dmuprUserProfile :: Lens' DescribeMyUserProfileResponse (Maybe SelfUserProfile)

A UserProfile object that describes the user's SSH information.