Network.AWS.OpsWorks.DescribePermissions

Description

Describes the permissions for a specified stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribePermissions.html

Request

data DescribePermissions

Request constructor

describePermissions :: DescribePermissions

DescribePermissions constructor.

The fields accessible through corresponding lenses are:

dpIamUserArn :: Maybe Text

dpStackId :: Maybe Text

Request lenses

dpIamUserArn :: Lens' DescribePermissions (Maybe Text)

The user's IAM ARN. For more information about IAM ARNs, see Using Identifiers .

dpStackId :: Lens' DescribePermissions (Maybe Text)

The stack ID.

Response

data DescribePermissionsResponse

Response constructor

describePermissionsResponse :: DescribePermissionsResponse

DescribePermissionsResponse constructor.

The fields accessible through corresponding lenses are:

dprPermissions :: [Permission]

Response lenses

dprPermissions :: Lens' DescribePermissionsResponse [Permission]

An array of Permission objects that describe the stack permissions.

If the request object contains only a stack ID, the array contains a Permission object with permissions for each of the stack IAM ARNs. If the request object contains only an IAM ARN, the array contains a Permission object with permissions for each of the user's stack IDs. If the request contains a stack ID and an IAM ARN, the array contains a single Permission object with permissions for the specified stack and IAM ARN.