Network.AWS.OpsWorks.DescribeRDSDBInstances

Description

Describes Amazon RDS instances.

Required Permissions : To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

This call accepts only one resource-identifying parameter.

Creating a Request

describeRDSDBInstances

Arguments

:: Text
drdiStackId

-> DescribeRDSDBInstances

Creates a value of DescribeRDSDBInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdiRDSDBInstanceARNs - An array containing the ARNs of the instances to be described.

drdiStackId - The stack ID that the instances are registered with. The operation returns descriptions of all registered Amazon RDS instances.

data DescribeRDSDBInstances

See: describeRDSDBInstances smart constructor.

Request Lenses

drdiRDSDBInstanceARNs :: Lens' DescribeRDSDBInstances [Text]

An array containing the ARNs of the instances to be described.

drdiStackId :: Lens' DescribeRDSDBInstances Text

The stack ID that the instances are registered with. The operation returns descriptions of all registered Amazon RDS instances.

Destructuring the Response

describeRDSDBInstancesResponse

Arguments

:: Int
drdirsResponseStatus

-> DescribeRDSDBInstancesResponse

Creates a value of DescribeRDSDBInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdirsRDSDBInstances - An a array of RdsDbInstance objects that describe the instances.

drdirsResponseStatus - -- | The response status code.

data DescribeRDSDBInstancesResponse

Contains the response to a DescribeRdsDbInstances request.

See: describeRDSDBInstancesResponse smart constructor.

Response Lenses

drdirsRDSDBInstances :: Lens' DescribeRDSDBInstancesResponse [RDSDBInstance]

An a array of RdsDbInstance objects that describe the instances.

drdirsResponseStatus :: Lens' DescribeRDSDBInstancesResponse Int

- | The response status code.