Network.AWS.OpsWorks.DescribeRdsDbInstances

Description

Describes Amazon RDS instances.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeRdsDbInstances.html

Request

data DescribeRdsDbInstances

Request constructor

describeRdsDbInstances

Arguments

:: Text
drdiStackId

-> DescribeRdsDbInstances

DescribeRdsDbInstances constructor.

The fields accessible through corresponding lenses are:

drdiRdsDbInstanceArns :: [Text]

drdiStackId :: Text

Request lenses

drdiRdsDbInstanceArns :: Lens' DescribeRdsDbInstances [Text]

An array containing the ARNs of the instances to be described.

drdiStackId :: Lens' DescribeRdsDbInstances Text

The stack ID that the instances are registered with. The operation returns descriptions of all registered Amazon RDS instances.

Response

data DescribeRdsDbInstancesResponse

Response constructor

describeRdsDbInstancesResponse :: DescribeRdsDbInstancesResponse

DescribeRdsDbInstancesResponse constructor.

The fields accessible through corresponding lenses are:

drdirRdsDbInstances :: [RdsDbInstance]

Response lenses

drdirRdsDbInstances :: Lens' DescribeRdsDbInstancesResponse [RdsDbInstance]

An a array of RdsDbInstance objects that describe the instances.