Network.AWS.OpsWorks.DescribeServiceErrors

Description

Describes AWS OpsWorks service errors.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeServiceErrors.html

Request

data DescribeServiceErrors

Request constructor

describeServiceErrors :: DescribeServiceErrors

DescribeServiceErrors constructor.

The fields accessible through corresponding lenses are:

dseInstanceId :: Maybe Text

dseServiceErrorIds :: [Text]

dseStackId :: Maybe Text

Request lenses

dseInstanceId :: Lens' DescribeServiceErrors (Maybe Text)

The instance ID. If you use this parameter, DescribeServiceErrors returns descriptions of the errors associated with the specified instance.

dseServiceErrorIds :: Lens' DescribeServiceErrors [Text]

An array of service error IDs. If you use this parameter, DescribeServiceErrors returns descriptions of the specified errors. Otherwise, it returns a description of every error.

dseStackId :: Lens' DescribeServiceErrors (Maybe Text)

The stack ID. If you use this parameter, DescribeServiceErrors returns descriptions of the errors associated with the specified stack.

Response

data DescribeServiceErrorsResponse

Response constructor

describeServiceErrorsResponse :: DescribeServiceErrorsResponse

DescribeServiceErrorsResponse constructor.

The fields accessible through corresponding lenses are:

dserServiceErrors :: [ServiceError']

Response lenses

dserServiceErrors :: Lens' DescribeServiceErrorsResponse [ServiceError']

An array of ServiceError objects that describe the specified service errors.