Network.AWS.OpsWorks.DescribeStackProvisioningParameters

Description

Requests a description of a stack's provisioning parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeStackProvisioningParameters.html

Request

data DescribeStackProvisioningParameters

Request constructor

describeStackProvisioningParameters

Arguments

:: Text
dsppStackId

-> DescribeStackProvisioningParameters

DescribeStackProvisioningParameters constructor.

The fields accessible through corresponding lenses are:

dsppStackId :: Text

Request lenses

dsppStackId :: Lens' DescribeStackProvisioningParameters Text

The stack ID

Response

data DescribeStackProvisioningParametersResponse

Response constructor

describeStackProvisioningParametersResponse :: DescribeStackProvisioningParametersResponse

DescribeStackProvisioningParametersResponse constructor.

The fields accessible through corresponding lenses are:

dspprAgentInstallerUrl :: Maybe Text

dspprParameters :: HashMap Text Text

Response lenses

dspprAgentInstallerUrl :: Lens' DescribeStackProvisioningParametersResponse (Maybe Text)

The AWS OpsWorks agent installer's URL.

dspprParameters :: Lens' DescribeStackProvisioningParametersResponse (HashMap Text Text)

An embedded object that contains the provisioning parameters.