Network.AWS.OpsWorks.DescribeStackSummary

Description

Describes the number of layers and apps in a specified stack, and the number of instances in each state, such as running_setup or online. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data DescribeStackSummary

Request constructor

describeStackSummary

Arguments

:: Text
dssStackId

-> DescribeStackSummary

Smart constructor for the minimum required parameters to construct a valid DescribeStackSummary request.

The fields accessible through corresponding lenses are:

StackId :: Text

Request lenses

dssStackId :: Lens' DescribeStackSummary Text

The stack ID.

Response

data DescribeStackSummaryResponse

Contains the response to a DescribeStackSummary request.

Response constructor

describeStackSummaryResponse :: DescribeStackSummaryResponse

Smart constructor for the minimum required parameters to construct a valid DescribeStackSummaryResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

StackSummary :: Maybe StackSummary

Response lenses

dssrStackSummary :: Lens' DescribeStackSummaryResponse (Maybe StackSummary)

A StackSummary object that contains the results.