Network.AWS.OpsWorks.DescribeStacks

Description

Requests a description of one or more stacks.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeStacks.html

Request

data DescribeStacks

Request constructor

describeStacks :: DescribeStacks

DescribeStacks constructor.

The fields accessible through corresponding lenses are:

dsStackIds :: [Text]

Request lenses

dsStackIds :: Lens' DescribeStacks [Text]

An array of stack IDs that specify the stacks to be described. If you omit this parameter, DescribeStacks returns a description of every stack.

Response

data DescribeStacksResponse

Response constructor

describeStacksResponse :: DescribeStacksResponse

DescribeStacksResponse constructor.

The fields accessible through corresponding lenses are:

dsrStacks :: [Stack]

Response lenses

dsrStacks :: Lens' DescribeStacksResponse [Stack]

An array of Stack objects that describe the stacks.