Network.AWS.OpsWorks.DescribeTimeBasedAutoScaling

Description

Describes time-based auto scaling configurations for specified instances. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data DescribeTimeBasedAutoScaling

Request constructor

describeTimeBasedAutoScaling

Arguments

:: [Text]
dtbasInstanceIds

-> DescribeTimeBasedAutoScaling

Smart constructor for the minimum required parameters to construct a valid DescribeTimeBasedAutoScaling request.

The fields accessible through corresponding lenses are:

InstanceIds :: [Text]

Request lenses

dtbasInstanceIds :: Lens' DescribeTimeBasedAutoScaling [Text]

An array of instance IDs.

Response

data DescribeTimeBasedAutoScalingResponse

Contains the response to a DescribeTimeBasedAutoScaling request.

Response constructor

describeTimeBasedAutoScalingResponse :: DescribeTimeBasedAutoScalingResponse

Smart constructor for the minimum required parameters to construct a valid DescribeTimeBasedAutoScalingResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TimeBasedAutoScalingConfigurations :: [TimeBasedAutoScalingConfiguration]

Response lenses

dtbasrTimeBasedAutoScalingConfigurations :: Lens' DescribeTimeBasedAutoScalingResponse [TimeBasedAutoScalingConfiguration]

An array of TimeBasedAutoScalingConfiguration objects that describe the configuration for the specified instances.