Network.AWS.OpsWorks.DescribeUserProfiles

Description

Describe specified users.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeUserProfiles.html

Request

data DescribeUserProfiles

Request constructor

describeUserProfiles :: DescribeUserProfiles

DescribeUserProfiles constructor.

The fields accessible through corresponding lenses are:

dupIamUserArns :: [Text]

Request lenses

dupIamUserArns :: Lens' DescribeUserProfiles [Text]

An array of IAM user ARNs that identify the users to be described.

Response

data DescribeUserProfilesResponse

Response constructor

describeUserProfilesResponse :: DescribeUserProfilesResponse

DescribeUserProfilesResponse constructor.

The fields accessible through corresponding lenses are:

duprUserProfiles :: [UserProfile]

Response lenses

duprUserProfiles :: Lens' DescribeUserProfilesResponse [UserProfile]

A Users object that describes the specified users.