Network.AWS.OpsWorks.DescribeVolumes

Description

Describes an instance's Amazon EBS volumes.

Required Permissions : To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

describeVolumes :: DescribeVolumes

Creates a value of DescribeVolumes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvInstanceId - The instance ID. If you use this parameter, DescribeVolumes returns descriptions of the volumes associated with the specified instance.

dvVolumeIds - Am array of volume IDs. If you use this parameter, DescribeVolumes returns descriptions of the specified volumes. Otherwise, it returns a description of every volume.

dvRAIdArrayId - The RAID array ID. If you use this parameter, DescribeVolumes returns descriptions of the volumes associated with the specified RAID array.

dvStackId - A stack ID. The action describes the stack's registered Amazon EBS volumes.

data DescribeVolumes

See: describeVolumes smart constructor.

Request Lenses

dvInstanceId :: Lens' DescribeVolumes (Maybe Text)

The instance ID. If you use this parameter, DescribeVolumes returns descriptions of the volumes associated with the specified instance.

dvVolumeIds :: Lens' DescribeVolumes [Text]

Am array of volume IDs. If you use this parameter, DescribeVolumes returns descriptions of the specified volumes. Otherwise, it returns a description of every volume.

dvRAIdArrayId :: Lens' DescribeVolumes (Maybe Text)

The RAID array ID. If you use this parameter, DescribeVolumes returns descriptions of the volumes associated with the specified RAID array.

dvStackId :: Lens' DescribeVolumes (Maybe Text)

A stack ID. The action describes the stack's registered Amazon EBS volumes.

Destructuring the Response

describeVolumesResponse

Arguments

:: Int
dvrsResponseStatus

-> DescribeVolumesResponse

Creates a value of DescribeVolumesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvrsVolumes - An array of volume IDs.

dvrsResponseStatus - -- | The response status code.

data DescribeVolumesResponse

Contains the response to a DescribeVolumes request.

See: describeVolumesResponse smart constructor.

Response Lenses

dvrsVolumes :: Lens' DescribeVolumesResponse [Volume]

An array of volume IDs.

dvrsResponseStatus :: Lens' DescribeVolumesResponse Int

- | The response status code.