Network.AWS.OpsWorks.DetachElasticLoadBalancer

Description

Detaches a specified Elastic Load Balancing instance from its layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data DetachElasticLoadBalancer

Request constructor

mkDetachElasticLoadBalancer

Arguments

:: Text
delb1ElasticLoadBalancerName

-> Text
delb1LayerId

-> DetachElasticLoadBalancer

Smart constructor for the minimum required parameters to construct a valid DetachElasticLoadBalancer request.

The fields accessible through corresponding lenses are:

ElasticLoadBalancerName :: Text

LayerId :: Text

Request lenses

delb1ElasticLoadBalancerName :: Lens' DetachElasticLoadBalancer Text

The Elastic Load Balancing instance's name.

delb1LayerId :: Lens' DetachElasticLoadBalancer Text

The ID of the layer that the Elastic Load Balancing instance is attached to.

Response

data DetachElasticLoadBalancerResponse

Response constructor

mkDetachElasticLoadBalancerResponse :: DetachElasticLoadBalancerResponse

Smart constructor for the minimum required parameters to construct a valid DetachElasticLoadBalancerResponse response.

This constructor is provided for convenience and testing purposes.