Network.AWS.OpsWorks.DisassociateElasticIp

Description

Disassociates an Elastic IP address from its instance. The address remains registered with the stack. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DisassociateElasticIp.html

Request

data DisassociateElasticIp

Request constructor

disassociateElasticIp

Arguments

:: Text
deiElasticIp

-> DisassociateElasticIp

DisassociateElasticIp constructor.

The fields accessible through corresponding lenses are:

deiElasticIp :: Text

Request lenses

deiElasticIp :: Lens' DisassociateElasticIp Text

The Elastic IP address.

Response

data DisassociateElasticIpResponse

Response constructor

disassociateElasticIpResponse :: DisassociateElasticIpResponse

DisassociateElasticIpResponse constructor.