Network.AWS.OpsWorks.RegisterElasticIp

Description

Registers an Elastic IP address with a specified stack. An address can be registered with only one stack at a time. If the address is already registered, you must first deregister it by calling DeregisterElasticIp. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_RegisterElasticIp.html

Request

data RegisterElasticIp

Request constructor

registerElasticIp

Arguments

:: Text
reiElasticIp

-> Text
reiStackId

-> RegisterElasticIp

RegisterElasticIp constructor.

The fields accessible through corresponding lenses are:

reiElasticIp :: Text

reiStackId :: Text

Request lenses

reiElasticIp :: Lens' RegisterElasticIp Text

The Elastic IP address.

reiStackId :: Lens' RegisterElasticIp Text

The stack ID.

Response

data RegisterElasticIpResponse

Response constructor

registerElasticIpResponse :: RegisterElasticIpResponse

RegisterElasticIpResponse constructor.

The fields accessible through corresponding lenses are:

reirElasticIp :: Maybe Text

Response lenses

reirElasticIp :: Lens' RegisterElasticIpResponse (Maybe Text)

The Elastic IP address.