Network.AWS.OpsWorks.RegisterInstance

Description

Registers instances with a specified stack that were created outside of AWS OpsWorks.

We do not recommend using this action to register instances. The complete registration operation has two primary steps, installing the AWS OpsWorks agent on the instance and registering the instance with the stack. RegisterInstance handles only the second step. You should instead use the AWS CLI register command, which performs the entire registration operation. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_RegisterInstance.html

Request

data RegisterInstance

Request constructor

registerInstance

Arguments

:: Text
riStackId

-> RegisterInstance

RegisterInstance constructor.

The fields accessible through corresponding lenses are:

riHostname :: Maybe Text

riInstanceIdentity :: Maybe InstanceIdentity

riPrivateIp :: Maybe Text

riPublicIp :: Maybe Text

riRsaPublicKey :: Maybe Text

riRsaPublicKeyFingerprint :: Maybe Text

riStackId :: Text

Request lenses

riHostname :: Lens' RegisterInstance (Maybe Text)

The instance's hostname.

riInstanceIdentity :: Lens' RegisterInstance (Maybe InstanceIdentity)

An InstanceIdentity object that contains the instance's identity.

riPrivateIp :: Lens' RegisterInstance (Maybe Text)

The instance's private IP address.

riPublicIp :: Lens' RegisterInstance (Maybe Text)

The instance's public IP address.

riRsaPublicKey :: Lens' RegisterInstance (Maybe Text)

The instances public RSA key. This key is used to encrypt communication between the instance and the service.

riRsaPublicKeyFingerprint :: Lens' RegisterInstance (Maybe Text)

The instances public RSA key fingerprint.

riStackId :: Lens' RegisterInstance Text

The ID of the stack that the instance is to be registered with.

Response

data RegisterInstanceResponse

Response constructor

registerInstanceResponse :: RegisterInstanceResponse

RegisterInstanceResponse constructor.

The fields accessible through corresponding lenses are:

rirInstanceId :: Maybe Text

Response lenses

rirInstanceId :: Lens' RegisterInstanceResponse (Maybe Text)

The registered instance's AWS OpsWorks ID.