Network.AWS.OpsWorks.RegisterRDSDBInstance

Description

Registers an Amazon RDS instance with a stack.

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

registerRDSDBInstance

Arguments

:: Text
rrdiStackId

-> Text
rrdiRDSDBInstanceARN

-> Text
rrdiDBUser

-> Text
rrdiDBPassword

-> RegisterRDSDBInstance

Creates a value of RegisterRDSDBInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrdiStackId - The stack ID.

rrdiRDSDBInstanceARN - The Amazon RDS instance's ARN.

rrdiDBUser - The database's master user name.

rrdiDBPassword - The database password.

data RegisterRDSDBInstance

See: registerRDSDBInstance smart constructor.

Request Lenses

rrdiStackId :: Lens' RegisterRDSDBInstance Text

The stack ID.

rrdiRDSDBInstanceARN :: Lens' RegisterRDSDBInstance Text

The Amazon RDS instance's ARN.

rrdiDBUser :: Lens' RegisterRDSDBInstance Text

The database's master user name.

rrdiDBPassword :: Lens' RegisterRDSDBInstance Text

The database password.

Destructuring the Response

registerRDSDBInstanceResponse :: RegisterRDSDBInstanceResponse

Creates a value of RegisterRDSDBInstanceResponse with the minimum fields required to make a request.

data RegisterRDSDBInstanceResponse

See: registerRDSDBInstanceResponse smart constructor.