Network.AWS.OpsWorks.RegisterRdsDbInstance

Description

Registers an Amazon RDS instance with a stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_RegisterRdsDbInstance.html

Request

data RegisterRdsDbInstance

Request constructor

registerRdsDbInstance

Arguments

:: Text
rrdiStackId

-> Text
rrdiRdsDbInstanceArn

-> Text
rrdiDbUser

-> Text
rrdiDbPassword

-> RegisterRdsDbInstance

RegisterRdsDbInstance constructor.

The fields accessible through corresponding lenses are:

rrdiDbPassword :: Text

rrdiDbUser :: Text

rrdiRdsDbInstanceArn :: Text

rrdiStackId :: Text

Request lenses

rrdiDbPassword :: Lens' RegisterRdsDbInstance Text

The database password.

rrdiDbUser :: Lens' RegisterRdsDbInstance Text

The database's master user name.

rrdiRdsDbInstanceArn :: Lens' RegisterRdsDbInstance Text

The Amazon RDS instance's ARN.

rrdiStackId :: Lens' RegisterRdsDbInstance Text

The stack ID.

Response

data RegisterRdsDbInstanceResponse

Response constructor

registerRdsDbInstanceResponse :: RegisterRdsDbInstanceResponse

RegisterRdsDbInstanceResponse constructor.