Network.AWS.OpsWorks.RegisterVolume

Description

Registers an Amazon EBS volume with a specified stack. A volume can be registered with only one stack at a time. If the volume is already registered, you must first deregister it by calling DeregisterVolume . For more information, see Resource Management .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

registerVolume

Arguments

:: Text
rvStackId

-> RegisterVolume

Creates a value of RegisterVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rvEC2VolumeId - The Amazon EBS volume ID.

rvStackId - The stack ID.

data RegisterVolume

See: registerVolume smart constructor.

Request Lenses

rvEC2VolumeId :: Lens' RegisterVolume (Maybe Text)

The Amazon EBS volume ID.

rvStackId :: Lens' RegisterVolume Text

The stack ID.

Destructuring the Response

registerVolumeResponse

Arguments

:: Int
rvrsResponseStatus

-> RegisterVolumeResponse

Creates a value of RegisterVolumeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rvrsVolumeId - The volume ID.

rvrsResponseStatus - -- | The response status code.

data RegisterVolumeResponse

Contains the response to a RegisterVolume request.

See: registerVolumeResponse smart constructor.

Response Lenses

rvrsVolumeId :: Lens' RegisterVolumeResponse (Maybe Text)

The volume ID.

rvrsResponseStatus :: Lens' RegisterVolumeResponse Int

- | The response status code.