Network.AWS.OpsWorks.SetPermission

Description

Specifies a user's permissions. For more information, see Security and Permissions .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

setPermission

Arguments

:: Text
spStackId

-> Text
spIAMUserARN

-> SetPermission

Creates a value of SetPermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spAllowSudo - The user is allowed to use sudo to elevate privileges.

spLevel - The user's permission level, which must be set to one of the following strings. You cannot set your own permissions level. * deny * show * deploy * manage * iam_only For more information on the permissions associated with these levels, see Managing User Permissions .

spAllowSSH - The user is allowed to use SSH to communicate with the instance.

spStackId - The stack ID.

spIAMUserARN - The user's IAM ARN. This can also be a federated user's ARN.

data SetPermission

See: setPermission smart constructor.

Request Lenses

spAllowSudo :: Lens' SetPermission (Maybe Bool)

The user is allowed to use sudo to elevate privileges.

spLevel :: Lens' SetPermission (Maybe Text)

The user's permission level, which must be set to one of the following strings. You cannot set your own permissions level. * deny * show * deploy * manage * iam_only For more information on the permissions associated with these levels, see Managing User Permissions .

spAllowSSH :: Lens' SetPermission (Maybe Bool)

The user is allowed to use SSH to communicate with the instance.

spStackId :: Lens' SetPermission Text

The stack ID.

spIAMUserARN :: Lens' SetPermission Text

The user's IAM ARN. This can also be a federated user's ARN.

Destructuring the Response

setPermissionResponse :: SetPermissionResponse

Creates a value of SetPermissionResponse with the minimum fields required to make a request.

data SetPermissionResponse

See: setPermissionResponse smart constructor.