Network.AWS.OpsWorks.SetTimeBasedAutoScaling

Description

Specify the time-based auto scaling configuration for a specified instance. For more information, see Managing Load with Time-based and Load-basedInstances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_SetTimeBasedAutoScaling.html

Request

data SetTimeBasedAutoScaling

Request constructor

setTimeBasedAutoScaling

Arguments

:: Text
stbasInstanceId

-> SetTimeBasedAutoScaling

SetTimeBasedAutoScaling constructor.

The fields accessible through corresponding lenses are:

stbasAutoScalingSchedule :: Maybe WeeklyAutoScalingSchedule

stbasInstanceId :: Text

Request lenses

stbasAutoScalingSchedule :: Lens' SetTimeBasedAutoScaling (Maybe WeeklyAutoScalingSchedule)

An AutoScalingSchedule with the instance schedule.

stbasInstanceId :: Lens' SetTimeBasedAutoScaling Text

The instance ID.

Response

data SetTimeBasedAutoScalingResponse

Response constructor

setTimeBasedAutoScalingResponse :: SetTimeBasedAutoScalingResponse

SetTimeBasedAutoScalingResponse constructor.