Network.AWS.OpsWorks.StartStack

Description

Starts a stack's instances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_StartStack.html

Request

data StartStack

Request constructor

startStack

Arguments

:: Text
ss2StackId

-> StartStack

StartStack constructor.

The fields accessible through corresponding lenses are:

ss2StackId :: Text

Request lenses

ss2StackId :: Lens' StartStack Text

The stack ID.

Response

data StartStackResponse

Response constructor

startStackResponse :: StartStackResponse

StartStackResponse constructor.