Network.AWS.OpsWorks.UpdateApp

Description

Updates a specified app. Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see <http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html Managing User Permissions>.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_UpdateApp.html

Request

data UpdateApp

Request constructor

updateApp

Arguments

:: Text
uaAppId

-> UpdateApp

UpdateApp constructor.

The fields accessible through corresponding lenses are:

uaAppId :: Text

uaAppSource :: Maybe Source

uaAttributes :: HashMap AppAttributesKeys Text

uaDataSources :: [DataSource]

uaDescription :: Maybe Text

uaDomains :: [Text]

uaEnableSsl :: Maybe Bool

uaName :: Maybe Text

uaSslConfiguration :: Maybe SslConfiguration

uaType :: Maybe AppType

Request lenses

uaAppId :: Lens' UpdateApp Text

The app ID.

uaAppSource :: Lens' UpdateApp (Maybe Source)

A Source object that specifies the app repository.

uaAttributes :: Lens' UpdateApp (HashMap AppAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

uaDataSources :: Lens' UpdateApp [DataSource]

The app's data sources.

uaDescription :: Lens' UpdateApp (Maybe Text)

A description of the app.

uaDomains :: Lens' UpdateApp [Text]

The app's virtual host settings, with multiple domains separated by commas. For example: ''www.example.com, example.com''.

uaEnableSsl :: Lens' UpdateApp (Maybe Bool)

Whether SSL is enabled for the app.

uaName :: Lens' UpdateApp (Maybe Text)

The app name.

uaSslConfiguration :: Lens' UpdateApp (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

uaType :: Lens' UpdateApp (Maybe AppType)

The app type.

Response

data UpdateAppResponse

Response constructor

updateAppResponse :: UpdateAppResponse

UpdateAppResponse constructor.