Network.AWS.OpsWorks.UpdateInstance

Description

Updates a specified instance. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see <http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html Managing User Permissions>.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_UpdateInstance.html

Request

data UpdateInstance

Request constructor

updateInstance

Arguments

:: Text
uiInstanceId

-> UpdateInstance

UpdateInstance constructor.

The fields accessible through corresponding lenses are:

uiAmiId :: Maybe Text

uiArchitecture :: Maybe Architecture

uiAutoScalingType :: Maybe AutoScalingType

uiEbsOptimized :: Maybe Bool

uiHostname :: Maybe Text

uiInstallUpdatesOnBoot :: Maybe Bool

uiInstanceId :: Text

uiInstanceType :: Maybe Text

uiLayerIds :: [Text]

uiOs :: Maybe Text

uiSshKeyName :: Maybe Text

Request lenses

uiAmiId :: Lens' UpdateInstance (Maybe Text)

A custom AMI ID to be used to create the instance. The AMI should be based on one of the standard AWS OpsWorks APIs: Amazon Linux or Ubuntu 12.04 LTS. For more information, see <http://docs.aws.amazon.com/opsworks/latest/userguide/workinginstances.html Instances>.

uiArchitecture :: Lens' UpdateInstance (Maybe Architecture)

The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see <http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html Instance Families and Types>.

uiAutoScalingType :: Lens' UpdateInstance (Maybe AutoScalingType)

The instance's auto scaling type, which has three possible values: AlwaysRunning: A 24/7 instance, which is not affected by auto scaling. TimeBasedAutoScaling: A time-based auto scaling instance, which is started and stopped based on a specified schedule. LoadBasedAutoScaling: A load-based auto scaling instance, which is started and stopped based on load metrics.

uiEbsOptimized :: Lens' UpdateInstance (Maybe Bool)

Whether this is an Amazon EBS-optimized instance.

uiHostname :: Lens' UpdateInstance (Maybe Text)

The instance host name.

uiInstallUpdatesOnBoot :: Lens' UpdateInstance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

uiInstanceId :: Lens' UpdateInstance Text

The instance ID.

uiInstanceType :: Lens' UpdateInstance (Maybe Text)

The instance type. AWS OpsWorks supports all instance types except Cluster Compute, Cluster GPU, and High Memory Cluster. For more information, see <http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html Instance Families and Types>. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

uiLayerIds :: Lens' UpdateInstance [Text]

The instance's layer IDs.

uiOs :: Lens' UpdateInstance (Maybe Text)

The instance operating system, which must be set to one of the following. Standard operating systems: Amazon Linux or Ubuntu 12.04 LTS Custom AMIs: Custom The default option is Amazon Linux. If you set this parameter to Custom, you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. For more information on the standard operating systems, see <http://docs.aws.amazon.com/opsworks/latest/userguide/workinginstances-os.html Operating Systems>For more information on how to use custom AMIs with OpsWorks, see <http://docs.aws.amazon.com/opsworks/latest/userguide/workinginstances-custom-ami.html Using Custom AMIs>.

uiSshKeyName :: Lens' UpdateInstance (Maybe Text)

The instance SSH key name.

Response

data UpdateInstanceResponse

Response constructor

updateInstanceResponse :: UpdateInstanceResponse

UpdateInstanceResponse constructor.