Network.AWS.OpsWorks.UpdateLayer

Description

Updates a specified layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data UpdateLayer

Request constructor

mkUpdateLayer

Arguments

:: Text
ulLayerId

-> UpdateLayer

Smart constructor for the minimum required parameters to construct a valid UpdateLayer request.

The fields accessible through corresponding lenses are:

LayerId :: Text

Name :: Maybe Text

Shortname :: Maybe Text

Attributes :: Map LayerAttributesKeys Text

CustomInstanceProfileArn :: Maybe Text

CustomSecurityGroupIds :: [Text]

Packages :: [Text]

VolumeConfigurations :: [VolumeConfiguration]

EnableAutoHealing :: Maybe Bool

AutoAssignElasticIps :: Maybe Bool

AutoAssignPublicIps :: Maybe Bool

CustomRecipes :: Maybe Recipes

InstallUpdatesOnBoot :: Maybe Bool

UseEbsOptimizedInstances :: Maybe Bool

Request lenses

ulLayerId :: Lens' UpdateLayer Text

The layer ID.

ulName :: Lens' UpdateLayer (Maybe Text)

The layer name, which is used by the console.

ulShortname :: Lens' UpdateLayer (Maybe Text)

The layer short name, which is used internally by AWS OpsWorksand by Chef. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters and must be in the following format: A[a-z0-9-_.]+Z.

ulAttributes :: Lens' UpdateLayer (Map LayerAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

ulCustomInstanceProfileArn :: Lens' UpdateLayer (Maybe Text)

The ARN of an IAM profile to be used for all of the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

ulCustomSecurityGroupIds :: Lens' UpdateLayer [Text]

An array containing the layer's custom security group IDs.

ulPackages :: Lens' UpdateLayer [Text]

An array of Package objects that describe the layer's packages.

ulVolumeConfigurations :: Lens' UpdateLayer [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

ulEnableAutoHealing :: Lens' UpdateLayer (Maybe Bool)

Whether to disable auto healing for the layer.

ulAutoAssignElasticIps :: Lens' UpdateLayer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

ulAutoAssignPublicIps :: Lens' UpdateLayer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

ulCustomRecipes :: Lens' UpdateLayer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer's custom recipes.

ulInstallUpdatesOnBoot :: Lens' UpdateLayer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances. We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

ulUseEbsOptimizedInstances :: Lens' UpdateLayer (Maybe Bool)

Whether to use Amazon EBS-optimized instances.

Response

data UpdateLayerResponse

Response constructor

mkUpdateLayerResponse :: UpdateLayerResponse

Smart constructor for the minimum required parameters to construct a valid UpdateLayerResponse response.

This constructor is provided for convenience and testing purposes.