Network.AWS.OpsWorks.UpdateMyUserProfile

Description

Updates a user's SSH public key.

Required Permissions: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_UpdateMyUserProfile.html

Request

data UpdateMyUserProfile

Request constructor

updateMyUserProfile :: UpdateMyUserProfile

UpdateMyUserProfile constructor.

The fields accessible through corresponding lenses are:

umupSshPublicKey :: Maybe Text

Request lenses

umupSshPublicKey :: Lens' UpdateMyUserProfile (Maybe Text)

The user's SSH public key.

Response

data UpdateMyUserProfileResponse

Response constructor

updateMyUserProfileResponse :: UpdateMyUserProfileResponse

UpdateMyUserProfileResponse constructor.