Network.AWS.OpsWorks.UpdateRdsDbInstance

Description

Updates an Amazon RDS instance.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_UpdateRdsDbInstance.html

Request

data UpdateRdsDbInstance

Request constructor

updateRdsDbInstance

Arguments

:: Text
urdiRdsDbInstanceArn

-> UpdateRdsDbInstance

UpdateRdsDbInstance constructor.

The fields accessible through corresponding lenses are:

urdiDbPassword :: Maybe Text

urdiDbUser :: Maybe Text

urdiRdsDbInstanceArn :: Text

Request lenses

urdiDbPassword :: Lens' UpdateRdsDbInstance (Maybe Text)

The database password.

urdiDbUser :: Lens' UpdateRdsDbInstance (Maybe Text)

The master user name.

urdiRdsDbInstanceArn :: Lens' UpdateRdsDbInstance Text

The Amazon RDS instance's ARN.

Response

data UpdateRdsDbInstanceResponse

Response constructor

updateRdsDbInstanceResponse :: UpdateRdsDbInstanceResponse

UpdateRdsDbInstanceResponse constructor.