Network.AWS.OpsWorks.UpdateUserProfile

Description

Updates a specified user profile.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_UpdateUserProfile.html

Request

data UpdateUserProfile

Request constructor

updateUserProfile

Arguments

:: Text
uupIamUserArn

-> UpdateUserProfile

UpdateUserProfile constructor.

The fields accessible through corresponding lenses are:

uupAllowSelfManagement :: Maybe Bool

uupIamUserArn :: Text

uupSshPublicKey :: Maybe Text

uupSshUsername :: Maybe Text

Request lenses

uupAllowSelfManagement :: Lens' UpdateUserProfile (Maybe Bool)

Whether users can specify their own SSH public key through the My Settings page. For more information, see Managing User Permissions.

uupIamUserArn :: Lens' UpdateUserProfile Text

The user IAM ARN.

uupSshPublicKey :: Lens' UpdateUserProfile (Maybe Text)

The user's new SSH public key.

uupSshUsername :: Lens' UpdateUserProfile (Maybe Text)

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], -, and '_'. If the specified name includes other punctuation marks, AWS OpsWorks removes them. For example, 'my.name' will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks generates one from the IAM user name.

Response

data UpdateUserProfileResponse

Response constructor

updateUserProfileResponse :: UpdateUserProfileResponse

UpdateUserProfileResponse constructor.