Network.AWS.OpsWorks.V2013_02_18.AssociateElasticIp

Description

Associates one of the stack's registered Elastic IP addresses with a specified instance. The address must first be registered with the stack by calling RegisterElasticIp. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

associateElasticIp

Arguments

:: Text
_aeirElasticIp

-> AssociateElasticIp

Minimum specification for a AssociateElasticIp request.

data AssociateElasticIp

Constructors

AssociateElasticIp

Fields

_aeirElasticIp :: Text

The Elastic IP address.

_aeirInstanceId :: Maybe Text

The instance ID.

aeirInstanceId :: Lens' AssociateElasticIp (Maybe Text)

aeirElasticIp :: Lens' AssociateElasticIp Text

data AssociateElasticIpResponse

Constructors

AssociateElasticIpResponse