Network.AWS.OpsWorks.V2013_02_18.CreateApp

Description

Creates an app for a specified stack. For more information, see Creating Apps. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

createApp

Arguments

:: AppType
_carType

-> Text
_carName

-> Text
_carStackId

-> CreateApp

Minimum specification for a CreateApp request.

data CreateApp

Constructors

CreateApp

Fields

_carType :: AppType

The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks deploys an application to those instances that are members of the corresponding layer.

_carName :: Text

The app name.

_carStackId :: Text

The stack ID.

_carAttributes :: Map AppAttributesKeys Text

One or more user-defined key/value pairs to be added to the stack attributes.

_carEnableSsl :: Maybe Bool

Whether to enable SSL for the app.

_carDataSources :: [DataSource]

The app's data source.

_carAppSource :: Maybe Source

A Source object that specifies the app repository.

_carSslConfiguration :: Maybe SslConfiguration

An SslConfiguration object with the SSL configuration.

_carShortname :: Maybe Text

The app's short name.

_carDescription :: Maybe Text

A description of the app.

_carDomains :: [Text]

The app virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'.

carType :: Lens' CreateApp AppType

carStackId :: Lens' CreateApp Text

carSslConfiguration :: Lens' CreateApp (Maybe SslConfiguration)

carShortname :: Lens' CreateApp (Maybe Text)

carName :: Lens' CreateApp Text

carEnableSsl :: Lens' CreateApp (Maybe Bool)

carDomains :: Lens' CreateApp [Text]

carDescription :: Lens' CreateApp (Maybe Text)

carDataSources :: Lens' CreateApp [DataSource]

carAttributes :: Lens' CreateApp (Map AppAttributesKeys Text)

carAppSource :: Lens' CreateApp (Maybe Source)

data CreateAppResponse

Constructors

CreateAppResponse

Fields

_casAppId :: Maybe Text

The app ID.

casAppId :: Iso' CreateAppResponse (Maybe Text)