Network.AWS.OpsWorks.V2013_02_18.CreateDeployment

Description

Deploys a stack or app. App deployment generates a deploy event, which runs the associated recipes and passes them a JSON stack configuration object that includes information about the app. Stack deployment runs the deploy recipes but does not raise an event. For more information, see Deploying Apps and Run Stack Commands. Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

createDeployment

Arguments

:: DeploymentCommand
_cdrCommand

-> Text
_cdrStackId

-> CreateDeployment

Minimum specification for a CreateDeployment request.

data CreateDeployment

Constructors

CreateDeployment

Fields

_cdrCommand :: DeploymentCommand

A DeploymentCommand object that specifies the deployment command and any associated arguments.

_cdrStackId :: Text

The stack ID.

_cdrCustomJson :: Maybe Text

A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format and must escape characters such as '"'.: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration JSON.

_cdrAppId :: Maybe Text

The app ID. This parameter is required for app deployments, but not for other deployment commands.

_cdrComment :: Maybe Text

A user-defined comment.

_cdrInstanceIds :: [Text]

The instance IDs for the deployment targets.

cdrStackId :: Lens' CreateDeployment Text

cdrInstanceIds :: Lens' CreateDeployment [Text]

cdrCustomJson :: Lens' CreateDeployment (Maybe Text)

cdrComment :: Lens' CreateDeployment (Maybe Text)

cdrCommand :: Lens' CreateDeployment DeploymentCommand

cdrAppId :: Lens' CreateDeployment (Maybe Text)

data CreateDeploymentResponse

Constructors

CreateDeploymentResponse

Fields

_cdsDeploymentId :: Maybe Text

The deployment ID, which can be used with other requests to identify the deployment.

cdsDeploymentId :: Iso' CreateDeploymentResponse (Maybe Text)