Network.AWS.OpsWorks.V2013_02_18.CreateUserProfile

Description

Creates a new user profile. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

createUserProfile

Arguments

:: Text
_cuprIamUserArn

-> CreateUserProfile

Minimum specification for a CreateUserProfile request.

data CreateUserProfile

Constructors

CreateUserProfile

Fields

_cuprIamUserArn :: Text

The user's IAM ARN.

_cuprAllowSelfManagement :: Maybe Bool

Whether users can specify their own SSH public key through the My Settings page. For more information, see Setting an IAM User's Public SSH Key.

_cuprSshPublicKey :: Maybe Text

The user's public SSH key.

_cuprSshUsername :: Maybe Text

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], -, and '_'. If the specified name includes other punctuation marks, AWS OpsWorks removes them. For example, my.name will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks generates one from the IAM user name.

cuprSshUsername :: Lens' CreateUserProfile (Maybe Text)

cuprSshPublicKey :: Lens' CreateUserProfile (Maybe Text)

cuprIamUserArn :: Lens' CreateUserProfile Text

cuprAllowSelfManagement :: Lens' CreateUserProfile (Maybe Bool)

data CreateUserProfileResponse

Constructors

CreateUserProfileResponse

Fields

_cupsIamUserArn :: Maybe Text

The user's IAM ARN.

cupsIamUserArn :: Iso' CreateUserProfileResponse (Maybe Text)