Network.AWS.OpsWorks.V2013_02_18.DeleteLayer

Description

Deletes a specified layer. You must first stop and then delete all associated instances. For more information, see How to Delete a Layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data DeleteLayer

Constructors

DeleteLayer

Fields

_dltLayerId :: Text

The layer ID.

dltLayerId :: Iso' DeleteLayer Text

data DeleteLayerResponse

Constructors

DeleteLayerResponse