Network.AWS.OpsWorks.V2013_02_18.DescribeApps

Description

Requests a description of a specified set of apps. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data DescribeApps

Request constructor

mkDescribeApps :: DescribeApps

Smart constructor for the minimum required parameters to construct a valid DescribeApps request.

The fields accessible through corresponding lenses are:

StackId :: Maybe Text

AppIds :: [Text]

Request lenses

da1StackId :: Lens' DescribeApps (Maybe Text)

The app stack ID. If you use this parameter, DescribeApps returns a description of the apps in the specified stack.

da1AppIds :: Lens' DescribeApps [Text]

An array of app IDs for the apps to be described. If you use this parameter, DescribeApps returns a description of the specified apps. Otherwise, it returns a description of every app.

Response

data DescribeAppsResponse

Contains the response to a DescribeApps request.

Response constructor

mkDescribeAppsResponse :: DescribeAppsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeAppsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Apps :: [App]

Response lenses

darApps :: Lens' DescribeAppsResponse [App]

An array of App objects that describe the specified apps.