Network.AWS.OpsWorks.V2013_02_18.DescribeDeployments

Description

Requests a description of a specified set of deployments. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeDeployments :: DescribeDeployments

Minimum specification for a DescribeDeployments request.

data DescribeDeployments

Constructors

DescribeDeployments

Fields

_ddrAppId :: Maybe Text

The app ID. If you include this parameter, DescribeDeployments returns a description of the commands associated with the specified app.

_ddrStackId :: Maybe Text

The stack ID. If you include this parameter, DescribeDeployments returns a description of the commands associated with the specified stack.

_ddrDeploymentIds :: [Text]

An array of deployment IDs to be described. If you include this parameter, DescribeDeployments returns a description of the specified deployments. Otherwise, it returns a description of every deployment.

ddrStackId :: Lens' DescribeDeployments (Maybe Text)

ddrDeploymentIds :: Lens' DescribeDeployments [Text]

ddrAppId :: Lens' DescribeDeployments (Maybe Text)

data DescribeDeploymentsResponse

Constructors

DescribeDeploymentsResponse

Fields

_ddsDeployments :: [Deployment]

An array of Deployment objects that describe the deployments.

ddsDeployments :: Iso' DescribeDeploymentsResponse [Deployment]