Network.AWS.OpsWorks.V2013_02_18.DescribeElasticIps

Description

Describes Elastic IP addresses. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeElasticIps :: DescribeElasticIps

Minimum specification for a DescribeElasticIps request.

data DescribeElasticIps

Constructors

DescribeElasticIps

Fields

_deitInstanceId :: Maybe Text

The instance ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses associated with the specified instance.

_deitStackId :: Maybe Text

A stack ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses that are registered with the specified stack.

_deitIps :: [Text]

An array of Elastic IP addresses to be described. If you include this parameter, DescribeElasticIps returns a description of the specified Elastic IP addresses. Otherwise, it returns a description of every Elastic IP address.

deitStackId :: Lens' DescribeElasticIps (Maybe Text)

deitIps :: Lens' DescribeElasticIps [Text]

deitInstanceId :: Lens' DescribeElasticIps (Maybe Text)

data DescribeElasticIpsResponse

Constructors

DescribeElasticIpsResponse

Fields

_deiuElasticIps :: [ElasticIp]

An ElasticIps object that describes the specified Elastic IP addresses.

deiuElasticIps :: Iso' DescribeElasticIpsResponse [ElasticIp]