Network.AWS.OpsWorks.V2013_02_18.DescribeElasticLoadBalancers

Description

Describes a stack's Elastic Load Balancing instances. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeElasticLoadBalancers :: DescribeElasticLoadBalancers

Minimum specification for a DescribeElasticLoadBalancers request.

data DescribeElasticLoadBalancers

Constructors

DescribeElasticLoadBalancers

Fields

_delbsStackId :: Maybe Text

A stack ID. The action describes the stack's Elastic Load Balancing instances.

_delbsLayerIds :: [Text]

A list of layer IDs. The action describes the Elastic Load Balancing instances for the specified layers.

delbsStackId :: Lens' DescribeElasticLoadBalancers (Maybe Text)

delbsLayerIds :: Lens' DescribeElasticLoadBalancers [Text]

data DescribeElasticLoadBalancersResponse

Constructors

DescribeElasticLoadBalancersResponse

Fields

_delbtElasticLoadBalancers :: [ElasticLoadBalancer]

A list of ElasticLoadBalancer objects that describe the specified Elastic Load Balancing instances.

delbtElasticLoadBalancers :: Iso' DescribeElasticLoadBalancersResponse [ElasticLoadBalancer]