Network.AWS.OpsWorks.V2013_02_18.DescribeInstances

Description

Requests a description of a set of instances. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data DescribeInstances

Request constructor

mkDescribeInstances :: DescribeInstances

Smart constructor for the minimum required parameters to construct a valid DescribeInstances request.

The fields accessible through corresponding lenses are:

StackId :: Maybe Text

LayerId :: Maybe Text

InstanceIds :: [Text]

Request lenses

di1StackId :: Lens' DescribeInstances (Maybe Text)

A stack ID. If you use this parameter, DescribeInstances returns descriptions of the instances associated with the specified stack.

di1LayerId :: Lens' DescribeInstances (Maybe Text)

A layer ID. If you use this parameter, DescribeInstances returns descriptions of the instances associated with the specified layer.

di1InstanceIds :: Lens' DescribeInstances [Text]

An array of instance IDs to be described. If you use this parameter, DescribeInstances returns a description of the specified instances. Otherwise, it returns a description of every instance.

Response

data DescribeInstancesResponse

Contains the response to a DescribeInstances request.

Response constructor

mkDescribeInstancesResponse :: DescribeInstancesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeInstancesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Instances :: [Instance]

Response lenses

dirInstances :: Lens' DescribeInstancesResponse [Instance]

An array of Instance objects that describe the instances.