Network.AWS.OpsWorks.V2013_02_18.DescribeLayers

Description

Requests a description of one or more layers in a specified stack. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeLayers :: DescribeLayers

Minimum specification for a DescribeLayers request.

data DescribeLayers

Constructors

DescribeLayers

Fields

_dlrStackId :: Maybe Text

The stack ID.

_dlrLayerIds :: [Text]

An array of layer IDs that specify the layers to be described. If you omit this parameter, DescribeLayers returns a description of every layer in the specified stack.

dlrStackId :: Lens' DescribeLayers (Maybe Text)

dlrLayerIds :: Lens' DescribeLayers [Text]

data DescribeLayersResponse

Constructors

DescribeLayersResponse

Fields

_dlsLayers :: [Layer]

An array of Layer objects that describe the layers.

dlsLayers :: Iso' DescribeLayersResponse [Layer]