Network.AWS.OpsWorks.V2013_02_18.DescribePermissions

Description

Describes the permissions for a specified stack. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data DescribePermissions

Request constructor

mkDescribePermissions :: DescribePermissions

Smart constructor for the minimum required parameters to construct a valid DescribePermissions request.

Request lenses

dpIamUserArn :: Lens' DescribePermissions (Maybe Text)

The user's IAM ARN. For more information about IAM ARNs, see Using Identifiers.

dpStackId :: Lens' DescribePermissions (Maybe Text)

The stack ID.

Response

data DescribePermissionsResponse

Contains the response to a DescribePermissions request.

Response lenses

dprsPermissions :: Lens' DescribePermissionsResponse [Permission]

An array of Permission objects that describe the stack permissions. If the request object contains only a stack ID, the array contains a Permission object with permissions for each of the stack IAM ARNs. If the request object contains only an IAM ARN, the array contains a Permission object with permissions for each of the user's stack IDs. If the request contains a stack ID and an IAM ARN, the array contains a single Permission object with permissions for the specified stack and IAM ARN.