Network.AWS.OpsWorks.V2013_02_18.DescribeServiceErrors

Description

Describes AWS OpsWorks service errors. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeServiceErrors :: DescribeServiceErrors

Minimum specification for a DescribeServiceErrors request.

data DescribeServiceErrors

Constructors

DescribeServiceErrors

Fields

_dserInstanceId :: Maybe Text

The instance ID. If you use this parameter, DescribeServiceErrors returns descriptions of the errors associated with the specified instance.

_dserStackId :: Maybe Text

The stack ID. If you use this parameter, DescribeServiceErrors returns descriptions of the errors associated with the specified stack.

_dserServiceErrorIds :: [Text]

An array of service error IDs. If you use this parameter, DescribeServiceErrors returns descriptions of the specified errors. Otherwise, it returns a description of every error.

dserStackId :: Lens' DescribeServiceErrors (Maybe Text)

dserServiceErrorIds :: Lens' DescribeServiceErrors [Text]

dserInstanceId :: Lens' DescribeServiceErrors (Maybe Text)

data DescribeServiceErrorsResponse

Constructors

DescribeServiceErrorsResponse

Fields

_dsesServiceErrors :: [ServiceError]

An array of ServiceError objects that describe the specified service errors.

dsesServiceErrors :: Iso' DescribeServiceErrorsResponse [ServiceError]