Network.AWS.OpsWorks.V2013_02_18.DescribeStacks

Description

Requests a description of one or more stacks. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeStacks :: DescribeStacks

Minimum specification for a DescribeStacks request.

data DescribeStacks

Constructors

DescribeStacks

Fields

_dssStackIds :: [Text]

An array of stack IDs that specify the stacks to be described. If you omit this parameter, DescribeStacks returns a description of every stack.

dssStackIds :: Iso' DescribeStacks [Text]

data DescribeStacksResponse

Constructors

DescribeStacksResponse

Fields

_dstStacks :: [Stack]

An array of Stack objects that describe the stacks.

dstStacks :: Iso' DescribeStacksResponse [Stack]