Network.AWS.OpsWorks.V2013_02_18.DescribeTimeBasedAutoScaling

Description

Describes time-based auto scaling configurations for specified instances. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data DescribeTimeBasedAutoScaling

Constructors

DescribeTimeBasedAutoScaling

Fields

_dtbasrInstanceIds :: [Text]

An array of instance IDs.

dtbasrInstanceIds :: Iso' DescribeTimeBasedAutoScaling [Text]

data DescribeTimeBasedAutoScalingResponse

Constructors

DescribeTimeBasedAutoScalingResponse

Fields

_dtbassTimeBasedAutoScalingConfigurations :: [TimeBasedAutoScalingConfiguration]

An array of TimeBasedAutoScalingConfiguration objects that describe the configuration for the specified instances.

dtbassTimeBasedAutoScalingConfigurations :: Iso' DescribeTimeBasedAutoScalingResponse [TimeBasedAutoScalingConfiguration]