Network.AWS.OpsWorks.V2013_02_18.DescribeUserProfiles

Description

Describe specified users. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeUserProfiles :: DescribeUserProfiles

Minimum specification for a DescribeUserProfiles request.

data DescribeUserProfiles

Constructors

DescribeUserProfiles

Fields

_duprIamUserArns :: [Text]

An array of IAM user ARNs that identify the users to be described.

duprIamUserArns :: Iso' DescribeUserProfiles [Text]

data DescribeUserProfilesResponse

Constructors

DescribeUserProfilesResponse

Fields

_dupsUserProfiles :: [UserProfile]

A Users object that describes the specified users.

dupsUserProfiles :: Iso' DescribeUserProfilesResponse [UserProfile]