Network.AWS.OpsWorks.V2013_02_18.DescribeVolumes

Description

Describes an instance's Amazon EBS volumes. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeVolumes :: DescribeVolumes

Minimum specification for a DescribeVolumes request.

data DescribeVolumes

Constructors

DescribeVolumes

Fields

_dvsInstanceId :: Maybe Text

The instance ID. If you use this parameter, DescribeVolumes returns descriptions of the volumes associated with the specified instance.

_dvsRaidArrayId :: Maybe Text

The RAID array ID. If you use this parameter, DescribeVolumes returns descriptions of the volumes associated with the specified RAID array.

_dvsStackId :: Maybe Text

A stack ID. The action describes the stack's registered Amazon EBS volumes.

_dvsVolumeIds :: [Text]

Am array of volume IDs. If you use this parameter, DescribeVolumes returns descriptions of the specified volumes. Otherwise, it returns a description of every volume.

dvsVolumeIds :: Lens' DescribeVolumes [Text]

dvsStackId :: Lens' DescribeVolumes (Maybe Text)

dvsRaidArrayId :: Lens' DescribeVolumes (Maybe Text)

dvsInstanceId :: Lens' DescribeVolumes (Maybe Text)

data DescribeVolumesResponse

Constructors

DescribeVolumesResponse

Fields

_dvtVolumes :: [Volume]

An array of volume IDs.

dvtVolumes :: Iso' DescribeVolumesResponse [Volume]