Network.AWS.OpsWorks.V2013_02_18.DisassociateElasticIp

Description

Disassociates an Elastic IP address from its instance. The address remains registered with the stack. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data DisassociateElasticIp

Constructors

DisassociateElasticIp

Fields

_deirElasticIp :: Text

The Elastic IP address.

deirElasticIp :: Iso' DisassociateElasticIp Text

data DisassociateElasticIpResponse

Constructors

DisassociateElasticIpResponse