Network.AWS.OpsWorks.V2013_02_18.GetHostnameSuggestion

Description

Gets a generated host name for the specified layer, based on the current host name theme. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data GetHostnameSuggestion

Constructors

GetHostnameSuggestion

Fields

_ghsrLayerId :: Text

The layer ID.

ghsrLayerId :: Iso' GetHostnameSuggestion Text

data GetHostnameSuggestionResponse

Constructors

GetHostnameSuggestionResponse

Fields

_ghssHostname :: Maybe Text

The generated host name.

_ghssLayerId :: Maybe Text

The layer ID.

ghssLayerId :: Lens' GetHostnameSuggestionResponse (Maybe Text)

ghssHostname :: Lens' GetHostnameSuggestionResponse (Maybe Text)