Network.AWS.OpsWorks.V2013_02_18.Monadic

Description

AWS OpsWorks provides a simple and flexible way to create and manage stacks and applications. With AWS OpsWorks, you can provision AWS resources, manage their configuration, deploy applications to those resources, and monitor their health.

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.OpsWorks with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.OpsWorks

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.OpsWorks.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Network.AWS.OpsWorks.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

AssignVolume

Assigns one of the stack's registered Amazon EBS volumes to a specified instance. The volume must first be registered with the stack by calling RegisterVolume. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: AssignVolume

assignVolume

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
avVolumeId

-> State AssignVolume a

-> m AssignVolumeResponse

assignVolumeCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
avVolumeId

-> State AssignVolume a

-> m (Either ServiceEr AssignVolumeResponse)

AssociateElasticIp

Associates one of the stack's registered Elastic IP addresses with a specified instance. The address must first be registered with the stack by calling RegisterElasticIp. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: AssociateElasticIp

associateElasticIp

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
aeiElasticIp

-> State AssociateElasticIp a

-> m AssociateElasticIpResponse

associateElasticIpCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
aeiElasticIp

-> State AssociateElasticIp a

-> m (Either ServiceEr AssociateElasticIpResponse)

AttachElasticLoadBalancer

Attaches an Elastic Load Balancing load balancer to a specified layer. For more information, see Elastic Load Balancing. You must create the Elastic Load Balancing instance separately, by using the Elastic Load Balancing console, API, or CLI. For more information, see Elastic Load Balancing Developer Guide. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: AttachElasticLoadBalancer

attachElasticLoadBalancer

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
aelbElasticLoadBalancerName

-> Text
aelbLayerId

-> State AttachElasticLoadBalancer a

-> m AttachElasticLoadBalancerResponse

attachElasticLoadBalancerCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
aelbElasticLoadBalancerName

-> Text
aelbLayerId

-> State AttachElasticLoadBalancer a

-> m (Either ServiceEr AttachElasticLoadBalancerResponse)

CloneStack

Creates a clone of a specified stack. For more information, see Clone a Stack. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: CloneStack

cloneStack

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
csSourceStackId

-> Text
csServiceRoleArn

-> State CloneStack a

-> m CloneStackResponse

cloneStackCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
csSourceStackId

-> Text
csServiceRoleArn

-> State CloneStack a

-> m (Either ServiceEr CloneStackResponse)

CreateApp

Creates an app for a specified stack. For more information, see Creating Apps. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: CreateApp

createApp

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
caStackId

-> Text
caName

-> AppType
caType

-> State CreateApp a

-> m CreateAppResponse

createAppCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
caStackId

-> Text
caName

-> AppType
caType

-> State CreateApp a

-> m (Either ServiceEr CreateAppResponse)

CreateDeployment

Deploys a stack or app. App deployment generates a deploy event, which runs the associated recipes and passes them a JSON stack configuration object that includes information about the app. Stack deployment runs the deploy recipes but does not raise an event. For more information, see Deploying Apps and Run Stack Commands. Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: CreateDeployment

createDeployment

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cdStackId

-> DeploymentCommand
cdCommand

-> State CreateDeployment a

-> m CreateDeploymentResponse

createDeploymentCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cdStackId

-> DeploymentCommand
cdCommand

-> State CreateDeployment a

-> m (Either ServiceEr CreateDeploymentResponse)

CreateInstance

Creates an instance in a specified stack. For more information, see Adding an Instance to a Layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: CreateInstance

createInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ciStackId

-> [Text]
ciLayerIds

-> Text
ciInstanceType

-> State CreateInstance a

-> m CreateInstanceResponse

createInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ciStackId

-> [Text]
ciLayerIds

-> Text
ciInstanceType

-> State CreateInstance a

-> m (Either ServiceEr CreateInstanceResponse)

CreateLayer

Creates a layer. For more information, see How to Create a Layer. You should use CreateLayer for noncustom layer types such as PHP App Server only if the stack does not have an existing layer of that type. A stack can have at most one instance of each noncustom layer; if you attempt to create a second instance, CreateLayer fails. A stack can have an arbitrary number of custom layers, so you can call CreateLayer as many times as you like for that layer type. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: CreateLayer

createLayer

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
clStackId

-> LayerType
clType

-> Text
clName

-> Text
clShortname

-> State CreateLayer a

-> m CreateLayerResponse

createLayerCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
clStackId

-> LayerType
clType

-> Text
clName

-> Text
clShortname

-> State CreateLayer a

-> m (Either ServiceEr CreateLayerResponse)

CreateStack

Creates a new stack. For more information, see Create a New Stack. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: CreateStack

createStack

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cs1Name

-> Text
cs1Region

-> Text
cs1ServiceRoleArn

-> Text
cs1DefaultInstanceProfileArn

-> State CreateStack a

-> m CreateStackResponse

createStackCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cs1Name

-> Text
cs1Region

-> Text
cs1ServiceRoleArn

-> Text
cs1DefaultInstanceProfileArn

-> State CreateStack a

-> m (Either ServiceEr CreateStackResponse)

CreateUserProfile

Creates a new user profile. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: CreateUserProfile

createUserProfile

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cupIamUserArn

-> State CreateUserProfile a

-> m CreateUserProfileResponse

createUserProfileCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cupIamUserArn

-> State CreateUserProfile a

-> m (Either ServiceEr CreateUserProfileResponse)

DeleteApp

Deletes a specified app. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DeleteApp

deleteApp

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
daAppId

-> State DeleteApp a

-> m DeleteAppResponse

deleteAppCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
daAppId

-> State DeleteApp a

-> m (Either ServiceEr DeleteAppResponse)

DeleteInstance

Deletes a specified instance. You must stop an instance before you can delete it. For more information, see Deleting Instances. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DeleteInstance

deleteInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
diInstanceId

-> State DeleteInstance a

-> m DeleteInstanceResponse

deleteInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
diInstanceId

-> State DeleteInstance a

-> m (Either ServiceEr DeleteInstanceResponse)

DeleteLayer

Deletes a specified layer. You must first stop and then delete all associated instances. For more information, see How to Delete a Layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DeleteLayer

deleteLayer

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dlLayerId

-> State DeleteLayer a

-> m DeleteLayerResponse

deleteLayerCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dlLayerId

-> State DeleteLayer a

-> m (Either ServiceEr DeleteLayerResponse)

DeleteStack

Deletes a specified stack. You must first delete all instances, layers, and apps. For more information, see Shut Down a Stack. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DeleteStack

deleteStack

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ds1StackId

-> State DeleteStack a

-> m DeleteStackResponse

deleteStackCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ds1StackId

-> State DeleteStack a

-> m (Either ServiceEr DeleteStackResponse)

DeleteUserProfile

Deletes a user profile. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DeleteUserProfile

deleteUserProfile

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dupIamUserArn

-> State DeleteUserProfile a

-> m DeleteUserProfileResponse

deleteUserProfileCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dupIamUserArn

-> State DeleteUserProfile a

-> m (Either ServiceEr DeleteUserProfileResponse)

DeregisterElasticIp

Deregisters a specified Elastic IP address. The address can then be registered by another stack. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DeregisterElasticIp

deregisterElasticIp

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
deiElasticIp

-> State DeregisterElasticIp a

-> m DeregisterElasticIpResponse

deregisterElasticIpCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
deiElasticIp

-> State DeregisterElasticIp a

-> m (Either ServiceEr DeregisterElasticIpResponse)

DeregisterRdsDbInstance

Deregisters an Amazon RDS instance.

See: DeregisterRdsDbInstance

deregisterRdsDbInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
drdiRdsDbInstanceArn

-> State DeregisterRdsDbInstance a

-> m DeregisterRdsDbInstanceResponse

deregisterRdsDbInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
drdiRdsDbInstanceArn

-> State DeregisterRdsDbInstance a

-> m (Either ServiceEr DeregisterRdsDbInstanceResponse)

DeregisterVolume

Deregisters an Amazon EBS volume. The volume can then be registered by another stack. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DeregisterVolume

deregisterVolume

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dvVolumeId

-> State DeregisterVolume a

-> m DeregisterVolumeResponse

deregisterVolumeCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dvVolumeId

-> State DeregisterVolume a

-> m (Either ServiceEr DeregisterVolumeResponse)

DescribeApps

Requests a description of a specified set of apps. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeApps

describeApps :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeApps a -> m DescribeAppsResponse

describeAppsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeApps a -> m (Either ServiceEr DescribeAppsResponse)

DescribeCommands

Describes the results of specified commands. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeCommands

describeCommands :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeCommands a -> m DescribeCommandsResponse

describeCommandsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeCommands a -> m (Either ServiceEr DescribeCommandsResponse)

DescribeDeployments

Requests a description of a specified set of deployments. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeDeployments

describeDeployments :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeDeployments a -> m DescribeDeploymentsResponse

describeDeploymentsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeDeployments a -> m (Either ServiceEr DescribeDeploymentsResponse)

DescribeElasticIps

Describes Elastic IP addresses. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeElasticIps

describeElasticIps :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeElasticIps a -> m DescribeElasticIpsResponse

describeElasticIpsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeElasticIps a -> m (Either ServiceEr DescribeElasticIpsResponse)

DescribeElasticLoadBalancers

Describes a stack's Elastic Load Balancing instances. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeElasticLoadBalancers

describeElasticLoadBalancers :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeElasticLoadBalancers a -> m DescribeElasticLoadBalancersResponse

describeElasticLoadBalancersCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeElasticLoadBalancers a -> m (Either ServiceEr DescribeElasticLoadBalancersResponse)

DescribeInstances

Requests a description of a set of instances. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeInstances

describeInstances :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeInstances a -> m DescribeInstancesResponse

describeInstancesCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeInstances a -> m (Either ServiceEr DescribeInstancesResponse)

DescribeLayers

Requests a description of one or more layers in a specified stack. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeLayers

describeLayers :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeLayers a -> m DescribeLayersResponse

describeLayersCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeLayers a -> m (Either ServiceEr DescribeLayersResponse)

DescribeLoadBasedAutoScaling

Describes load-based auto scaling configurations for specified layers. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeLoadBasedAutoScaling

describeLoadBasedAutoScaling

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Text]
dlbasLayerIds

-> State DescribeLoadBasedAutoScaling a

-> m DescribeLoadBasedAutoScalingResponse

describeLoadBasedAutoScalingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Text]
dlbasLayerIds

-> State DescribeLoadBasedAutoScaling a

-> m (Either ServiceEr DescribeLoadBasedAutoScalingResponse)

DescribeMyUserProfile

Describes a user's SSH information. Required Permissions: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeMyUserProfile

describeMyUserProfile :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeMyUserProfile a -> m DescribeMyUserProfileResponse

describeMyUserProfileCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeMyUserProfile a -> m (Either ServiceEr DescribeMyUserProfileResponse)

DescribePermissions

Describes the permissions for a specified stack. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribePermissions

describePermissions :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribePermissions a -> m DescribePermissionsResponse

describePermissionsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribePermissions a -> m (Either ServiceEr DescribePermissionsResponse)

DescribeRaidArrays

Describe an instance's RAID arrays. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeRaidArrays

describeRaidArrays :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeRaidArrays a -> m DescribeRaidArraysResponse

describeRaidArraysCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeRaidArrays a -> m (Either ServiceEr DescribeRaidArraysResponse)

DescribeRdsDbInstances

Describes Amazon RDS instances.

See: DescribeRdsDbInstances

describeRdsDbInstances

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
drdi1StackId

-> State DescribeRdsDbInstances a

-> m DescribeRdsDbInstancesResponse

describeRdsDbInstancesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
drdi1StackId

-> State DescribeRdsDbInstances a

-> m (Either ServiceEr DescribeRdsDbInstancesResponse)

DescribeServiceErrors

Describes AWS OpsWorks service errors. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeServiceErrors

describeServiceErrors :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeServiceErrors a -> m DescribeServiceErrorsResponse

describeServiceErrorsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeServiceErrors a -> m (Either ServiceEr DescribeServiceErrorsResponse)

DescribeStackSummary

Describes the number of layers and apps in a specified stack, and the number of instances in each state, such as running_setup or online. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeStackSummary

describeStackSummary

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dssStackId

-> State DescribeStackSummary a

-> m DescribeStackSummaryResponse

describeStackSummaryCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dssStackId

-> State DescribeStackSummary a

-> m (Either ServiceEr DescribeStackSummaryResponse)

DescribeStacks

Requests a description of one or more stacks. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeStacks

describeStacks :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeStacks a -> m DescribeStacksResponse

describeStacksCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeStacks a -> m (Either ServiceEr DescribeStacksResponse)

DescribeTimeBasedAutoScaling

Describes time-based auto scaling configurations for specified instances. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeTimeBasedAutoScaling

describeTimeBasedAutoScaling

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Text]
dtbasInstanceIds

-> State DescribeTimeBasedAutoScaling a

-> m DescribeTimeBasedAutoScalingResponse

describeTimeBasedAutoScalingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Text]
dtbasInstanceIds

-> State DescribeTimeBasedAutoScaling a

-> m (Either ServiceEr DescribeTimeBasedAutoScalingResponse)

DescribeUserProfiles

Describe specified users. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeUserProfiles

describeUserProfiles :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeUserProfiles a -> m DescribeUserProfilesResponse

describeUserProfilesCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeUserProfiles a -> m (Either ServiceEr DescribeUserProfilesResponse)

DescribeVolumes

Describes an instance's Amazon EBS volumes. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DescribeVolumes

describeVolumes :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeVolumes a -> m DescribeVolumesResponse

describeVolumesCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeVolumes a -> m (Either ServiceEr DescribeVolumesResponse)

DetachElasticLoadBalancer

Detaches a specified Elastic Load Balancing instance from its layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DetachElasticLoadBalancer

detachElasticLoadBalancer

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
delb1ElasticLoadBalancerName

-> Text
delb1LayerId

-> State DetachElasticLoadBalancer a

-> m DetachElasticLoadBalancerResponse

detachElasticLoadBalancerCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
delb1ElasticLoadBalancerName

-> Text
delb1LayerId

-> State DetachElasticLoadBalancer a

-> m (Either ServiceEr DetachElasticLoadBalancerResponse)

DisassociateElasticIp

Disassociates an Elastic IP address from its instance. The address remains registered with the stack. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: DisassociateElasticIp

disassociateElasticIp

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dei2ElasticIp

-> State DisassociateElasticIp a

-> m DisassociateElasticIpResponse

disassociateElasticIpCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dei2ElasticIp

-> State DisassociateElasticIp a

-> m (Either ServiceEr DisassociateElasticIpResponse)

GetHostnameSuggestion

Gets a generated host name for the specified layer, based on the current host name theme. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: GetHostnameSuggestion

getHostnameSuggestion

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ghsLayerId

-> State GetHostnameSuggestion a

-> m GetHostnameSuggestionResponse

getHostnameSuggestionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ghsLayerId

-> State GetHostnameSuggestion a

-> m (Either ServiceEr GetHostnameSuggestionResponse)

RebootInstance

Reboots a specified instance. For more information, see Starting, Stopping, and Rebooting Instances. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: RebootInstance

rebootInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
riInstanceId

-> State RebootInstance a

-> m RebootInstanceResponse

rebootInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
riInstanceId

-> State RebootInstance a

-> m (Either ServiceEr RebootInstanceResponse)

RegisterElasticIp

Registers an Elastic IP address with a specified stack. An address can be registered with only one stack at a time. If the address is already registered, you must first deregister it by calling DeregisterElasticIp. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: RegisterElasticIp

registerElasticIp

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
reiElasticIp

-> Text
reiStackId

-> State RegisterElasticIp a

-> m RegisterElasticIpResponse

registerElasticIpCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
reiElasticIp

-> Text
reiStackId

-> State RegisterElasticIp a

-> m (Either ServiceEr RegisterElasticIpResponse)

RegisterRdsDbInstance

Registers an Amazon RDS instance with a stack.

See: RegisterRdsDbInstance

registerRdsDbInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
rrdiStackId

-> Text
rrdiRdsDbInstanceArn

-> Text
rrdiDbUser

-> Text
rrdiDbPassword

-> State RegisterRdsDbInstance a

-> m RegisterRdsDbInstanceResponse

registerRdsDbInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
rrdiStackId

-> Text
rrdiRdsDbInstanceArn

-> Text
rrdiDbUser

-> Text
rrdiDbPassword

-> State RegisterRdsDbInstance a

-> m (Either ServiceEr RegisterRdsDbInstanceResponse)

RegisterVolume

Registers an Amazon EBS volume with a specified stack. A volume can be registered with only one stack at a time. If the volume is already registered, you must first deregister it by calling DeregisterVolume. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: RegisterVolume

registerVolume

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
rvStackId

-> State RegisterVolume a

-> m RegisterVolumeResponse

registerVolumeCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
rvStackId

-> State RegisterVolume a

-> m (Either ServiceEr RegisterVolumeResponse)

SetLoadBasedAutoScaling

Specify the load-based auto scaling configuration for a specified layer. For more information, see Managing Load with Time-based and Load-based Instances. To use load-based auto scaling, you must create a set of load-based auto scaling instances. Load-based auto scaling operates only on the instances from that set, so you must ensure that you have created enough instances to handle the maximum anticipated load. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: SetLoadBasedAutoScaling

setLoadBasedAutoScaling

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
slbasLayerId

-> State SetLoadBasedAutoScaling a

-> m SetLoadBasedAutoScalingResponse

setLoadBasedAutoScalingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
slbasLayerId

-> State SetLoadBasedAutoScaling a

-> m (Either ServiceEr SetLoadBasedAutoScalingResponse)

SetPermission

Specifies a user's permissions. For more information, see Security and Permissions. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: SetPermission

setPermission

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
spStackId

-> Text
spIamUserArn

-> State SetPermission a

-> m SetPermissionResponse

setPermissionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
spStackId

-> Text
spIamUserArn

-> State SetPermission a

-> m (Either ServiceEr SetPermissionResponse)

SetTimeBasedAutoScaling

Specify the time-based auto scaling configuration for a specified instance. For more information, see Managing Load with Time-based and Load-based Instances. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: SetTimeBasedAutoScaling

setTimeBasedAutoScaling

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
stbasInstanceId

-> State SetTimeBasedAutoScaling a

-> m SetTimeBasedAutoScalingResponse

setTimeBasedAutoScalingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
stbasInstanceId

-> State SetTimeBasedAutoScaling a

-> m (Either ServiceEr SetTimeBasedAutoScalingResponse)

StartInstance

Starts a specified instance. For more information, see Starting, Stopping, and Rebooting Instances. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: StartInstance

startInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
siInstanceId

-> State StartInstance a

-> m StartInstanceResponse

startInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
siInstanceId

-> State StartInstance a

-> m (Either ServiceEr StartInstanceResponse)

StartStack

Starts a stack's instances. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: StartStack

startStack

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ss1StackId

-> State StartStack a

-> m StartStackResponse

startStackCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ss1StackId

-> State StartStack a

-> m (Either ServiceEr StartStackResponse)

StopInstance

Stops a specified instance. When you stop a standard instance, the data disappears and must be reinstalled when you restart the instance. You can stop an Amazon EBS-backed instance without losing data. For more information, see Starting, Stopping, and Rebooting Instances. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: StopInstance

stopInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
si1InstanceId

-> State StopInstance a

-> m StopInstanceResponse

stopInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
si1InstanceId

-> State StopInstance a

-> m (Either ServiceEr StopInstanceResponse)

StopStack

Stops a specified stack. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: StopStack

stopStack

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ss2StackId

-> State StopStack a

-> m StopStackResponse

stopStackCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ss2StackId

-> State StopStack a

-> m (Either ServiceEr StopStackResponse)

UnassignVolume

Unassigns an assigned Amazon EBS volume. The volume remains registered with the stack. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UnassignVolume

unassignVolume

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uvVolumeId

-> State UnassignVolume a

-> m UnassignVolumeResponse

unassignVolumeCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uvVolumeId

-> State UnassignVolume a

-> m (Either ServiceEr UnassignVolumeResponse)

UpdateApp

Updates a specified app. Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UpdateApp

updateApp

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uaAppId

-> State UpdateApp a

-> m UpdateAppResponse

updateAppCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uaAppId

-> State UpdateApp a

-> m (Either ServiceEr UpdateAppResponse)

UpdateElasticIp

Updates a registered Elastic IP address's name. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UpdateElasticIp

updateElasticIp

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ueiElasticIp

-> State UpdateElasticIp a

-> m UpdateElasticIpResponse

updateElasticIpCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ueiElasticIp

-> State UpdateElasticIp a

-> m (Either ServiceEr UpdateElasticIpResponse)

UpdateInstance

Updates a specified instance. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UpdateInstance

updateInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uiInstanceId

-> State UpdateInstance a

-> m UpdateInstanceResponse

updateInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uiInstanceId

-> State UpdateInstance a

-> m (Either ServiceEr UpdateInstanceResponse)

UpdateLayer

Updates a specified layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UpdateLayer

updateLayer

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ulLayerId

-> State UpdateLayer a

-> m UpdateLayerResponse

updateLayerCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ulLayerId

-> State UpdateLayer a

-> m (Either ServiceEr UpdateLayerResponse)

UpdateMyUserProfile

Updates a user's SSH public key. Required Permissions: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UpdateMyUserProfile

updateMyUserProfile :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State UpdateMyUserProfile a -> m UpdateMyUserProfileResponse

updateMyUserProfileCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State UpdateMyUserProfile a -> m (Either ServiceEr UpdateMyUserProfileResponse)

UpdateRdsDbInstance

Updates an Amazon RDS instance.

See: UpdateRdsDbInstance

updateRdsDbInstance

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
urdiRdsDbInstanceArn

-> State UpdateRdsDbInstance a

-> m UpdateRdsDbInstanceResponse

updateRdsDbInstanceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
urdiRdsDbInstanceArn

-> State UpdateRdsDbInstance a

-> m (Either ServiceEr UpdateRdsDbInstanceResponse)

UpdateStack

Updates a specified stack. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UpdateStack

updateStack

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
usStackId

-> State UpdateStack a

-> m UpdateStackResponse

updateStackCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
usStackId

-> State UpdateStack a

-> m (Either ServiceEr UpdateStackResponse)

UpdateUserProfile

Updates a specified user profile. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UpdateUserProfile

updateUserProfile

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uupIamUserArn

-> State UpdateUserProfile a

-> m UpdateUserProfileResponse

updateUserProfileCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uupIamUserArn

-> State UpdateUserProfile a

-> m (Either ServiceEr UpdateUserProfileResponse)

UpdateVolume

Updates an Amazon EBS volume's name or mount point. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

See: UpdateVolume

updateVolume

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uv1VolumeId

-> State UpdateVolume a

-> m UpdateVolumeResponse

updateVolumeCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uv1VolumeId

-> State UpdateVolume a

-> m (Either ServiceEr UpdateVolumeResponse)

Re-exported

module Network.AWS.OpsWorks.V2013_02_18

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()