Network.AWS.OpsWorks.V2013_02_18.RegisterElasticIp

Description

Registers an Elastic IP address with a specified stack. An address can be registered with only one stack at a time. If the address is already registered, you must first deregister it by calling DeregisterElasticIp. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data RegisterElasticIp

Constructors

RegisterElasticIp

Fields

_reirElasticIp :: Text

The Elastic IP address.

_reirStackId :: Text

The stack ID.

reirStackId :: Lens' RegisterElasticIp Text

reirElasticIp :: Lens' RegisterElasticIp Text

data RegisterElasticIpResponse

Constructors

RegisterElasticIpResponse

Fields

_reisElasticIp :: Maybe Text

The Elastic IP address.

reisElasticIp :: Iso' RegisterElasticIpResponse (Maybe Text)