Network.AWS.OpsWorks.V2013_02_18.RegisterRdsDbInstance

Description

Registers an Amazon RDS instance with a stack.

Request

data RegisterRdsDbInstance

Request constructor

mkRegisterRdsDbInstance

Arguments

:: Text
rrdiStackId

-> Text
rrdiRdsDbInstanceArn

-> Text
rrdiDbUser

-> Text
rrdiDbPassword

-> RegisterRdsDbInstance

Smart constructor for the minimum required parameters to construct a valid RegisterRdsDbInstance request.

Request lenses

rrdiStackId :: Lens' RegisterRdsDbInstance Text

The stack ID.

rrdiRdsDbInstanceArn :: Lens' RegisterRdsDbInstance Text

The Amazon RDS instance's ARN.

rrdiDbUser :: Lens' RegisterRdsDbInstance Text

The database's master user name.

rrdiDbPassword :: Lens' RegisterRdsDbInstance Text

The database password.

Response

data RegisterRdsDbInstanceResponse

Response constructor

mkRegisterRdsDbInstanceResponse :: RegisterRdsDbInstanceResponse

Smart constructor for the minimum required parameters to construct a valid RegisterRdsDbInstanceResponse response.

This constructor is provided for convenience and testing purposes.