Network.AWS.OpsWorks.V2013_02_18.RegisterVolume

Description

Registers an Amazon EBS volume with a specified stack. A volume can be registered with only one stack at a time. If the volume is already registered, you must first deregister it by calling DeregisterVolume. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

registerVolume

Arguments

:: Text
_rvrStackId

-> RegisterVolume

Minimum specification for a RegisterVolume request.

data RegisterVolume

Constructors

RegisterVolume

Fields

_rvrStackId :: Text

The stack ID.

_rvrEc2VolumeId :: Maybe Text

The Amazon EBS volume ID.

rvrStackId :: Lens' RegisterVolume Text

rvrEc2VolumeId :: Lens' RegisterVolume (Maybe Text)

data RegisterVolumeResponse

Constructors

RegisterVolumeResponse

Fields

_rvsVolumeId :: Maybe Text

The volume ID.

rvsVolumeId :: Iso' RegisterVolumeResponse (Maybe Text)