Network.AWS.OpsWorks.V2013_02_18.SetLoadBasedAutoScaling

Description

Specify the load-based auto scaling configuration for a specified layer. For more information, see Managing Load with Time-based and Load-based Instances. To use load-based auto scaling, you must create a set of load-based auto scaling instances. Load-based auto scaling operates only on the instances from that set, so you must ensure that you have created enough instances to handle the maximum anticipated load. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

setLoadBasedAutoScaling

Arguments

:: Text
_slbasrLayerId

-> SetLoadBasedAutoScaling

Minimum specification for a SetLoadBasedAutoScaling request.

data SetLoadBasedAutoScaling

Constructors

SetLoadBasedAutoScaling

Fields

_slbasrLayerId :: Text

The layer ID.

_slbasrUpScaling :: Maybe AutoScalingThresholds

An AutoScalingThresholds object with the upscaling threshold configuration. If the load exceeds these thresholds for a specified amount of time, AWS OpsWorks starts a specified number of instances.

_slbasrDownScaling :: Maybe AutoScalingThresholds

An AutoScalingThresholds object with the downscaling threshold configuration. If the load falls below these thresholds for a specified amount of time, AWS OpsWorks stops a specified number of instances.

_slbasrEnable :: Maybe Bool

Enables load-based auto scaling for the layer.

slbasrUpScaling :: Lens' SetLoadBasedAutoScaling (Maybe AutoScalingThresholds)

slbasrLayerId :: Lens' SetLoadBasedAutoScaling Text

slbasrEnable :: Lens' SetLoadBasedAutoScaling (Maybe Bool)

slbasrDownScaling :: Lens' SetLoadBasedAutoScaling (Maybe AutoScalingThresholds)

data SetLoadBasedAutoScalingResponse

Constructors

SetLoadBasedAutoScalingResponse