Network.AWS.OpsWorks.V2013_02_18.SetPermission

Description

Specifies a user's permissions. For more information, see Security and Permissions. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

setPermission

Arguments

:: Text
_sprIamUserArn

-> Text
_sprStackId

-> SetPermission

Minimum specification for a SetPermission request.

data SetPermission

Constructors

SetPermission

Fields

_sprIamUserArn :: Text

The user's IAM ARN.

_sprStackId :: Text

The stack ID.

_sprAllowSudo :: Maybe Bool

The user is allowed to use sudo to elevate privileges.

_sprAllowSsh :: Maybe Bool

The user is allowed to use SSH to communicate with the instance.

_sprLevel :: Maybe Text

The user's permission level, which must be set to one of the following strings. You cannot set your own permissions level. deny show deploy manage iam_only For more information on the permissions associated with these levels, see Managing User Permissions.

sprStackId :: Lens' SetPermission Text

sprLevel :: Lens' SetPermission (Maybe Text)

sprIamUserArn :: Lens' SetPermission Text

sprAllowSudo :: Lens' SetPermission (Maybe Bool)

sprAllowSsh :: Lens' SetPermission (Maybe Bool)

data SetPermissionResponse

Constructors

SetPermissionResponse