Network.AWS.OpsWorks.V2013_02_18.StopStack

Description

Stops a specified stack. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data StopStack

Constructors

StopStack

Fields

_ssssssssszStackId :: Text

The stack ID.

ssssssssszStackId :: Iso' StopStack Text

data StopStackResponse

Constructors

StopStackResponse