Network.AWS.OpsWorks.V2013_02_18.Types

Description

AWS OpsWorks provides a simple and flexible way to create and manage stacks and applications. With AWS OpsWorks, you can provision AWS resources, manage their configuration, deploy applications to those resources, and monitor their health.

Service

data OpsWorks

Supported version (2013-02-18) of the AWS OpsWorks service.

AppAttributesKeys

data AppAttributesKeys

Constructors

AppAttributesKeysAutoBundleOnDeploy
AutoBundleOnDeploy

AppAttributesKeysDocumentRoot
DocumentRoot

AppAttributesKeysRailsEnv
RailsEnv

AppType

data AppType

Constructors

AppTypeNodejs
nodejs

AppTypeOther
other

AppTypePhp
php

AppTypeRails
rails

AppTypeStatic
static

Architecture

data Architecture

Constructors

ArchitectureI386
i386

ArchitectureX8664
x86_64

AutoScalingType

data AutoScalingType

Constructors

AutoScalingTypeLoad
load

AutoScalingTypeTimer
timer

DeploymentCommandName

data DeploymentCommandName

Constructors

DeploymentCommandNameDeploy
deploy

DeploymentCommandNameExecuteRecipes
execute_recipes

DeploymentCommandNameInstallDependencies
install_dependencies

DeploymentCommandNameRestart
restart

DeploymentCommandNameRollback
rollback

DeploymentCommandNameStart
start

DeploymentCommandNameStop
stop

DeploymentCommandNameUndeploy
undeploy

DeploymentCommandNameUpdateCustomCookbooks
update_custom_cookbooks

DeploymentCommandNameUpdateDependencies
update_dependencies

LayerAttributesKeys

data LayerAttributesKeys

Constructors

LayerAttributesKeysBundlerVersion
BundlerVersion

LayerAttributesKeysEnableHaproxyStats
EnableHaproxyStats

LayerAttributesKeysGangliaPassword
GangliaPassword

LayerAttributesKeysGangliaUrl
GangliaUrl

LayerAttributesKeysGangliaUser
GangliaUser

LayerAttributesKeysHaproxyHealthCheckMethod
HaproxyHealthCheckMethod

LayerAttributesKeysHaproxyHealthCheckUrl
HaproxyHealthCheckUrl

LayerAttributesKeysHaproxyStatsPassword
HaproxyStatsPassword

LayerAttributesKeysHaproxyStatsUrl
HaproxyStatsUrl

LayerAttributesKeysHaproxyStatsUser
HaproxyStatsUser

LayerAttributesKeysJavaAppServer
JavaAppServer

LayerAttributesKeysJavaAppServerVersion
JavaAppServerVersion

LayerAttributesKeysJvm
Jvm

LayerAttributesKeysJvmOptions
JvmOptions

LayerAttributesKeysJvmVersion
JvmVersion

LayerAttributesKeysManageBundler
ManageBundler

LayerAttributesKeysMemcachedMemory
MemcachedMemory

LayerAttributesKeysMysqlRootPassword
MysqlRootPassword

LayerAttributesKeysMysqlRootPasswordUbiquitous
MysqlRootPasswordUbiquitous

LayerAttributesKeysNodejsVersion
NodejsVersion

LayerAttributesKeysPassengerVersion
PassengerVersion

LayerAttributesKeysRailsStack
RailsStack

LayerAttributesKeysRubyVersion
RubyVersion

LayerAttributesKeysRubygemsVersion
RubygemsVersion

LayerType

data LayerType

Constructors

LayerTypeCustom
custom

LayerTypeDbMaster
db-master

LayerTypeLb
lb

LayerTypeMemcached
memcached

LayerTypeMonitoringMaster
monitoring-master

LayerTypeNodejsApp
nodejs-app

LayerTypePhpApp
php-app

LayerTypeRailsApp
rails-app

LayerTypeWeb
web

RootDeviceType

data RootDeviceType

Constructors

RootDeviceTypeEbs
ebs

RootDeviceTypeInstanceStore
instance-store

SourceType

data SourceType

Constructors

SourceTypeArchive
archive

SourceTypeGit
git

SourceTypeS3
s3

SourceTypeSvn
svn

StackAttributesKeys

data StackAttributesKeys

Constructors

StackAttributesKeysColor
Color

App

data App

A description of the app.

mkApp :: App

Smart constructor for the minimum required fields to construct a valid App data type to populate a request.

aAppId :: Lens' App (Maybe Text)

The app ID.

aStackId :: Lens' App (Maybe Text)

The app stack ID.

aShortname :: Lens' App (Maybe Text)

The app's short name.

aName :: Lens' App (Maybe Text)

The app name.

aDescription :: Lens' App (Maybe Text)

A description of the app.

aDataSources :: Lens' App [DataSource]

The app's data sources.

aType :: Lens' App (Maybe AppType)

The app type.

aAppSource :: Lens' App (Maybe Source)

A Source object that describes the app repository.

aDomains :: Lens' App [Text]

The app vhost settings with multiple domains separated by commas. For example: 'www.example.com, example.com'.

aEnableSsl :: Lens' App (Maybe Bool)

Whether to enable SSL for the app.

aSslConfiguration :: Lens' App (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

aAttributes :: Lens' App (Map AppAttributesKeys Text)

The stack attributes.

aCreatedAt :: Lens' App (Maybe Text)

When the app was created.

AutoScalingThresholds

data AutoScalingThresholds

A LoadBasedAutoscalingInstruction object that describes the upscaling configuration, which defines how and when AWS OpsWorks increases the number of instances.

mkAutoScalingThresholds :: AutoScalingThresholds

Smart constructor for the minimum required fields to construct a valid AutoScalingThresholds data type to populate a request.

astInstanceCount :: Lens' AutoScalingThresholds (Maybe Integer)

The number of instances to add or remove when the load exceeds a threshold.

astThresholdsWaitTime :: Lens' AutoScalingThresholds (Maybe Integer)

The amount of time, in minutes, that the load must exceed a threshold before more instances are added or removed.

astIgnoreMetricsTime :: Lens' AutoScalingThresholds (Maybe Integer)

The amount of time (in minutes) after a scaling event occurs that AWS OpsWorks should ignore metrics and not raise any additional scaling events. For example, AWS OpsWorks adds new instances following an upscaling event but the instances won't start reducing the load until they have been booted and configured. There is no point in raising additional scaling events during that operation, which typically takes several minutes. IgnoreMetricsTime allows you to direct AWS OpsWorks to not raise any scaling events long enough to get the new instances online.

astCpuThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The CPU utilization threshold, as a percent of the available CPU.

astMemoryThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The memory utilization threshold, as a percent of the available memory.

astLoadThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The load threshold. For more information about how load is computed, see Load (computing).

ChefConfiguration

data ChefConfiguration

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version on Chef 11.10 stacks. For more information, see Create a New Stack.

mkChefConfiguration :: ChefConfiguration

Smart constructor for the minimum required fields to construct a valid ChefConfiguration data type to populate a request.

ccManageBerkshelf :: Lens' ChefConfiguration (Maybe Bool)

Whether to enable Berkshelf.

ccBerkshelfVersion :: Lens' ChefConfiguration (Maybe Text)

The Berkshelf version.

Command

data Command

Describes a command.

mkCommand :: Command

Smart constructor for the minimum required fields to construct a valid Command data type to populate a request.

cCommandId :: Lens' Command (Maybe Text)

The command ID.

cInstanceId :: Lens' Command (Maybe Text)

The ID of the instance where the command was executed.

cDeploymentId :: Lens' Command (Maybe Text)

The command deployment ID.

cCreatedAt :: Lens' Command (Maybe Text)

Date and time when the command was run.

cAcknowledgedAt :: Lens' Command (Maybe Text)

Date and time when the command was acknowledged.

cCompletedAt :: Lens' Command (Maybe Text)

Date when the command completed.

cStatus :: Lens' Command (Maybe Text)

The command status: failed successful skipped pending.

cExitCode :: Lens' Command (Maybe Integer)

The command exit code.

cLogUrl :: Lens' Command (Maybe Text)

The URL of the command log.

cType :: Lens' Command (Maybe Text)

The command type: deploy rollback start stop restart undeploy update_dependencies install_dependencies update_custom_cookbooks execute_recipes.

DataSource

data DataSource

Describes an app's data source.

mkDataSource :: DataSource

Smart constructor for the minimum required fields to construct a valid DataSource data type to populate a request.

dsType :: Lens' DataSource (Maybe Text)

The data source's type, AutoSelectOpsworksMysqlInstance, OpsworksMysqlInstance, or RdsDbInstance.

dsArn :: Lens' DataSource (Maybe Text)

The data source's ARN.

dsDatabaseName :: Lens' DataSource (Maybe Text)

The database name.

Deployment

data Deployment

Describes a deployment of a stack or app.

mkDeployment :: Deployment

Smart constructor for the minimum required fields to construct a valid Deployment data type to populate a request.

dDeploymentId :: Lens' Deployment (Maybe Text)

The deployment ID.

dStackId :: Lens' Deployment (Maybe Text)

The stack ID.

dAppId :: Lens' Deployment (Maybe Text)

The app ID.

dCreatedAt :: Lens' Deployment (Maybe Text)

Date when the deployment was created.

dCompletedAt :: Lens' Deployment (Maybe Text)

Date when the deployment completed.

dDuration :: Lens' Deployment (Maybe Integer)

The deployment duration.

dIamUserArn :: Lens' Deployment (Maybe Text)

The user's IAM ARN.

dComment :: Lens' Deployment (Maybe Text)

A user-defined comment.

dCommand :: Lens' Deployment (Maybe DeploymentCommand)

Used to specify a deployment operation.

dStatus :: Lens' Deployment (Maybe Text)

The deployment status: running successful failed.

dCustomJson :: Lens' Deployment (Maybe Text)

A string that contains user-defined custom JSON. It is used to override the corresponding default stack configuration JSON values for stack. The string should be in the following format and must escape characters such as '"'.: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration JSON.

dInstanceIds :: Lens' Deployment [Text]

The IDs of the target instances.

DeploymentCommand

data DeploymentCommand

A DeploymentCommand object that specifies the deployment command and any associated arguments.

mkDeploymentCommand

Arguments

:: DeploymentCommandName
dcName

-> DeploymentCommand

Smart constructor for the minimum required fields to construct a valid DeploymentCommand data type to populate a request.

dcName :: Lens' DeploymentCommand DeploymentCommandName

Specifies the operation. You can specify only one command. For stacks, the following commands are available: execute_recipes: Execute one or more recipes. To specify the recipes, set an Args parameter named recipes to the list of recipes to be executed. For example, to execute phpapp::appsetup, set Args to {"recipes":["phpapp::appsetup"]}. install_dependencies: Install the stack's dependencies. update_custom_cookbooks: Update the stack's custom cookbooks. update_dependencies: Update the stack's dependencies. For apps, the following commands are available: deploy: Deploy an app. Rails apps have an optional Args parameter named migrate. Set Args to {"migrate":["true"]} to migrate the database. The default setting is {"migrate":["false"]}. rollback Roll the app back to the previous version. When you update an app, AWS OpsWorks stores the previous version, up to a maximum of five versions. You can use this command to roll an app back as many as four versions. start: Start the app's web or application server. stop: Stop the app's web or application server. restart: Restart the app's web or application server. undeploy: Undeploy the app.

dcArgs :: Lens' DeploymentCommand (Map Text [Text])

The arguments of those commands that take arguments. It should be set to a JSON object with the following format: {"arg_name":["value1", "value2", ...]}.

ElasticIp

data ElasticIp

Describes an Elastic IP address.

mkElasticIp :: ElasticIp

Smart constructor for the minimum required fields to construct a valid ElasticIp data type to populate a request.

eiIp :: Lens' ElasticIp (Maybe Text)

The IP address.

eiName :: Lens' ElasticIp (Maybe Text)

The name.

eiDomain :: Lens' ElasticIp (Maybe Text)

The domain.

eiRegion :: Lens' ElasticIp (Maybe Text)

The AWS region. For more information, see Regions and Endpoints.

eiInstanceId :: Lens' ElasticIp (Maybe Text)

The ID of the instance that the address is attached to.

ElasticLoadBalancer

data ElasticLoadBalancer

Describes an Elastic Load Balancing instance.

mkElasticLoadBalancer :: ElasticLoadBalancer

Smart constructor for the minimum required fields to construct a valid ElasticLoadBalancer data type to populate a request.

elbElasticLoadBalancerName :: Lens' ElasticLoadBalancer (Maybe Text)

The Elastic Load Balancing instance's name.

elbRegion :: Lens' ElasticLoadBalancer (Maybe Text)

The instance's AWS region.

elbDnsName :: Lens' ElasticLoadBalancer (Maybe Text)

The instance's public DNS name.

elbStackId :: Lens' ElasticLoadBalancer (Maybe Text)

The ID of the stack that the instance is associated with.

elbLayerId :: Lens' ElasticLoadBalancer (Maybe Text)

The ID of the layer that the instance is attached to.

elbVpcId :: Lens' ElasticLoadBalancer (Maybe Text)

The VPC ID.

elbAvailabilityZones :: Lens' ElasticLoadBalancer [Text]

A list of Availability Zones.

elbSubnetIds :: Lens' ElasticLoadBalancer [Text]

A list of subnet IDs, if the stack is running in a VPC.

elbEc2InstanceIds :: Lens' ElasticLoadBalancer [Text]

A list of the EC2 instances that the Elastic Load Balancing instance is managing traffic for.

Instance

data Instance

Describes an instance.

mkInstance :: Instance

Smart constructor for the minimum required fields to construct a valid Instance data type to populate a request.

iInstanceId :: Lens' Instance (Maybe Text)

The instance ID.

iEc2InstanceId :: Lens' Instance (Maybe Text)

The ID of the associated Amazon EC2 instance.

iVirtualizationType :: Lens' Instance (Maybe Text)

The instance's virtualization type, paravirtual or hvm.

iHostname :: Lens' Instance (Maybe Text)

The instance host name.

iStackId :: Lens' Instance (Maybe Text)

The stack ID.

iLayerIds :: Lens' Instance [Text]

An array containing the instance layer IDs.

iSecurityGroupIds :: Lens' Instance [Text]

An array containing the instance security group IDs.

iInstanceType :: Lens' Instance (Maybe Text)

The instance type. AWS OpsWorks supports all instance types except Cluster Compute, Cluster GPU, and High Memory Cluster. For more information, see Instance Families and Types. The parameter values that specify the various types are in the API Name column of the Available Instance Types table.

iInstanceProfileArn :: Lens' Instance (Maybe Text)

The ARN of the instance's IAM profile. For more information about IAM ARNs, see Using Identifiers.

iStatus :: Lens' Instance (Maybe Text)

The instance status: requested booting running_setup online setup_failed start_failed terminating terminated stopped connection_lost.

iOs :: Lens' Instance (Maybe Text)

The instance operating system.

iAmiId :: Lens' Instance (Maybe Text)

A custom AMI ID to be used to create the instance. The AMI should be based on one of the standard AWS OpsWorks APIs: Amazon Linux or Ubuntu 12.04 LTS. For more information, see Instances.

iAvailabilityZone :: Lens' Instance (Maybe Text)

The instance Availability Zone. For more information, see Regions and Endpoints.

iSubnetId :: Lens' Instance (Maybe Text)

The instance's subnet ID, if the stack is running in a VPC.

iPublicDns :: Lens' Instance (Maybe Text)

The instance public DNS name.

iPrivateDns :: Lens' Instance (Maybe Text)

The instance private DNS name.

iPublicIp :: Lens' Instance (Maybe Text)

The instance public IP address.

iPrivateIp :: Lens' Instance (Maybe Text)

The instance private IP address.

iElasticIp :: Lens' Instance (Maybe Text)

The instance Elastic IP address .

iAutoScalingType :: Lens' Instance (Maybe AutoScalingType)

The instance's auto scaling type, which has three possible values: AlwaysRunning: A 24/7 instance, which is not affected by auto scaling. TimeBasedAutoScaling: A time-based auto scaling instance, which is started and stopped based on a specified schedule. LoadBasedAutoScaling: A load-based auto scaling instance, which is started and stopped based on load metrics.

iSshKeyName :: Lens' Instance (Maybe Text)

The instance SSH key name.

iSshHostRsaKeyFingerprint :: Lens' Instance (Maybe Text)

The SSH key's RSA fingerprint.

iSshHostDsaKeyFingerprint :: Lens' Instance (Maybe Text)

The SSH key's DSA fingerprint.

iCreatedAt :: Lens' Instance (Maybe Text)

The time that the instance was created.

iLastServiceErrorId :: Lens' Instance (Maybe Text)

The ID of the last service error. For more information, call DescribeServiceErrors.

iArchitecture :: Lens' Instance (Maybe Architecture)

The instance architecture, "i386" or "x86_64".

iRootDeviceType :: Lens' Instance (Maybe RootDeviceType)

The instance root device type. For more information, see Storage for the Root Device.

iRootDeviceVolumeId :: Lens' Instance (Maybe Text)

The root device volume ID.

iInstallUpdatesOnBoot :: Lens' Instance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances. We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

iEbsOptimized :: Lens' Instance (Maybe Bool)

Whether this is an Amazon EBS-optimized instance.

InstancesCount

data InstancesCount

An InstancesCount object with the number of instances in each status.

mkInstancesCount :: InstancesCount

Smart constructor for the minimum required fields to construct a valid InstancesCount data type to populate a request.

icBooting :: Lens' InstancesCount (Maybe Integer)

The number of instances with booting status.

icConnectionLost :: Lens' InstancesCount (Maybe Integer)

The number of instances with connection_lost status.

icOnline :: Lens' InstancesCount (Maybe Integer)

The number of instances with online status.

icPending :: Lens' InstancesCount (Maybe Integer)

The number of instances with pending status.

icRebooting :: Lens' InstancesCount (Maybe Integer)

The number of instances with rebooting status.

icRequested :: Lens' InstancesCount (Maybe Integer)

The number of instances with requested status.

icRunningSetup :: Lens' InstancesCount (Maybe Integer)

The number of instances with running_setup status.

icSetupFailed :: Lens' InstancesCount (Maybe Integer)

The number of instances with setup_failed status.

icShuttingDown :: Lens' InstancesCount (Maybe Integer)

The number of instances with shutting_down status.

icStartFailed :: Lens' InstancesCount (Maybe Integer)

The number of instances with start_failed status.

icStopped :: Lens' InstancesCount (Maybe Integer)

The number of instances with stopped status.

icStopping :: Lens' InstancesCount (Maybe Integer)

The number of instances with stopping status.

icTerminated :: Lens' InstancesCount (Maybe Integer)

The number of instances with terminated status.

icTerminating :: Lens' InstancesCount (Maybe Integer)

The number of instances with terminating status.

Layer

data Layer

Describes a layer.

mkLayer :: Layer

Smart constructor for the minimum required fields to construct a valid Layer data type to populate a request.

lStackId :: Lens' Layer (Maybe Text)

The layer stack ID.

lLayerId :: Lens' Layer (Maybe Text)

The layer ID.

lType :: Lens' Layer (Maybe LayerType)

The layer type, which must be one of the following: Custom GangliaMonitoringMaster HaProxy MemcachedServer MySqlMaster NodeJsAppServer PhpAppServer RailsAppServer WebServer.

lName :: Lens' Layer (Maybe Text)

The layer name.

lShortname :: Lens' Layer (Maybe Text)

The layer short name.

lAttributes :: Lens' Layer (Map LayerAttributesKeys Text)

The layer attributes.

lCustomInstanceProfileArn :: Lens' Layer (Maybe Text)

The ARN of the default IAM profile to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

lCustomSecurityGroupIds :: Lens' Layer [Text]

An array containing the layer's custom security group IDs.

lDefaultSecurityGroupNames :: Lens' Layer [Text]

An array containing the layer's security group names.

lPackages :: Lens' Layer [Text]

An array of Package objects that describe the layer's packages.

lVolumeConfigurations :: Lens' Layer [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

lEnableAutoHealing :: Lens' Layer (Maybe Bool)

Whether auto healing is disabled for the layer.

lAutoAssignElasticIps :: Lens' Layer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

lAutoAssignPublicIps :: Lens' Layer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

lDefaultRecipes :: Lens' Layer (Maybe Recipes)

AWS OpsWorks supports five lifecycle events, setup, configuration, deploy, undeploy, and shutdown. For each layer, AWS OpsWorks runs a set of standard recipes for each event. In addition, you can provide custom recipes for any or all layers and events. AWS OpsWorks runs custom event recipes after the standard recipes. LayerCustomRecipes specifies the custom recipes for a particular layer to be run in response to each of the five events. To specify a recipe, use the cookbook's directory name in the repository followed by two colons and the recipe name, which is the recipe's file name without the .rb extension. For example: phpapp2::dbsetup specifies the dbsetup.rb recipe in the repository's phpapp2 folder.

lCustomRecipes :: Lens' Layer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer's custom recipes.

lCreatedAt :: Lens' Layer (Maybe Text)

Date when the layer was created.

lInstallUpdatesOnBoot :: Lens' Layer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances. We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

lUseEbsOptimizedInstances :: Lens' Layer (Maybe Bool)

Whether the layer uses Amazon EBS-optimized instances.

LoadBasedAutoScalingConfiguration

data LoadBasedAutoScalingConfiguration

Describes a layer's load-based auto scaling configuration.

mkLoadBasedAutoScalingConfiguration :: LoadBasedAutoScalingConfiguration

Smart constructor for the minimum required fields to construct a valid LoadBasedAutoScalingConfiguration data type to populate a request.

lbascLayerId :: Lens' LoadBasedAutoScalingConfiguration (Maybe Text)

The layer ID.

lbascEnable :: Lens' LoadBasedAutoScalingConfiguration (Maybe Bool)

Whether load-based auto scaling is enabled for the layer.

lbascUpScaling :: Lens' LoadBasedAutoScalingConfiguration (Maybe AutoScalingThresholds)

A LoadBasedAutoscalingInstruction object that describes the upscaling configuration, which defines how and when AWS OpsWorks increases the number of instances.

lbascDownScaling :: Lens' LoadBasedAutoScalingConfiguration (Maybe AutoScalingThresholds)

A LoadBasedAutoscalingInstruction object that describes the downscaling configuration, which defines how and when AWS OpsWorks reduces the number of instances.

Permission

data Permission

Describes stack or user permissions.

mkPermission :: Permission

Smart constructor for the minimum required fields to construct a valid Permission data type to populate a request.

pStackId :: Lens' Permission (Maybe Text)

A stack ID.

pIamUserArn :: Lens' Permission (Maybe Text)

The Amazon Resource Name (ARN) for an AWS Identity and Access Management (IAM) role. For more information about IAM ARNs, see Using Identifiers.

pAllowSsh :: Lens' Permission (Maybe Bool)

Whether the user can use SSH.

pAllowSudo :: Lens' Permission (Maybe Bool)

Whether the user can use sudo.

pLevel :: Lens' Permission (Maybe Text)

The user's permission level, which must be the following: deny show deploy manage iam_only For more information on the permissions associated with these levels, see Managing User Permissions.

RaidArray

data RaidArray

Describes an instance's RAID array.

mkRaidArray :: RaidArray

Smart constructor for the minimum required fields to construct a valid RaidArray data type to populate a request.

raRaidArrayId :: Lens' RaidArray (Maybe Text)

The array ID.

raInstanceId :: Lens' RaidArray (Maybe Text)

The instance ID.

raName :: Lens' RaidArray (Maybe Text)

The array name.

raRaidLevel :: Lens' RaidArray (Maybe Integer)

The RAID level.

raNumberOfDisks :: Lens' RaidArray (Maybe Integer)

The number of disks in the array.

raSize :: Lens' RaidArray (Maybe Integer)

The array's size.

raDevice :: Lens' RaidArray (Maybe Text)

The array's Linux device. For example devmdadm0.

raMountPoint :: Lens' RaidArray (Maybe Text)

The array's mount point.

raAvailabilityZone :: Lens' RaidArray (Maybe Text)

The array's Availability Zone. For more information, see Regions and Endpoints.

raCreatedAt :: Lens' RaidArray (Maybe Text)

When the RAID array was created.

raVolumeType :: Lens' RaidArray (Maybe Text)

The volume type, standard or PIOPS.

raIops :: Lens' RaidArray (Maybe Integer)

For PIOPS volumes, the IOPS per disk.

RdsDbInstance

data RdsDbInstance

Describes an Amazon RDS instance.

mkRdsDbInstance :: RdsDbInstance

Smart constructor for the minimum required fields to construct a valid RdsDbInstance data type to populate a request.

rdiRdsDbInstanceArn :: Lens' RdsDbInstance (Maybe Text)

The instance's ARN.

rdiDbInstanceIdentifier :: Lens' RdsDbInstance (Maybe Text)

The DB instance identifier.

rdiDbUser :: Lens' RdsDbInstance (Maybe Text)

The master user name.

rdiDbPassword :: Lens' RdsDbInstance (Maybe Text)

The database password.

rdiRegion :: Lens' RdsDbInstance (Maybe Text)

The instance's AWS region.

rdiAddress :: Lens' RdsDbInstance (Maybe Text)

The instance's address.

rdiEngine :: Lens' RdsDbInstance (Maybe Text)

The instance's database engine.

rdiStackId :: Lens' RdsDbInstance (Maybe Text)

The ID of the stack that the instance is registered with.

rdiMissingOnRds :: Lens' RdsDbInstance (Maybe Bool)

Set to true if AWS OpsWorks was unable to discover the Amazon RDS instance. AWS OpsWorks attempts to discover the instance only once. If this value is set to true, you must deregister the instance and then register it again.

Recipes

data Recipes

A LayerCustomRecipes object that specifies the layer custom recipes.

mkRecipes :: Recipes

Smart constructor for the minimum required fields to construct a valid Recipes data type to populate a request.

rSetup :: Lens' Recipes [Text]

An array of custom recipe names to be run following a setup event.

rConfigure :: Lens' Recipes [Text]

An array of custom recipe names to be run following a configure event.

rDeploy :: Lens' Recipes [Text]

An array of custom recipe names to be run following a deploy event.

rUndeploy :: Lens' Recipes [Text]

An array of custom recipe names to be run following a undeploy event.

rShutdown :: Lens' Recipes [Text]

An array of custom recipe names to be run following a shutdown event.

SelfUserProfile

data SelfUserProfile

A UserProfile object that describes the user's SSH information.

mkSelfUserProfile :: SelfUserProfile

Smart constructor for the minimum required fields to construct a valid SelfUserProfile data type to populate a request.

supIamUserArn :: Lens' SelfUserProfile (Maybe Text)

The user's IAM ARN.

supName :: Lens' SelfUserProfile (Maybe Text)

The user's name.

supSshUsername :: Lens' SelfUserProfile (Maybe Text)

The user's SSH user name.

supSshPublicKey :: Lens' SelfUserProfile (Maybe Text)

The user's SSH public key.

ServiceError

data ServiceError

Describes an AWS OpsWorks service error.

mkServiceError :: ServiceError

Smart constructor for the minimum required fields to construct a valid ServiceError data type to populate a request.

seServiceErrorId :: Lens' ServiceError (Maybe Text)

The error ID.

seStackId :: Lens' ServiceError (Maybe Text)

The stack ID.

seInstanceId :: Lens' ServiceError (Maybe Text)

The instance ID.

seType :: Lens' ServiceError (Maybe Text)

The error type.

seMessage :: Lens' ServiceError (Maybe Text)

A message that describes the error.

seCreatedAt :: Lens' ServiceError (Maybe Text)

When the error occurred.

Source

data Source

Contains the information required to retrieve an app or cookbook from a repository. For more information, see Creating Apps or Custom Recipes and Cookbooks.

mkSource :: Source

Smart constructor for the minimum required fields to construct a valid Source data type to populate a request.

sType :: Lens' Source (Maybe SourceType)

The repository type.

sUrl :: Lens' Source (Maybe Text)

The source URL.

sUsername :: Lens' Source (Maybe Text)

This parameter depends on the repository type. For Amazon S3 bundles, set Username to the appropriate IAM access key ID. For HTTP bundles, Git repositories, and Subversion repositories, set Username to the user name.

sPassword :: Lens' Source (Maybe Text)

This parameter depends on the repository type. For Amazon S3 bundles, set Password to the appropriate IAM secret access key. For HTTP bundles and Subversion repositories, set Password to the password. For more information on how to safely handle IAM credentials, see .

sSshKey :: Lens' Source (Maybe Text)

The repository's SSH key.

sRevision :: Lens' Source (Maybe Text)

The application's version. AWS OpsWorks enables you to easily deploy new versions of an application. One of the simplest approaches is to have branches or revisions in your repository that represent different versions that can potentially be deployed.

SslConfiguration

data SslConfiguration

An SslConfiguration object with the SSL configuration.

mkSslConfiguration

Arguments

:: Text
scCertificate

-> Text
scPrivateKey

-> SslConfiguration

Smart constructor for the minimum required fields to construct a valid SslConfiguration data type to populate a request.

scCertificate :: Lens' SslConfiguration Text

The contents of the certificate's domain.crt file.

scPrivateKey :: Lens' SslConfiguration Text

The private key; the contents of the certificate's domain.kex file.

scChain :: Lens' SslConfiguration (Maybe Text)

Optional. Can be used to specify an intermediate certificate authority key or client authentication.

Stack

data Stack

Describes a stack.

mkStack :: Stack

Smart constructor for the minimum required fields to construct a valid Stack data type to populate a request.

srStackId :: Lens' Stack (Maybe Text)

The stack ID.

srName :: Lens' Stack (Maybe Text)

The stack name.

srArn :: Lens' Stack (Maybe Text)

The stack's ARN.

srRegion :: Lens' Stack (Maybe Text)

The stack AWS region, such as "us-east-1". For more information about AWS regions, see Regions and Endpoints.

srVpcId :: Lens' Stack (Maybe Text)

The VPC ID, if the stack is running in a VPC.

srAttributes :: Lens' Stack (Map StackAttributesKeys Text)

The stack's attributes.

srServiceRoleArn :: Lens' Stack (Maybe Text)

The stack AWS Identity and Access Management (IAM) role.

srDefaultInstanceProfileArn :: Lens' Stack (Maybe Text)

The ARN of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

srDefaultOs :: Lens' Stack (Maybe Text)

The stack's default operating system, which must be set to Amazon Linux or Ubuntu 12.04 LTS. The default option is Amazon Linux.

srHostnameTheme :: Lens' Stack (Maybe Text)

The stack host name theme, with spaces replaced by underscores.

srDefaultAvailabilityZone :: Lens' Stack (Maybe Text)

The stack's default Availability Zone. For more information, see Regions and Endpoints.

srDefaultSubnetId :: Lens' Stack (Maybe Text)

The default subnet ID, if the stack is running in a VPC.

srCustomJson :: Lens' Stack (Maybe Text)

A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format and must escape characters such as '"'.: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration JSON.

srConfigurationManager :: Lens' Stack (Maybe StackConfigurationManager)

The configuration manager.

srChefConfiguration :: Lens' Stack (Maybe ChefConfiguration)

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version. For more information, see Create a New Stack.

srUseCustomCookbooks :: Lens' Stack (Maybe Bool)

Whether the stack uses custom cookbooks.

srUseOpsworksSecurityGroups :: Lens' Stack (Maybe Bool)

Whether the stack automatically associates the AWS OpsWorks built-in security groups with the stack's layers.

srCustomCookbooksSource :: Lens' Stack (Maybe Source)

Contains the information required to retrieve an app or cookbook from a repository. For more information, see Creating Apps or Custom Recipes and Cookbooks.

srDefaultSshKeyName :: Lens' Stack (Maybe Text)

A default SSH key for the stack's instances. You can override this value when you create or update an instance.

srCreatedAt :: Lens' Stack (Maybe Text)

Date when the stack was created.

srDefaultRootDeviceType :: Lens' Stack (Maybe RootDeviceType)

The default root device type. This value is used by default for all instances in the stack, but you can override it when you create an instance. For more information, see Storage for the Root Device.

StackConfigurationManager

data StackConfigurationManager

The configuration manager. When you clone a stack we recommend that you use the configuration manager to specify the Chef version, 0.9, 11.4, or 11.10. The default value is currently 11.4.

mkStackConfigurationManager :: StackConfigurationManager

Smart constructor for the minimum required fields to construct a valid StackConfigurationManager data type to populate a request.

scmName :: Lens' StackConfigurationManager (Maybe Text)

The name. This parameter must be set to Chef.

scmVersion :: Lens' StackConfigurationManager (Maybe Text)

The Chef version. This parameter must be set to 0.9, 11.4, or 11.10. The default value is 11.4.

StackSummary

data StackSummary

A StackSummary object that contains the results.

mkStackSummary :: StackSummary

Smart constructor for the minimum required fields to construct a valid StackSummary data type to populate a request.

ssStackId :: Lens' StackSummary (Maybe Text)

The stack ID.

ssName :: Lens' StackSummary (Maybe Text)

The stack name.

ssArn :: Lens' StackSummary (Maybe Text)

The stack's ARN.

ssLayersCount :: Lens' StackSummary (Maybe Integer)

The number of layers.

ssAppsCount :: Lens' StackSummary (Maybe Integer)

The number of apps.

ssInstancesCount :: Lens' StackSummary (Maybe InstancesCount)

An InstancesCount object with the number of instances in each status.

TimeBasedAutoScalingConfiguration

data TimeBasedAutoScalingConfiguration

Describes an instance's time-based auto scaling configuration.

mkTimeBasedAutoScalingConfiguration :: TimeBasedAutoScalingConfiguration

Smart constructor for the minimum required fields to construct a valid TimeBasedAutoScalingConfiguration data type to populate a request.

tbascInstanceId :: Lens' TimeBasedAutoScalingConfiguration (Maybe Text)

The instance ID.

tbascAutoScalingSchedule :: Lens' TimeBasedAutoScalingConfiguration (Maybe WeeklyAutoScalingSchedule)

A WeeklyAutoScalingSchedule object with the instance schedule.

UserProfile

data UserProfile

Describes a user's SSH information.

mkUserProfile :: UserProfile

Smart constructor for the minimum required fields to construct a valid UserProfile data type to populate a request.

upIamUserArn :: Lens' UserProfile (Maybe Text)

The user's IAM ARN.

upName :: Lens' UserProfile (Maybe Text)

The user's name.

upSshUsername :: Lens' UserProfile (Maybe Text)

The user's SSH user name.

upSshPublicKey :: Lens' UserProfile (Maybe Text)

The user's SSH public key.

upAllowSelfManagement :: Lens' UserProfile (Maybe Bool)

Whether users can specify their own SSH public key through the My Settings page. For more information, see Managing User Permissions.

Volume

data Volume

Describes an instance's Amazon EBS volume.

mkVolume :: Volume

Smart constructor for the minimum required fields to construct a valid Volume data type to populate a request.

vVolumeId :: Lens' Volume (Maybe Text)

The volume ID.

vEc2VolumeId :: Lens' Volume (Maybe Text)

The Amazon EC2 volume ID.

vName :: Lens' Volume (Maybe Text)

The volume name.

vRaidArrayId :: Lens' Volume (Maybe Text)

The RAID array ID.

vInstanceId :: Lens' Volume (Maybe Text)

The instance ID.

vStatus :: Lens' Volume (Maybe Text)

The value returned by DescribeVolumes.

vSize :: Lens' Volume (Maybe Integer)

The volume size.

vDevice :: Lens' Volume (Maybe Text)

The device name.

vMountPoint :: Lens' Volume (Maybe Text)

The volume mount point. For example "devsdh".

vRegion :: Lens' Volume (Maybe Text)

The AWS region. For more information about AWS regions, see Regions and Endpoints.

vAvailabilityZone :: Lens' Volume (Maybe Text)

The volume Availability Zone. For more information, see Regions and Endpoints.

vVolumeType :: Lens' Volume (Maybe Text)

The volume type, standard or PIOPS.

vIops :: Lens' Volume (Maybe Integer)

For PIOPS volumes, the IOPS per disk.

VolumeConfiguration

data VolumeConfiguration

Describes an Amazon EBS volume configuration.

mkVolumeConfiguration

Arguments

:: Text
vcMountPoint

-> Integer
vcNumberOfDisks

-> Integer
vcSize

-> VolumeConfiguration

Smart constructor for the minimum required fields to construct a valid VolumeConfiguration data type to populate a request.

vcMountPoint :: Lens' VolumeConfiguration Text

The volume mount point. For example "devsdh".

vcRaidLevel :: Lens' VolumeConfiguration (Maybe Integer)

The volume RAID level.

vcNumberOfDisks :: Lens' VolumeConfiguration Integer

The number of disks in the volume.

vcSize :: Lens' VolumeConfiguration Integer

The volume size.

vcVolumeType :: Lens' VolumeConfiguration (Maybe Text)

The volume type, standard or PIOPS.

vcIops :: Lens' VolumeConfiguration (Maybe Integer)

For PIOPS volumes, the IOPS per disk.

WeeklyAutoScalingSchedule

data WeeklyAutoScalingSchedule

A WeeklyAutoScalingSchedule object with the instance schedule.

mkWeeklyAutoScalingSchedule :: WeeklyAutoScalingSchedule

Smart constructor for the minimum required fields to construct a valid WeeklyAutoScalingSchedule data type to populate a request.

wassMonday :: Lens' WeeklyAutoScalingSchedule (Map Text Text)

The schedule for Monday.

wassTuesday :: Lens' WeeklyAutoScalingSchedule (Map Text Text)

The schedule for Tuesday.

wassWednesday :: Lens' WeeklyAutoScalingSchedule (Map Text Text)

The schedule for Wednesday.

wassThursday :: Lens' WeeklyAutoScalingSchedule (Map Text Text)

The schedule for Thursday.

wassFriday :: Lens' WeeklyAutoScalingSchedule (Map Text Text)

The schedule for Friday.

wassSaturday :: Lens' WeeklyAutoScalingSchedule (Map Text Text)

The schedule for Saturday.

wassSunday :: Lens' WeeklyAutoScalingSchedule (Map Text Text)

The schedule for Sunday.