Network.AWS.OpsWorks.V2013_02_18.UpdateInstance

Description

Updates a specified instance. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

updateInstance

Arguments

:: Text
_uirInstanceId

-> UpdateInstance

Minimum specification for a UpdateInstance request.

data UpdateInstance

Constructors

UpdateInstance

Fields

_uirInstanceId :: Text

The instance ID.

_uirArchitecture :: Maybe Architecture

The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

_uirAutoScalingType :: Maybe AutoScalingType

The instance's auto scaling type, which has three possible values: AlwaysRunning: A 24/7 instance, which is not affected by auto scaling. TimeBasedAutoScaling: A time-based auto scaling instance, which is started and stopped based on a specified schedule. LoadBasedAutoScaling: A load-based auto scaling instance, which is started and stopped based on load metrics.

_uirInstallUpdatesOnBoot :: Maybe Bool

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances. We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

_uirEbsOptimized :: Maybe Bool

Whether this is an Amazon EBS-optimized instance.

_uirHostname :: Maybe Text

The instance host name.

_uirSshKeyName :: Maybe Text

The instance SSH key name.

_uirInstanceType :: Maybe Text

The instance type. AWS OpsWorks supports all instance types except Cluster Compute, Cluster GPU, and High Memory Cluster. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

_uirOs :: Maybe Text

The instance operating system, which must be set to one of the following. Standard operating systems: Amazon Linux or Ubuntu 12.04 LTS Custom AMIs: Custom The default option is Amazon Linux. If you set this parameter to Custom, you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. For more information on the standard operating systems, see Operating SystemsFor more information on how to use custom AMIs with OpsWorks, see Using Custom AMIs.

_uirAmiId :: Maybe Text

A custom AMI ID to be used to create the instance. The AMI should be based on one of the standard AWS OpsWorks APIs: Amazon Linux or Ubuntu 12.04 LTS. For more information, see Instances.

_uirLayerIds :: [Text]

The instance's layer IDs.

uirSshKeyName :: Lens' UpdateInstance (Maybe Text)

uirOs :: Lens' UpdateInstance (Maybe Text)

uirLayerIds :: Lens' UpdateInstance [Text]

uirInstanceType :: Lens' UpdateInstance (Maybe Text)

uirInstanceId :: Lens' UpdateInstance Text

uirInstallUpdatesOnBoot :: Lens' UpdateInstance (Maybe Bool)

uirHostname :: Lens' UpdateInstance (Maybe Text)

uirEbsOptimized :: Lens' UpdateInstance (Maybe Bool)

uirAutoScalingType :: Lens' UpdateInstance (Maybe AutoScalingType)

uirArchitecture :: Lens' UpdateInstance (Maybe Architecture)

uirAmiId :: Lens' UpdateInstance (Maybe Text)

data UpdateInstanceResponse

Constructors

UpdateInstanceResponse