Network.AWS.OpsWorks.V2013_02_18.UpdateLayer

Description

Updates a specified layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

updateLayer

Arguments

:: Text
_ulrLayerId

-> UpdateLayer

Minimum specification for a UpdateLayer request.

data UpdateLayer

Constructors

UpdateLayer

Fields

_ulrLayerId :: Text

The layer ID.

_ulrInstallUpdatesOnBoot :: Maybe Bool

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances. We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

_ulrEnableAutoHealing :: Maybe Bool

Whether to disable auto healing for the layer.

_ulrAutoAssignPublicIps :: Maybe Bool

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

_ulrUseEbsOptimizedInstances :: Maybe Bool

Whether to use Amazon EBS-optimized instances.

_ulrAutoAssignElasticIps :: Maybe Bool

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

_ulrAttributes :: Map LayerAttributesKeys Text

One or more user-defined key/value pairs to be added to the stack attributes.

_ulrCustomRecipes :: Maybe Recipes

A LayerCustomRecipes object that specifies the layer's custom recipes.

_ulrCustomInstanceProfileArn :: Maybe Text

The ARN of an IAM profile to be used for all of the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

_ulrShortname :: Maybe Text

The layer short name, which is used internally by AWS OpsWorksand by Chef. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters and must be in the following format: A[a-z0-9-_.]+Z.

_ulrName :: Maybe Text

The layer name, which is used by the console.

_ulrCustomSecurityGroupIds :: [Text]

An array containing the layer's custom security group IDs.

_ulrPackages :: [Text]

An array of Package objects that describe the layer's packages.

_ulrVolumeConfigurations :: [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

ulrVolumeConfigurations :: Lens' UpdateLayer [VolumeConfiguration]

ulrUseEbsOptimizedInstances :: Lens' UpdateLayer (Maybe Bool)

ulrShortname :: Lens' UpdateLayer (Maybe Text)

ulrPackages :: Lens' UpdateLayer [Text]

ulrName :: Lens' UpdateLayer (Maybe Text)

ulrLayerId :: Lens' UpdateLayer Text

ulrInstallUpdatesOnBoot :: Lens' UpdateLayer (Maybe Bool)

ulrEnableAutoHealing :: Lens' UpdateLayer (Maybe Bool)

ulrCustomSecurityGroupIds :: Lens' UpdateLayer [Text]

ulrCustomRecipes :: Lens' UpdateLayer (Maybe Recipes)

ulrCustomInstanceProfileArn :: Lens' UpdateLayer (Maybe Text)

ulrAutoAssignPublicIps :: Lens' UpdateLayer (Maybe Bool)

ulrAutoAssignElasticIps :: Lens' UpdateLayer (Maybe Bool)

ulrAttributes :: Lens' UpdateLayer (Map LayerAttributesKeys Text)

data UpdateLayerResponse

Constructors

UpdateLayerResponse