Network.AWS.OpsWorks.V2013_02_18.UpdateMyUserProfile

Description

Updates a user's SSH public key. Required Permissions: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data UpdateMyUserProfile

Request constructor

mkUpdateMyUserProfile :: UpdateMyUserProfile

Smart constructor for the minimum required parameters to construct a valid UpdateMyUserProfile request.

The fields accessible through corresponding lenses are:

SshPublicKey :: Maybe Text

Request lenses

umupSshPublicKey :: Lens' UpdateMyUserProfile (Maybe Text)

The user's SSH public key.

Response

data UpdateMyUserProfileResponse

Response constructor

mkUpdateMyUserProfileResponse :: UpdateMyUserProfileResponse

Smart constructor for the minimum required parameters to construct a valid UpdateMyUserProfileResponse response.

This constructor is provided for convenience and testing purposes.