Network.AWS.OpsWorks.V2013_02_18.UpdateUserProfile

Description

Updates a specified user profile. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

updateUserProfile

Arguments

:: Text
_uuprIamUserArn

-> UpdateUserProfile

Minimum specification for a UpdateUserProfile request.

data UpdateUserProfile

Constructors

UpdateUserProfile

Fields

_uuprIamUserArn :: Text

The user IAM ARN.

_uuprAllowSelfManagement :: Maybe Bool

Whether users can specify their own SSH public key through the My Settings page. For more information, see Managing User Permissions.

_uuprSshPublicKey :: Maybe Text

The user's new SSH public key.

_uuprSshUsername :: Maybe Text

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], -, and '_'. If the specified name includes other punctuation marks, AWS OpsWorks removes them. For example, my.name will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks generates one from the IAM user name.

uuprSshUsername :: Lens' UpdateUserProfile (Maybe Text)

uuprSshPublicKey :: Lens' UpdateUserProfile (Maybe Text)

uuprIamUserArn :: Lens' UpdateUserProfile Text

uuprAllowSelfManagement :: Lens' UpdateUserProfile (Maybe Bool)

data UpdateUserProfileResponse

Constructors

UpdateUserProfileResponse