Network.AWS.OpsWorks.AssignVolume

Description

Assigns one of the stack's registered Amazon EBS volumes to a specified instance. The volume must first be registered with the stack by calling RegisterVolume . After you register the volume, you must call UpdateVolume to specify a mount point before calling AssignVolume . For more information, see Resource Management .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

assignVolume

Arguments

:: Text
avVolumeId

-> AssignVolume

Creates a value of AssignVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avInstanceId - The instance ID.

avVolumeId - The volume ID.

data AssignVolume

See: assignVolume smart constructor.

Request Lenses

avInstanceId :: Lens' AssignVolume (Maybe Text)

The instance ID.

avVolumeId :: Lens' AssignVolume Text

The volume ID.

Destructuring the Response

assignVolumeResponse :: AssignVolumeResponse

Creates a value of AssignVolumeResponse with the minimum fields required to make a request.

data AssignVolumeResponse

See: assignVolumeResponse smart constructor.