Network.AWS.OpsWorks.AssociateElasticIp

Description

Associates one of the stack's registered Elastic IP addresses with a specified instance. The address must first be registered with the stack by calling RegisterElasticIp. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_AssociateElasticIp.html

Request

data AssociateElasticIp

Request constructor

associateElasticIp

Arguments

:: Text
aeiElasticIp

-> AssociateElasticIp

AssociateElasticIp constructor.

The fields accessible through corresponding lenses are:

aeiElasticIp :: Text

aeiInstanceId :: Maybe Text

Request lenses

aeiElasticIp :: Lens' AssociateElasticIp Text

The Elastic IP address.

aeiInstanceId :: Lens' AssociateElasticIp (Maybe Text)

The instance ID.

Response

data AssociateElasticIpResponse

Response constructor

associateElasticIpResponse :: AssociateElasticIpResponse

AssociateElasticIpResponse constructor.