Network.AWS.OpsWorks.CloneStack

Description

Creates a clone of a specified stack. For more information, see Clone a Stack . By default, all parameters are set to the values used by the parent stack.

Required Permissions : To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

cloneStack

Arguments

:: Text
cSourceStackId

-> Text
cServiceRoleARN

-> CloneStack

Creates a value of CloneStack with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cDefaultInstanceProfileARN - The Amazon Resource Name (ARN) of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers .

cCloneAppIds - A list of source stack app IDs to be included in the cloned stack.

cDefaultRootDeviceType - The default root device type. This value is used by default for all instances in the cloned stack, but you can override it when you create an instance. For more information, see Storage for the Root Device .

cVPCId - The ID of the VPC that the cloned stack is to be launched into. It must be in the specified region. All instances are launched into this VPC, and you cannot change the ID later. * If your account supports EC2 Classic, the default value is no VPC. * If your account does not support EC2 Classic, the default value is the default VPC for the specified region. If the VPC ID corresponds to a default VPC and you have specified either the DefaultAvailabilityZone or the DefaultSubnetId parameter only, AWS OpsWorks Stacks infers the value of the other parameter. If you specify neither parameter, AWS OpsWorks Stacks sets these parameters to the first valid Availability Zone for the specified region and the corresponding default VPC subnet ID, respectively. If you specify a nondefault VPC ID, note the following: * It must belong to a VPC in your account that is in the specified region. * You must specify a value for DefaultSubnetId . For more information on how to use AWS OpsWorks Stacks with a VPC, see Running a Stack in a VPC . For more information on default VPC and EC2 Classic, see Supported Platforms .

cChefConfiguration - A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version on Chef 11.10 stacks. For more information, see Create a New Stack .

cAgentVersion - The default AWS OpsWorks Stacks agent version. You have the following options: * Auto-update - Set this parameter to LATEST . AWS OpsWorks Stacks automatically installs new agent versions on the stack's instances as soon as they are available. * Fixed version - Set this parameter to your preferred agent version. To update the agent version, you must edit the stack configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the stack's instances. The default setting is LATEST . To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions . AgentVersion cannot be set to Chef 12.2.

cDefaultSSHKeyName - A default Amazon EC2 key pair name. The default value is none. If you specify a key pair name, AWS OpsWorks installs the public key on the instance and you can use the private key with an SSH client to log in to the instance. For more information, see Using SSH to Communicate with an Instance and Managing SSH Access . You can override this setting by specifying a different key pair, or no key pair, when you create an instance .

cCustomJSON - A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes

cClonePermissions - Whether to clone the source stack's permissions.

cCustomCookbooksSource - Undocumented member.

cDefaultAvailabilityZone - The cloned stack's default Availability Zone, which must be in the specified region. For more information, see Regions and Endpoints . If you also specify a value for DefaultSubnetId , the subnet must be in the same zone. For more information, see the VpcId parameter description.

cAttributes - A list of stack attributes and values as key/value pairs to be added to the cloned stack.

cName - The cloned stack name.

cDefaultOS - The stack's operating system, which must be set to one of the following. * A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2017.03 , Amazon Linux 2016.09 , Amazon Linux 2016.03 , Amazon Linux 2015.09 , or Amazon Linux 2015.03 . * A supported Ubuntu operating system, such as Ubuntu 16.04 LTS , Ubuntu 14.04 LTS , or Ubuntu 12.04 LTS . * CentOS Linux 7 * Red Hat Enterprise Linux 7 * Microsoft Windows Server 2012 R2 Base , Microsoft Windows Server 2012 R2 with SQL Server Express , Microsoft Windows Server 2012 R2 with SQL Server Standard , or Microsoft Windows Server 2012 R2 with SQL Server Web . * A custom AMI: Custom . You specify the custom AMI you want to use when you create instances. For more information on how to use custom AMIs with OpsWorks, see Using Custom AMIs . The default option is the parent stack's operating system. For more information on the supported operating systems, see AWS OpsWorks Stacks Operating Systems .

cUseOpsworksSecurityGroups - Whether to associate the AWS OpsWorks Stacks built-in security groups with the stack's layers. AWS OpsWorks Stacks provides a standard set of built-in security groups, one for each layer, which are associated with layers by default. With UseOpsworksSecurityGroups you can instead provide your own custom security groups. UseOpsworksSecurityGroups has the following settings: * True - AWS OpsWorks Stacks automatically associates the appropriate built-in security group with each layer (default setting). You can associate additional security groups with a layer after you create it but you cannot delete the built-in security group. * False - AWS OpsWorks Stacks does not associate built-in security groups with layers. You must create appropriate Amazon Elastic Compute Cloud (Amazon EC2) security groups and associate a security group with each layer that you create. However, you can still manually associate a built-in security group with a layer on creation; custom security groups are required only for those layers that need custom settings. For more information, see Create a New Stack .

cUseCustomCookbooks - Whether to use custom cookbooks.

cDefaultSubnetId - The stack's default VPC subnet ID. This parameter is required if you specify a value for the VpcId parameter. All instances are launched into this subnet unless you specify otherwise when you create the instance. If you also specify a value for DefaultAvailabilityZone , the subnet must be in that zone. For information on default values and when this parameter is required, see the VpcId parameter description.

cRegion - The cloned stack AWS region, such as "ap-northeast-2". For more information about AWS regions, see Regions and Endpoints .

cConfigurationManager - The configuration manager. When you clone a stack we recommend that you use the configuration manager to specify the Chef version: 12, 11.10, or 11.4 for Linux stacks, or 12.2 for Windows stacks. The default value for Linux stacks is currently 12.

cHostnameTheme - The stack's host name theme, with spaces are replaced by underscores. The theme is used to generate host names for the stack's instances. By default, HostnameTheme is set to Layer_Dependent , which creates host names by appending integers to the layer's short name. The other themes are: * Baked_Goods * Clouds * Europe_Cities * Fruits * Greek_Deities * Legendary_creatures_from_Japan * Planets_and_Moons * Roman_Deities * Scottish_Islands * US_Cities * Wild_Cats To obtain a generated host name, call GetHostNameSuggestion , which returns a host name based on the current theme.

cSourceStackId - The source stack ID.

cServiceRoleARN - The stack AWS Identity and Access Management (IAM) role, which allows AWS OpsWorks Stacks to work with AWS resources on your behalf. You must set this parameter to the Amazon Resource Name (ARN) for an existing IAM role. If you create a stack by using the AWS OpsWorks Stacks console, it creates the role for you. You can obtain an existing stack's IAM ARN programmatically by calling DescribePermissions . For more information about IAM ARNs, see Using Identifiers .

data CloneStack

See: cloneStack smart constructor.

Request Lenses

cDefaultInstanceProfileARN :: Lens' CloneStack (Maybe Text)

The Amazon Resource Name (ARN) of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers .

cCloneAppIds :: Lens' CloneStack [Text]

A list of source stack app IDs to be included in the cloned stack.

cDefaultRootDeviceType :: Lens' CloneStack (Maybe RootDeviceType)

The default root device type. This value is used by default for all instances in the cloned stack, but you can override it when you create an instance. For more information, see Storage for the Root Device .

cVPCId :: Lens' CloneStack (Maybe Text)

The ID of the VPC that the cloned stack is to be launched into. It must be in the specified region. All instances are launched into this VPC, and you cannot change the ID later. * If your account supports EC2 Classic, the default value is no VPC. * If your account does not support EC2 Classic, the default value is the default VPC for the specified region. If the VPC ID corresponds to a default VPC and you have specified either the DefaultAvailabilityZone or the DefaultSubnetId parameter only, AWS OpsWorks Stacks infers the value of the other parameter. If you specify neither parameter, AWS OpsWorks Stacks sets these parameters to the first valid Availability Zone for the specified region and the corresponding default VPC subnet ID, respectively. If you specify a nondefault VPC ID, note the following: * It must belong to a VPC in your account that is in the specified region. * You must specify a value for DefaultSubnetId . For more information on how to use AWS OpsWorks Stacks with a VPC, see Running a Stack in a VPC . For more information on default VPC and EC2 Classic, see Supported Platforms .

cChefConfiguration :: Lens' CloneStack (Maybe ChefConfiguration)

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version on Chef 11.10 stacks. For more information, see Create a New Stack .

cAgentVersion :: Lens' CloneStack (Maybe Text)

The default AWS OpsWorks Stacks agent version. You have the following options: * Auto-update - Set this parameter to LATEST . AWS OpsWorks Stacks automatically installs new agent versions on the stack's instances as soon as they are available. * Fixed version - Set this parameter to your preferred agent version. To update the agent version, you must edit the stack configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the stack's instances. The default setting is LATEST . To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions . AgentVersion cannot be set to Chef 12.2.

cDefaultSSHKeyName :: Lens' CloneStack (Maybe Text)

A default Amazon EC2 key pair name. The default value is none. If you specify a key pair name, AWS OpsWorks installs the public key on the instance and you can use the private key with an SSH client to log in to the instance. For more information, see Using SSH to Communicate with an Instance and Managing SSH Access . You can override this setting by specifying a different key pair, or no key pair, when you create an instance .

cCustomJSON :: Lens' CloneStack (Maybe Text)

A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes

cClonePermissions :: Lens' CloneStack (Maybe Bool)

Whether to clone the source stack's permissions.

cCustomCookbooksSource :: Lens' CloneStack (Maybe Source)

Undocumented member.

cDefaultAvailabilityZone :: Lens' CloneStack (Maybe Text)

The cloned stack's default Availability Zone, which must be in the specified region. For more information, see Regions and Endpoints . If you also specify a value for DefaultSubnetId , the subnet must be in the same zone. For more information, see the VpcId parameter description.

cAttributes :: Lens' CloneStack (HashMap StackAttributesKeys Text)

A list of stack attributes and values as key/value pairs to be added to the cloned stack.

cName :: Lens' CloneStack (Maybe Text)

The cloned stack name.

cDefaultOS :: Lens' CloneStack (Maybe Text)

The stack's operating system, which must be set to one of the following. * A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2017.03 , Amazon Linux 2016.09 , Amazon Linux 2016.03 , Amazon Linux 2015.09 , or Amazon Linux 2015.03 . * A supported Ubuntu operating system, such as Ubuntu 16.04 LTS , Ubuntu 14.04 LTS , or Ubuntu 12.04 LTS . * CentOS Linux 7 * Red Hat Enterprise Linux 7 * Microsoft Windows Server 2012 R2 Base , Microsoft Windows Server 2012 R2 with SQL Server Express , Microsoft Windows Server 2012 R2 with SQL Server Standard , or Microsoft Windows Server 2012 R2 with SQL Server Web . * A custom AMI: Custom . You specify the custom AMI you want to use when you create instances. For more information on how to use custom AMIs with OpsWorks, see Using Custom AMIs . The default option is the parent stack's operating system. For more information on the supported operating systems, see AWS OpsWorks Stacks Operating Systems .

cUseOpsworksSecurityGroups :: Lens' CloneStack (Maybe Bool)

Whether to associate the AWS OpsWorks Stacks built-in security groups with the stack's layers. AWS OpsWorks Stacks provides a standard set of built-in security groups, one for each layer, which are associated with layers by default. With UseOpsworksSecurityGroups you can instead provide your own custom security groups. UseOpsworksSecurityGroups has the following settings: * True - AWS OpsWorks Stacks automatically associates the appropriate built-in security group with each layer (default setting). You can associate additional security groups with a layer after you create it but you cannot delete the built-in security group. * False - AWS OpsWorks Stacks does not associate built-in security groups with layers. You must create appropriate Amazon Elastic Compute Cloud (Amazon EC2) security groups and associate a security group with each layer that you create. However, you can still manually associate a built-in security group with a layer on creation; custom security groups are required only for those layers that need custom settings. For more information, see Create a New Stack .

cUseCustomCookbooks :: Lens' CloneStack (Maybe Bool)

Whether to use custom cookbooks.

cDefaultSubnetId :: Lens' CloneStack (Maybe Text)

The stack's default VPC subnet ID. This parameter is required if you specify a value for the VpcId parameter. All instances are launched into this subnet unless you specify otherwise when you create the instance. If you also specify a value for DefaultAvailabilityZone , the subnet must be in that zone. For information on default values and when this parameter is required, see the VpcId parameter description.

cRegion :: Lens' CloneStack (Maybe Text)

The cloned stack AWS region, such as "ap-northeast-2". For more information about AWS regions, see Regions and Endpoints .

cConfigurationManager :: Lens' CloneStack (Maybe StackConfigurationManager)

The configuration manager. When you clone a stack we recommend that you use the configuration manager to specify the Chef version: 12, 11.10, or 11.4 for Linux stacks, or 12.2 for Windows stacks. The default value for Linux stacks is currently 12.

cHostnameTheme :: Lens' CloneStack (Maybe Text)

The stack's host name theme, with spaces are replaced by underscores. The theme is used to generate host names for the stack's instances. By default, HostnameTheme is set to Layer_Dependent , which creates host names by appending integers to the layer's short name. The other themes are: * Baked_Goods * Clouds * Europe_Cities * Fruits * Greek_Deities * Legendary_creatures_from_Japan * Planets_and_Moons * Roman_Deities * Scottish_Islands * US_Cities * Wild_Cats To obtain a generated host name, call GetHostNameSuggestion , which returns a host name based on the current theme.

cSourceStackId :: Lens' CloneStack Text

The source stack ID.

cServiceRoleARN :: Lens' CloneStack Text

The stack AWS Identity and Access Management (IAM) role, which allows AWS OpsWorks Stacks to work with AWS resources on your behalf. You must set this parameter to the Amazon Resource Name (ARN) for an existing IAM role. If you create a stack by using the AWS OpsWorks Stacks console, it creates the role for you. You can obtain an existing stack's IAM ARN programmatically by calling DescribePermissions . For more information about IAM ARNs, see Using Identifiers .

Destructuring the Response

cloneStackResponse

Arguments

:: Int
csrsResponseStatus

-> CloneStackResponse

Creates a value of CloneStackResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csrsStackId - The cloned stack ID.

csrsResponseStatus - -- | The response status code.

data CloneStackResponse

Contains the response to a CloneStack request.

See: cloneStackResponse smart constructor.

Response Lenses

csrsStackId :: Lens' CloneStackResponse (Maybe Text)

The cloned stack ID.

csrsResponseStatus :: Lens' CloneStackResponse Int

- | The response status code.