Network.AWS.OpsWorks.CreateApp

Description

Creates an app for a specified stack. For more information, see Creating Apps .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

createApp

Arguments

:: Text
caStackId

-> Text
caName

-> AppType
caType

-> CreateApp

Creates a value of CreateApp with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caSSLConfiguration - An SslConfiguration object with the SSL configuration.

caEnvironment - An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instance. For more information, see Environment Variables . There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 10 KB (10240 Bytes). This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 10KB)."

caEnableSSL - Whether to enable SSL for the app.

caShortname - The app's short name.

caDataSources - The app's data source.

caAppSource - A Source object that specifies the app repository.

caAttributes - One or more user-defined key/value pairs to be added to the stack attributes.

caDomains - The app virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'

caDescription - A description of the app.

caStackId - The stack ID.

caName - The app name.

caType - The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks Stacks deploys an application to those instances that are members of the corresponding layer. If your app isn't one of the standard types, or you prefer to implement your own Deploy recipes, specify other .

data CreateApp

See: createApp smart constructor.

Request Lenses

caSSLConfiguration :: Lens' CreateApp (Maybe SSLConfiguration)

An SslConfiguration object with the SSL configuration.

caEnvironment :: Lens' CreateApp [EnvironmentVariable]

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instance. For more information, see Environment Variables . There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 10 KB (10240 Bytes). This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 10KB)."

caEnableSSL :: Lens' CreateApp (Maybe Bool)

Whether to enable SSL for the app.

caShortname :: Lens' CreateApp (Maybe Text)

The app's short name.

caDataSources :: Lens' CreateApp [DataSource]

The app's data source.

caAppSource :: Lens' CreateApp (Maybe Source)

A Source object that specifies the app repository.

caAttributes :: Lens' CreateApp (HashMap AppAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

caDomains :: Lens' CreateApp [Text]

The app virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'

caDescription :: Lens' CreateApp (Maybe Text)

A description of the app.

caStackId :: Lens' CreateApp Text

The stack ID.

caName :: Lens' CreateApp Text

The app name.

caType :: Lens' CreateApp AppType

The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks Stacks deploys an application to those instances that are members of the corresponding layer. If your app isn't one of the standard types, or you prefer to implement your own Deploy recipes, specify other .

Destructuring the Response

createAppResponse

Arguments

:: Int
carsResponseStatus

-> CreateAppResponse

Creates a value of CreateAppResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

carsAppId - The app ID.

carsResponseStatus - -- | The response status code.

data CreateAppResponse

Contains the response to a CreateApp request.

See: createAppResponse smart constructor.

Response Lenses

carsAppId :: Lens' CreateAppResponse (Maybe Text)

The app ID.

carsResponseStatus :: Lens' CreateAppResponse Int

- | The response status code.