Network.AWS.OpsWorks.CreateApp

Description

Creates an app for a specified stack. For more information, see Creating Apps. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data CreateApp

Request constructor

createApp

Arguments

:: Text
caStackId

-> Text
caName

-> AppType
caType

-> CreateApp

Smart constructor for the minimum required parameters to construct a valid CreateApp request.

The fields accessible through corresponding lenses are:

StackId :: Text

Shortname :: Maybe Text

Name :: Text

Description :: Maybe Text

DataSources :: [DataSource]

Type :: AppType

AppSource :: Maybe Source'

Domains :: [Text]

EnableSsl :: Maybe Bool

SslConfiguration :: Maybe SslConfiguration

Attributes :: Map AppAttributesKeys Text

Environment :: [EnvironmentVariable]

Request lenses

caStackId :: Lens' CreateApp Text

The stack ID.

caShortname :: Lens' CreateApp (Maybe Text)

The app's short name.

caName :: Lens' CreateApp Text

The app name.

caDescription :: Lens' CreateApp (Maybe Text)

A description of the app.

caDataSources :: Lens' CreateApp [DataSource]

The app's data source.

caType :: Lens' CreateApp AppType

The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks deploys an application to those instances that are members of the corresponding layer.

caAppSource :: Lens' CreateApp (Maybe Source')

A Source object that specifies the app repository.

caDomains :: Lens' CreateApp [Text]

The app virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'.

caEnableSsl :: Lens' CreateApp (Maybe Bool)

Whether to enable SSL for the app.

caSslConfiguration :: Lens' CreateApp (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

caAttributes :: Lens' CreateApp (Map AppAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

caEnvironment :: Lens' CreateApp [EnvironmentVariable]

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. You can specify up to ten environment variables. After you deploy the app, these variables are defined on the associated app server instance. This parameter is supported only by Chef 11.10 stacks. If you have specified one or more environment variables, you cannot modify the stack's Chef version.

Response

data CreateAppResponse

Contains the response to a CreateApp request.

Response constructor

createAppResponse :: CreateAppResponse

Smart constructor for the minimum required parameters to construct a valid CreateAppResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AppId :: Maybe Text

Response lenses

carAppId :: Lens' CreateAppResponse (Maybe Text)

The app ID.