Network.AWS.OpsWorks.CreateInstance

Description

Creates an instance in a specified stack. For more information, see Adding an Instance to a Layer .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

createInstance

Arguments

:: Text
ciStackId

-> Text
ciInstanceType

-> CreateInstance

Creates a value of CreateInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciInstallUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true . To control when updates are installed, set this value to false . You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

ciVirtualizationType - The instance's virtualization type, paravirtual or hvm .

ciHostname - The instance host name.

ciSSHKeyName - The instance's Amazon EC2 key-pair name.

ciAgentVersion - The default AWS OpsWorks Stacks agent version. You have the following options: * INHERIT - Use the stack's default agent version setting. * version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance. The default setting is INHERIT . To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions . AgentVersion cannot be set to Chef 12.2.

ciSubnetId - The ID of the instance's subnet. If the stack is running in a VPC, you can use this parameter to override the stack's default subnet ID value and direct AWS OpsWorks Stacks to launch the instance in a different subnet.

ciEBSOptimized - Whether to create an Amazon EBS-optimized instance.

ciOS - The instance's operating system, which must be set to one of the following. * A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2017.03 , Amazon Linux 2016.09 , Amazon Linux 2016.03 , Amazon Linux 2015.09 , or Amazon Linux 2015.03 . * A supported Ubuntu operating system, such as Ubuntu 16.04 LTS , Ubuntu 14.04 LTS , or Ubuntu 12.04 LTS . * CentOS Linux 7 * Red Hat Enterprise Linux 7 * A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base , Microsoft Windows Server 2012 R2 with SQL Server Express , Microsoft Windows Server 2012 R2 with SQL Server Standard , or Microsoft Windows Server 2012 R2 with SQL Server Web . * A custom AMI: Custom . For more information on the supported operating systems, see AWS OpsWorks Stacks Operating Systems . The default option is the current Amazon Linux version. If you set this parameter to Custom , you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. Block device mappings are not supported if the value is Custom . For more information on the supported operating systems, see Operating Systems For more information on how to use custom AMIs with AWS OpsWorks Stacks, see Using Custom AMIs .

ciAvailabilityZone - The instance Availability Zone. For more information, see Regions and Endpoints .

ciTenancy - The instance's tenancy option. The default option is no tenancy, or if the instance is running in a VPC, inherit tenancy settings from the VPC. The following are valid values for this parameter: dedicated , default , or host . Because there are costs associated with changes in tenancy options, we recommend that you research tenancy options before choosing them for your instances. For more information about dedicated hosts, see Dedicated Hosts Overview and Amazon EC2 Dedicated Hosts . For more information about dedicated instances, see Dedicated Instances and Amazon EC2 Dedicated Instances .

ciAutoScalingType - For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

ciArchitecture - The instance architecture. The default option is x86_64 . Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types .

ciAMIId - A custom AMI ID to be used to create the instance. The AMI should be based on one of the supported operating systems. For more information, see Using Custom AMIs .

ciRootDeviceType - The instance root device type. For more information, see Storage for the Root Device .

ciBlockDeviceMappings - An array of BlockDeviceMapping objects that specify the instance's block devices. For more information, see Block Device Mapping . Note that block device mappings are not supported for custom AMIs.

ciStackId - The stack ID.

ciLayerIds - An array that contains the instance's layer IDs.

ciInstanceType - The instance type, such as t2.micro . For a list of supported instance types, open the stack in the console, choose Instances , and choose + Instance . The Size list contains the currently supported types. For more information, see Instance Families and Types . The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

data CreateInstance

See: createInstance smart constructor.

Request Lenses

ciInstallUpdatesOnBoot :: Lens' CreateInstance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true . To control when updates are installed, set this value to false . You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

ciVirtualizationType :: Lens' CreateInstance (Maybe Text)

The instance's virtualization type, paravirtual or hvm .

ciHostname :: Lens' CreateInstance (Maybe Text)

The instance host name.

ciSSHKeyName :: Lens' CreateInstance (Maybe Text)

The instance's Amazon EC2 key-pair name.

ciAgentVersion :: Lens' CreateInstance (Maybe Text)

The default AWS OpsWorks Stacks agent version. You have the following options: * INHERIT - Use the stack's default agent version setting. * version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance. The default setting is INHERIT . To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions . AgentVersion cannot be set to Chef 12.2.

ciSubnetId :: Lens' CreateInstance (Maybe Text)

The ID of the instance's subnet. If the stack is running in a VPC, you can use this parameter to override the stack's default subnet ID value and direct AWS OpsWorks Stacks to launch the instance in a different subnet.

ciEBSOptimized :: Lens' CreateInstance (Maybe Bool)

Whether to create an Amazon EBS-optimized instance.

ciOS :: Lens' CreateInstance (Maybe Text)

The instance's operating system, which must be set to one of the following. * A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2017.03 , Amazon Linux 2016.09 , Amazon Linux 2016.03 , Amazon Linux 2015.09 , or Amazon Linux 2015.03 . * A supported Ubuntu operating system, such as Ubuntu 16.04 LTS , Ubuntu 14.04 LTS , or Ubuntu 12.04 LTS . * CentOS Linux 7 * Red Hat Enterprise Linux 7 * A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base , Microsoft Windows Server 2012 R2 with SQL Server Express , Microsoft Windows Server 2012 R2 with SQL Server Standard , or Microsoft Windows Server 2012 R2 with SQL Server Web . * A custom AMI: Custom . For more information on the supported operating systems, see AWS OpsWorks Stacks Operating Systems . The default option is the current Amazon Linux version. If you set this parameter to Custom , you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. Block device mappings are not supported if the value is Custom . For more information on the supported operating systems, see Operating Systems For more information on how to use custom AMIs with AWS OpsWorks Stacks, see Using Custom AMIs .

ciAvailabilityZone :: Lens' CreateInstance (Maybe Text)

The instance Availability Zone. For more information, see Regions and Endpoints .

ciTenancy :: Lens' CreateInstance (Maybe Text)

The instance's tenancy option. The default option is no tenancy, or if the instance is running in a VPC, inherit tenancy settings from the VPC. The following are valid values for this parameter: dedicated , default , or host . Because there are costs associated with changes in tenancy options, we recommend that you research tenancy options before choosing them for your instances. For more information about dedicated hosts, see Dedicated Hosts Overview and Amazon EC2 Dedicated Hosts . For more information about dedicated instances, see Dedicated Instances and Amazon EC2 Dedicated Instances .

ciAutoScalingType :: Lens' CreateInstance (Maybe AutoScalingType)

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

ciArchitecture :: Lens' CreateInstance (Maybe Architecture)

The instance architecture. The default option is x86_64 . Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types .

ciAMIId :: Lens' CreateInstance (Maybe Text)

A custom AMI ID to be used to create the instance. The AMI should be based on one of the supported operating systems. For more information, see Using Custom AMIs .

ciRootDeviceType :: Lens' CreateInstance (Maybe RootDeviceType)

The instance root device type. For more information, see Storage for the Root Device .

ciBlockDeviceMappings :: Lens' CreateInstance [BlockDeviceMapping]

An array of BlockDeviceMapping objects that specify the instance's block devices. For more information, see Block Device Mapping . Note that block device mappings are not supported for custom AMIs.

ciStackId :: Lens' CreateInstance Text

The stack ID.

ciLayerIds :: Lens' CreateInstance [Text]

An array that contains the instance's layer IDs.

ciInstanceType :: Lens' CreateInstance Text

The instance type, such as t2.micro . For a list of supported instance types, open the stack in the console, choose Instances , and choose + Instance . The Size list contains the currently supported types. For more information, see Instance Families and Types . The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

Destructuring the Response

createInstanceResponse

Arguments

:: Int
cirsResponseStatus

-> CreateInstanceResponse

Creates a value of CreateInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cirsInstanceId - The instance ID.

cirsResponseStatus - -- | The response status code.

data CreateInstanceResponse

Contains the response to a CreateInstance request.

See: createInstanceResponse smart constructor.

Response Lenses

cirsInstanceId :: Lens' CreateInstanceResponse (Maybe Text)

The instance ID.

cirsResponseStatus :: Lens' CreateInstanceResponse Int

- | The response status code.