Network.AWS.OpsWorks.CreateLayer

Description

Creates a layer. For more information, see How to Create a Layer .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

createLayer

Arguments

:: Text
clStackId

-> LayerType
clType

-> Text
clName

-> Text
clShortname

-> CreateLayer

Creates a value of CreateLayer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clCustomInstanceProfileARN - The ARN of an IAM profile to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers .

clCustomSecurityGroupIds - An array containing the layer custom security group IDs.

clInstallUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true . To control when updates are installed, set this value to false . You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

clCloudWatchLogsConfiguration - Specifies CloudWatch Logs configuration options for the layer. For more information, see CloudWatchLogsLogStream .

clLifecycleEventConfiguration - A LifeCycleEventConfiguration object that you can use to configure the Shutdown event to specify an execution timeout and enable or disable Elastic Load Balancer connection draining.

clCustomRecipes - A LayerCustomRecipes object that specifies the layer custom recipes.

clCustomJSON - A JSON-formatted string containing custom stack configuration and deployment attributes to be installed on the layer's instances. For more information, see Using Custom JSON . This feature is supported as of version 1.7.42 of the AWS CLI.

clVolumeConfigurations - A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

clEnableAutoHealing - Whether to disable auto healing for the layer.

clPackages - An array of Package objects that describes the layer packages.

clAttributes - One or more user-defined key-value pairs to be added to the stack attributes. To create a cluster layer, set the EcsClusterArn attribute to the cluster's ARN.

clAutoAssignPublicIPs - For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer .

clUseEBSOptimizedInstances - Whether to use Amazon EBS-optimized instances.

clAutoAssignElasticIPs - Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer .

clStackId - The layer stack ID.

clType - The layer type. A stack cannot have more than one built-in layer of the same type. It can have any number of custom layers. Built-in layers are not available in Chef 12 stacks.

clName - The layer name, which is used by the console.

clShortname - For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorks Stacks and by Chef recipes. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters, which are limited to the alphanumeric characters, -, '_', and .. The built-in layers' short names are defined by AWS OpsWorks Stacks. For more information, see the Layer Reference .

data CreateLayer

See: createLayer smart constructor.

Request Lenses

clCustomInstanceProfileARN :: Lens' CreateLayer (Maybe Text)

The ARN of an IAM profile to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers .

clCustomSecurityGroupIds :: Lens' CreateLayer [Text]

An array containing the layer custom security group IDs.

clInstallUpdatesOnBoot :: Lens' CreateLayer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true . To control when updates are installed, set this value to false . You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

clCloudWatchLogsConfiguration :: Lens' CreateLayer (Maybe CloudWatchLogsConfiguration)

Specifies CloudWatch Logs configuration options for the layer. For more information, see CloudWatchLogsLogStream .

clLifecycleEventConfiguration :: Lens' CreateLayer (Maybe LifecycleEventConfiguration)

A LifeCycleEventConfiguration object that you can use to configure the Shutdown event to specify an execution timeout and enable or disable Elastic Load Balancer connection draining.

clCustomRecipes :: Lens' CreateLayer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer custom recipes.

clCustomJSON :: Lens' CreateLayer (Maybe Text)

A JSON-formatted string containing custom stack configuration and deployment attributes to be installed on the layer's instances. For more information, see Using Custom JSON . This feature is supported as of version 1.7.42 of the AWS CLI.

clVolumeConfigurations :: Lens' CreateLayer [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

clEnableAutoHealing :: Lens' CreateLayer (Maybe Bool)

Whether to disable auto healing for the layer.

clPackages :: Lens' CreateLayer [Text]

An array of Package objects that describes the layer packages.

clAttributes :: Lens' CreateLayer (HashMap LayerAttributesKeys Text)

One or more user-defined key-value pairs to be added to the stack attributes. To create a cluster layer, set the EcsClusterArn attribute to the cluster's ARN.

clAutoAssignPublicIPs :: Lens' CreateLayer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer .

clUseEBSOptimizedInstances :: Lens' CreateLayer (Maybe Bool)

Whether to use Amazon EBS-optimized instances.

clAutoAssignElasticIPs :: Lens' CreateLayer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer .

clStackId :: Lens' CreateLayer Text

The layer stack ID.

clType :: Lens' CreateLayer LayerType

The layer type. A stack cannot have more than one built-in layer of the same type. It can have any number of custom layers. Built-in layers are not available in Chef 12 stacks.

clName :: Lens' CreateLayer Text

The layer name, which is used by the console.

clShortname :: Lens' CreateLayer Text

For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorks Stacks and by Chef recipes. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters, which are limited to the alphanumeric characters, -, '_', and .. The built-in layers' short names are defined by AWS OpsWorks Stacks. For more information, see the Layer Reference .

Destructuring the Response

createLayerResponse

Arguments

:: Int
clrsResponseStatus

-> CreateLayerResponse

Creates a value of CreateLayerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clrsLayerId - The layer ID.

clrsResponseStatus - -- | The response status code.

data CreateLayerResponse

Contains the response to a CreateLayer request.

See: createLayerResponse smart constructor.

Response Lenses

clrsLayerId :: Lens' CreateLayerResponse (Maybe Text)

The layer ID.

clrsResponseStatus :: Lens' CreateLayerResponse Int

- | The response status code.