Network.AWS.OpsWorks.CreateLayer

Description

Creates a layer. For more information, see How to Create a Layer.

You should use CreateLayer for noncustom layer types such as PHP App Server only if the stack does not have an existing layer of that type. A stack can have at most one instance of each noncustom layer; if you attempt to create a second instance, CreateLayer fails. A stack can have an arbitrary number of custom layers, so you can call CreateLayer as many times as you like for that layer type.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_CreateLayer.html

Request

data CreateLayer

Request constructor

createLayer

Arguments

:: Text
clStackId

-> LayerType
clType

-> Text
clName

-> Text
clShortname

-> CreateLayer

CreateLayer constructor.

The fields accessible through corresponding lenses are:

clAttributes :: HashMap LayerAttributesKeys Text

clAutoAssignElasticIps :: Maybe Bool

clAutoAssignPublicIps :: Maybe Bool

clCustomInstanceProfileArn :: Maybe Text

clCustomRecipes :: Maybe Recipes

clCustomSecurityGroupIds :: [Text]

clEnableAutoHealing :: Maybe Bool

clInstallUpdatesOnBoot :: Maybe Bool

clLifecycleEventConfiguration :: Maybe LifecycleEventConfiguration

clName :: Text

clPackages :: [Text]

clShortname :: Text

clStackId :: Text

clType :: LayerType

clUseEbsOptimizedInstances :: Maybe Bool

clVolumeConfigurations :: [VolumeConfiguration]

Request lenses

clAttributes :: Lens' CreateLayer (HashMap LayerAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

clAutoAssignElasticIps :: Lens' CreateLayer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

clAutoAssignPublicIps :: Lens' CreateLayer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How toEdit a Layer.

clCustomInstanceProfileArn :: Lens' CreateLayer (Maybe Text)

The ARN of an IAM profile that to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

clCustomRecipes :: Lens' CreateLayer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer custom recipes.

clCustomSecurityGroupIds :: Lens' CreateLayer [Text]

An array containing the layer custom security group IDs.

clEnableAutoHealing :: Lens' CreateLayer (Maybe Bool)

Whether to disable auto healing for the layer.

clInstallUpdatesOnBoot :: Lens' CreateLayer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or 'apt-get' (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

clLifecycleEventConfiguration :: Lens' CreateLayer (Maybe LifecycleEventConfiguration)

A LifeCycleEventConfiguration object that you can use to configure the Shutdown event to specify an execution timeout and enable or disable Elastic Load Balancer connection draining.

clName :: Lens' CreateLayer Text

The layer name, which is used by the console.

clPackages :: Lens' CreateLayer [Text]

An array of Package objects that describe the layer packages.

clShortname :: Lens' CreateLayer Text

For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorks and by Chef recipes. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters, which are limited to the alphanumeric characters, -, '_', and ..

The built-in layers' short names are defined by AWS OpsWorks. For more information, see the Layer Reference

clStackId :: Lens' CreateLayer Text

The layer stack ID.

clType :: Lens' CreateLayer LayerType

The layer type. A stack cannot have more than one built-in layer of the same type. It can have any number of custom layers.

clUseEbsOptimizedInstances :: Lens' CreateLayer (Maybe Bool)

Whether to use Amazon EBS-optimized instances.

clVolumeConfigurations :: Lens' CreateLayer [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

Response

data CreateLayerResponse

Response constructor

createLayerResponse :: CreateLayerResponse

CreateLayerResponse constructor.

The fields accessible through corresponding lenses are:

clrLayerId :: Maybe Text

Response lenses

clrLayerId :: Lens' CreateLayerResponse (Maybe Text)

The layer ID.