Network.AWS.OpsWorks.CreateStack

Description

Creates a new stack. For more information, see Create a New Stack .

Required Permissions : To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

createStack

Arguments

:: Text
csName

-> Text
csRegion

-> Text
csServiceRoleARN

-> Text
csDefaultInstanceProfileARN

-> CreateStack

Creates a value of CreateStack with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csDefaultRootDeviceType - The default root device type. This value is the default for all instances in the stack, but you can override it when you create an instance. The default option is instance-store . For more information, see Storage for the Root Device .

csVPCId - The ID of the VPC that the stack is to be launched into. The VPC must be in the stack's region. All instances are launched into this VPC. You cannot change the ID later. * If your account supports EC2-Classic, the default value is no VPC . * If your account does not support EC2-Classic, the default value is the default VPC for the specified region. If the VPC ID corresponds to a default VPC and you have specified either the DefaultAvailabilityZone or the DefaultSubnetId parameter only, AWS OpsWorks Stacks infers the value of the other parameter. If you specify neither parameter, AWS OpsWorks Stacks sets these parameters to the first valid Availability Zone for the specified region and the corresponding default VPC subnet ID, respectively. If you specify a nondefault VPC ID, note the following: * It must belong to a VPC in your account that is in the specified region. * You must specify a value for DefaultSubnetId . For more information on how to use AWS OpsWorks Stacks with a VPC, see Running a Stack in a VPC . For more information on default VPC and EC2-Classic, see Supported Platforms .

csChefConfiguration - A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version on Chef 11.10 stacks. For more information, see Create a New Stack .

csAgentVersion - The default AWS OpsWorks Stacks agent version. You have the following options: * Auto-update - Set this parameter to LATEST . AWS OpsWorks Stacks automatically installs new agent versions on the stack's instances as soon as they are available. * Fixed version - Set this parameter to your preferred agent version. To update the agent version, you must edit the stack configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the stack's instances. The default setting is the most recent release of the agent. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions . AgentVersion cannot be set to Chef 12.2.

csDefaultSSHKeyName - A default Amazon EC2 key pair name. The default value is none. If you specify a key pair name, AWS OpsWorks installs the public key on the instance and you can use the private key with an SSH client to log in to the instance. For more information, see Using SSH to Communicate with an Instance and Managing SSH Access . You can override this setting by specifying a different key pair, or no key pair, when you create an instance .

csCustomJSON - A string that contains user-defined, custom JSON. It can be used to override the corresponding default stack configuration attribute values or to pass data to recipes. The string should be in the following format: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes .

csCustomCookbooksSource - Undocumented member.

csDefaultAvailabilityZone - The stack's default Availability Zone, which must be in the specified region. For more information, see Regions and Endpoints . If you also specify a value for DefaultSubnetId , the subnet must be in the same zone. For more information, see the VpcId parameter description.

csAttributes - One or more user-defined key-value pairs to be added to the stack attributes.

csDefaultOS - The stack's default operating system, which is installed on every instance unless you specify a different operating system when you create the instance. You can specify one of the following. * A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2017.03 , Amazon Linux 2016.09 , Amazon Linux 2016.03 , Amazon Linux 2015.09 , or Amazon Linux 2015.03 . * A supported Ubuntu operating system, such as Ubuntu 16.04 LTS , Ubuntu 14.04 LTS , or Ubuntu 12.04 LTS . * CentOS Linux 7 * Red Hat Enterprise Linux 7 * A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base , Microsoft Windows Server 2012 R2 with SQL Server Express , Microsoft Windows Server 2012 R2 with SQL Server Standard , or Microsoft Windows Server 2012 R2 with SQL Server Web . * A custom AMI: Custom . You specify the custom AMI you want to use when you create instances. For more information, see Using Custom AMIs . The default option is the current Amazon Linux version. For more information on the supported operating systems, see AWS OpsWorks Stacks Operating Systems .

csUseOpsworksSecurityGroups - Whether to associate the AWS OpsWorks Stacks built-in security groups with the stack's layers. AWS OpsWorks Stacks provides a standard set of built-in security groups, one for each layer, which are associated with layers by default. With UseOpsworksSecurityGroups you can instead provide your own custom security groups. UseOpsworksSecurityGroups has the following settings: * True - AWS OpsWorks Stacks automatically associates the appropriate built-in security group with each layer (default setting). You can associate additional security groups with a layer after you create it, but you cannot delete the built-in security group. * False - AWS OpsWorks Stacks does not associate built-in security groups with layers. You must create appropriate EC2 security groups and associate a security group with each layer that you create. However, you can still manually associate a built-in security group with a layer on creation; custom security groups are required only for those layers that need custom settings. For more information, see Create a New Stack .

csUseCustomCookbooks - Whether the stack uses custom cookbooks.

csDefaultSubnetId - The stack's default VPC subnet ID. This parameter is required if you specify a value for the VpcId parameter. All instances are launched into this subnet unless you specify otherwise when you create the instance. If you also specify a value for DefaultAvailabilityZone , the subnet must be in that zone. For information on default values and when this parameter is required, see the VpcId parameter description.

csConfigurationManager - The configuration manager. When you create a stack we recommend that you use the configuration manager to specify the Chef version: 12, 11.10, or 11.4 for Linux stacks, or 12.2 for Windows stacks. The default value for Linux stacks is currently 11.4.

csHostnameTheme - The stack's host name theme, with spaces replaced by underscores. The theme is used to generate host names for the stack's instances. By default, HostnameTheme is set to Layer_Dependent , which creates host names by appending integers to the layer's short name. The other themes are: * Baked_Goods * Clouds * Europe_Cities * Fruits * Greek_Deities * Legendary_creatures_from_Japan * Planets_and_Moons * Roman_Deities * Scottish_Islands * US_Cities * Wild_Cats To obtain a generated host name, call GetHostNameSuggestion , which returns a host name based on the current theme.

csName - The stack name.

csRegion - The stack's AWS region, such as "ap-south-1". For more information about Amazon regions, see Regions and Endpoints .

csServiceRoleARN - The stack's AWS Identity and Access Management (IAM) role, which allows AWS OpsWorks Stacks to work with AWS resources on your behalf. You must set this parameter to the Amazon Resource Name (ARN) for an existing IAM role. For more information about IAM ARNs, see Using Identifiers .

csDefaultInstanceProfileARN - The Amazon Resource Name (ARN) of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers .

data CreateStack

See: createStack smart constructor.

Request Lenses

csDefaultRootDeviceType :: Lens' CreateStack (Maybe RootDeviceType)

The default root device type. This value is the default for all instances in the stack, but you can override it when you create an instance. The default option is instance-store . For more information, see Storage for the Root Device .

csVPCId :: Lens' CreateStack (Maybe Text)

The ID of the VPC that the stack is to be launched into. The VPC must be in the stack's region. All instances are launched into this VPC. You cannot change the ID later. * If your account supports EC2-Classic, the default value is no VPC . * If your account does not support EC2-Classic, the default value is the default VPC for the specified region. If the VPC ID corresponds to a default VPC and you have specified either the DefaultAvailabilityZone or the DefaultSubnetId parameter only, AWS OpsWorks Stacks infers the value of the other parameter. If you specify neither parameter, AWS OpsWorks Stacks sets these parameters to the first valid Availability Zone for the specified region and the corresponding default VPC subnet ID, respectively. If you specify a nondefault VPC ID, note the following: * It must belong to a VPC in your account that is in the specified region. * You must specify a value for DefaultSubnetId . For more information on how to use AWS OpsWorks Stacks with a VPC, see Running a Stack in a VPC . For more information on default VPC and EC2-Classic, see Supported Platforms .

csChefConfiguration :: Lens' CreateStack (Maybe ChefConfiguration)

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version on Chef 11.10 stacks. For more information, see Create a New Stack .

csAgentVersion :: Lens' CreateStack (Maybe Text)

The default AWS OpsWorks Stacks agent version. You have the following options: * Auto-update - Set this parameter to LATEST . AWS OpsWorks Stacks automatically installs new agent versions on the stack's instances as soon as they are available. * Fixed version - Set this parameter to your preferred agent version. To update the agent version, you must edit the stack configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the stack's instances. The default setting is the most recent release of the agent. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions . AgentVersion cannot be set to Chef 12.2.

csDefaultSSHKeyName :: Lens' CreateStack (Maybe Text)

A default Amazon EC2 key pair name. The default value is none. If you specify a key pair name, AWS OpsWorks installs the public key on the instance and you can use the private key with an SSH client to log in to the instance. For more information, see Using SSH to Communicate with an Instance and Managing SSH Access . You can override this setting by specifying a different key pair, or no key pair, when you create an instance .

csCustomJSON :: Lens' CreateStack (Maybe Text)

A string that contains user-defined, custom JSON. It can be used to override the corresponding default stack configuration attribute values or to pass data to recipes. The string should be in the following format: "{"key1": "value1", "key2": "value2",...}" For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes .

csCustomCookbooksSource :: Lens' CreateStack (Maybe Source)

Undocumented member.

csDefaultAvailabilityZone :: Lens' CreateStack (Maybe Text)

The stack's default Availability Zone, which must be in the specified region. For more information, see Regions and Endpoints . If you also specify a value for DefaultSubnetId , the subnet must be in the same zone. For more information, see the VpcId parameter description.

csAttributes :: Lens' CreateStack (HashMap StackAttributesKeys Text)

One or more user-defined key-value pairs to be added to the stack attributes.

csDefaultOS :: Lens' CreateStack (Maybe Text)

The stack's default operating system, which is installed on every instance unless you specify a different operating system when you create the instance. You can specify one of the following. * A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2017.03 , Amazon Linux 2016.09 , Amazon Linux 2016.03 , Amazon Linux 2015.09 , or Amazon Linux 2015.03 . * A supported Ubuntu operating system, such as Ubuntu 16.04 LTS , Ubuntu 14.04 LTS , or Ubuntu 12.04 LTS . * CentOS Linux 7 * Red Hat Enterprise Linux 7 * A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base , Microsoft Windows Server 2012 R2 with SQL Server Express , Microsoft Windows Server 2012 R2 with SQL Server Standard , or Microsoft Windows Server 2012 R2 with SQL Server Web . * A custom AMI: Custom . You specify the custom AMI you want to use when you create instances. For more information, see Using Custom AMIs . The default option is the current Amazon Linux version. For more information on the supported operating systems, see AWS OpsWorks Stacks Operating Systems .

csUseOpsworksSecurityGroups :: Lens' CreateStack (Maybe Bool)

Whether to associate the AWS OpsWorks Stacks built-in security groups with the stack's layers. AWS OpsWorks Stacks provides a standard set of built-in security groups, one for each layer, which are associated with layers by default. With UseOpsworksSecurityGroups you can instead provide your own custom security groups. UseOpsworksSecurityGroups has the following settings: * True - AWS OpsWorks Stacks automatically associates the appropriate built-in security group with each layer (default setting). You can associate additional security groups with a layer after you create it, but you cannot delete the built-in security group. * False - AWS OpsWorks Stacks does not associate built-in security groups with layers. You must create appropriate EC2 security groups and associate a security group with each layer that you create. However, you can still manually associate a built-in security group with a layer on creation; custom security groups are required only for those layers that need custom settings. For more information, see Create a New Stack .

csUseCustomCookbooks :: Lens' CreateStack (Maybe Bool)

Whether the stack uses custom cookbooks.

csDefaultSubnetId :: Lens' CreateStack (Maybe Text)

The stack's default VPC subnet ID. This parameter is required if you specify a value for the VpcId parameter. All instances are launched into this subnet unless you specify otherwise when you create the instance. If you also specify a value for DefaultAvailabilityZone , the subnet must be in that zone. For information on default values and when this parameter is required, see the VpcId parameter description.

csConfigurationManager :: Lens' CreateStack (Maybe StackConfigurationManager)

The configuration manager. When you create a stack we recommend that you use the configuration manager to specify the Chef version: 12, 11.10, or 11.4 for Linux stacks, or 12.2 for Windows stacks. The default value for Linux stacks is currently 11.4.

csHostnameTheme :: Lens' CreateStack (Maybe Text)

The stack's host name theme, with spaces replaced by underscores. The theme is used to generate host names for the stack's instances. By default, HostnameTheme is set to Layer_Dependent , which creates host names by appending integers to the layer's short name. The other themes are: * Baked_Goods * Clouds * Europe_Cities * Fruits * Greek_Deities * Legendary_creatures_from_Japan * Planets_and_Moons * Roman_Deities * Scottish_Islands * US_Cities * Wild_Cats To obtain a generated host name, call GetHostNameSuggestion , which returns a host name based on the current theme.

csName :: Lens' CreateStack Text

The stack name.

csRegion :: Lens' CreateStack Text

The stack's AWS region, such as "ap-south-1". For more information about Amazon regions, see Regions and Endpoints .

csServiceRoleARN :: Lens' CreateStack Text

The stack's AWS Identity and Access Management (IAM) role, which allows AWS OpsWorks Stacks to work with AWS resources on your behalf. You must set this parameter to the Amazon Resource Name (ARN) for an existing IAM role. For more information about IAM ARNs, see Using Identifiers .

csDefaultInstanceProfileARN :: Lens' CreateStack Text

The Amazon Resource Name (ARN) of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers .

Destructuring the Response

createStackResponse

Arguments

:: Int
crsResponseStatus

-> CreateStackResponse

Creates a value of CreateStackResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsStackId - The stack ID, which is an opaque string that you use to identify the stack when performing actions such as DescribeStacks .

crsResponseStatus - -- | The response status code.

data CreateStackResponse

Contains the response to a CreateStack request.

See: createStackResponse smart constructor.

Response Lenses

crsStackId :: Lens' CreateStackResponse (Maybe Text)

The stack ID, which is an opaque string that you use to identify the stack when performing actions such as DescribeStacks .

crsResponseStatus :: Lens' CreateStackResponse Int

- | The response status code.