Network.AWS.OpsWorks.CreateUserProfile

Description

Creates a new user profile.

Required Permissions : To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

createUserProfile

Arguments

:: Text
cupIAMUserARN

-> CreateUserProfile

Creates a value of CreateUserProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cupAllowSelfManagement - Whether users can specify their own SSH public key through the My Settings page. For more information, see Setting an IAM User's Public SSH Key .

cupSSHPublicKey - The user's public SSH key.

cupSSHUsername - The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], -, and '_'. If the specified name includes other punctuation marks, AWS OpsWorks Stacks removes them. For example, my.name will be changed to myname . If you do not specify an SSH user name, AWS OpsWorks Stacks generates one from the IAM user name.

cupIAMUserARN - The user's IAM ARN; this can also be a federated user's ARN.

data CreateUserProfile

See: createUserProfile smart constructor.

Request Lenses

cupAllowSelfManagement :: Lens' CreateUserProfile (Maybe Bool)

Whether users can specify their own SSH public key through the My Settings page. For more information, see Setting an IAM User's Public SSH Key .

cupSSHPublicKey :: Lens' CreateUserProfile (Maybe Text)

The user's public SSH key.

cupSSHUsername :: Lens' CreateUserProfile (Maybe Text)

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], -, and '_'. If the specified name includes other punctuation marks, AWS OpsWorks Stacks removes them. For example, my.name will be changed to myname . If you do not specify an SSH user name, AWS OpsWorks Stacks generates one from the IAM user name.

cupIAMUserARN :: Lens' CreateUserProfile Text

The user's IAM ARN; this can also be a federated user's ARN.

Destructuring the Response

createUserProfileResponse

Arguments

:: Int
cuprsResponseStatus

-> CreateUserProfileResponse

Creates a value of CreateUserProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cuprsIAMUserARN - The user's IAM ARN.

cuprsResponseStatus - -- | The response status code.

data CreateUserProfileResponse

Contains the response to a CreateUserProfile request.

See: createUserProfileResponse smart constructor.

Response Lenses

cuprsIAMUserARN :: Lens' CreateUserProfileResponse (Maybe Text)

The user's IAM ARN.

cuprsResponseStatus :: Lens' CreateUserProfileResponse Int

- | The response status code.