Network.AWS.OpsWorks.DescribeApps

Description

Requests a description of a specified set of apps.

Required Permissions : To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

describeApps :: DescribeApps

Creates a value of DescribeApps with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daAppIds - An array of app IDs for the apps to be described. If you use this parameter, DescribeApps returns a description of the specified apps. Otherwise, it returns a description of every app.

daStackId - The app stack ID. If you use this parameter, DescribeApps returns a description of the apps in the specified stack.

data DescribeApps

See: describeApps smart constructor.

Request Lenses

daAppIds :: Lens' DescribeApps [Text]

An array of app IDs for the apps to be described. If you use this parameter, DescribeApps returns a description of the specified apps. Otherwise, it returns a description of every app.

daStackId :: Lens' DescribeApps (Maybe Text)

The app stack ID. If you use this parameter, DescribeApps returns a description of the apps in the specified stack.

Destructuring the Response

describeAppsResponse

Arguments

:: Int
darsResponseStatus

-> DescribeAppsResponse

Creates a value of DescribeAppsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsApps - An array of App objects that describe the specified apps.

darsResponseStatus - -- | The response status code.

data DescribeAppsResponse

Contains the response to a DescribeApps request.

See: describeAppsResponse smart constructor.

Response Lenses

darsApps :: Lens' DescribeAppsResponse [App]

An array of App objects that describe the specified apps.

darsResponseStatus :: Lens' DescribeAppsResponse Int

- | The response status code.