Network.AWS.OpsWorks.DescribeInstances

Description

Requests a description of a set of instances.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeInstances.html

Request

data DescribeInstances

Request constructor

describeInstances :: DescribeInstances

DescribeInstances constructor.

The fields accessible through corresponding lenses are:

diInstanceIds :: [Text]

diLayerId :: Maybe Text

diStackId :: Maybe Text

Request lenses

diInstanceIds :: Lens' DescribeInstances [Text]

An array of instance IDs to be described. If you use this parameter, DescribeInstances returns a description of the specified instances. Otherwise, it returns a description of every instance.

diLayerId :: Lens' DescribeInstances (Maybe Text)

A layer ID. If you use this parameter, DescribeInstances returns descriptions of the instances associated with the specified layer.

diStackId :: Lens' DescribeInstances (Maybe Text)

A stack ID. If you use this parameter, DescribeInstances returns descriptions of the instances associated with the specified stack.

Response

data DescribeInstancesResponse

Response constructor

describeInstancesResponse :: DescribeInstancesResponse

DescribeInstancesResponse constructor.

The fields accessible through corresponding lenses are:

dirInstances :: [Instance]

Response lenses

dirInstances :: Lens' DescribeInstancesResponse [Instance]

An array of Instance objects that describe the instances.