Network.AWS.OpsWorks.DescribeLoadBasedAutoScaling

Description

Describes load-based auto scaling configurations for specified layers.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeLoadBasedAutoScaling.html

Request

data DescribeLoadBasedAutoScaling

Request constructor

describeLoadBasedAutoScaling :: DescribeLoadBasedAutoScaling

DescribeLoadBasedAutoScaling constructor.

The fields accessible through corresponding lenses are:

dlbasLayerIds :: [Text]

Request lenses

dlbasLayerIds :: Lens' DescribeLoadBasedAutoScaling [Text]

An array of layer IDs.

Response

data DescribeLoadBasedAutoScalingResponse

Response constructor

describeLoadBasedAutoScalingResponse :: DescribeLoadBasedAutoScalingResponse

DescribeLoadBasedAutoScalingResponse constructor.

The fields accessible through corresponding lenses are:

dlbasrLoadBasedAutoScalingConfigurations :: [LoadBasedAutoScalingConfiguration]

Response lenses

dlbasrLoadBasedAutoScalingConfigurations :: Lens' DescribeLoadBasedAutoScalingResponse [LoadBasedAutoScalingConfiguration]

An array of LoadBasedAutoScalingConfiguration objects that describe each layer's configuration.