Network.AWS.OpsWorks.DescribeMyUserProfile

Description

Describes a user's SSH information.

Required Permissions : To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

describeMyUserProfile :: DescribeMyUserProfile

Creates a value of DescribeMyUserProfile with the minimum fields required to make a request.

data DescribeMyUserProfile

See: describeMyUserProfile smart constructor.

Destructuring the Response

describeMyUserProfileResponse

Arguments

:: Int
dmuprsResponseStatus

-> DescribeMyUserProfileResponse

Creates a value of DescribeMyUserProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmuprsUserProfile - A UserProfile object that describes the user's SSH information.

dmuprsResponseStatus - -- | The response status code.

data DescribeMyUserProfileResponse

Contains the response to a DescribeMyUserProfile request.

See: describeMyUserProfileResponse smart constructor.

Response Lenses

dmuprsUserProfile :: Lens' DescribeMyUserProfileResponse (Maybe SelfUserProfile)

A UserProfile object that describes the user's SSH information.

dmuprsResponseStatus :: Lens' DescribeMyUserProfileResponse Int

- | The response status code.