Network.AWS.OpsWorks.DescribeServiceErrors

Description

Describes AWS OpsWorks Stacks service errors.

Required Permissions : To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

This call accepts only one resource-identifying parameter.

Creating a Request

describeServiceErrors :: DescribeServiceErrors

Creates a value of DescribeServiceErrors with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dseInstanceId - The instance ID. If you use this parameter, DescribeServiceErrors returns descriptions of the errors associated with the specified instance.

dseStackId - The stack ID. If you use this parameter, DescribeServiceErrors returns descriptions of the errors associated with the specified stack.

dseServiceErrorIds - An array of service error IDs. If you use this parameter, DescribeServiceErrors returns descriptions of the specified errors. Otherwise, it returns a description of every error.

data DescribeServiceErrors

See: describeServiceErrors smart constructor.

Request Lenses

dseInstanceId :: Lens' DescribeServiceErrors (Maybe Text)

The instance ID. If you use this parameter, DescribeServiceErrors returns descriptions of the errors associated with the specified instance.

dseStackId :: Lens' DescribeServiceErrors (Maybe Text)

The stack ID. If you use this parameter, DescribeServiceErrors returns descriptions of the errors associated with the specified stack.

dseServiceErrorIds :: Lens' DescribeServiceErrors [Text]

An array of service error IDs. If you use this parameter, DescribeServiceErrors returns descriptions of the specified errors. Otherwise, it returns a description of every error.

Destructuring the Response

describeServiceErrorsResponse

Arguments

:: Int
dsersResponseStatus

-> DescribeServiceErrorsResponse

Creates a value of DescribeServiceErrorsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsersServiceErrors - An array of ServiceError objects that describe the specified service errors.

dsersResponseStatus - -- | The response status code.

data DescribeServiceErrorsResponse

Contains the response to a DescribeServiceErrors request.

See: describeServiceErrorsResponse smart constructor.

Response Lenses

dsersServiceErrors :: Lens' DescribeServiceErrorsResponse [ServiceError']

An array of ServiceError objects that describe the specified service errors.

dsersResponseStatus :: Lens' DescribeServiceErrorsResponse Int

- | The response status code.