Network.AWS.OpsWorks.DescribeTimeBasedAutoScaling

Description

Describes time-based auto scaling configurations for specified instances.

Required Permissions : To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

describeTimeBasedAutoScaling :: DescribeTimeBasedAutoScaling

Creates a value of DescribeTimeBasedAutoScaling with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtbasInstanceIds - An array of instance IDs.

data DescribeTimeBasedAutoScaling

See: describeTimeBasedAutoScaling smart constructor.

Request Lenses

dtbasInstanceIds :: Lens' DescribeTimeBasedAutoScaling [Text]

An array of instance IDs.

Destructuring the Response

describeTimeBasedAutoScalingResponse

Arguments

:: Int
dtbasrsResponseStatus

-> DescribeTimeBasedAutoScalingResponse

Creates a value of DescribeTimeBasedAutoScalingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtbasrsTimeBasedAutoScalingConfigurations - An array of TimeBasedAutoScalingConfiguration objects that describe the configuration for the specified instances.

dtbasrsResponseStatus - -- | The response status code.

data DescribeTimeBasedAutoScalingResponse

Contains the response to a DescribeTimeBasedAutoScaling request.

See: describeTimeBasedAutoScalingResponse smart constructor.

Response Lenses

dtbasrsTimeBasedAutoScalingConfigurations :: Lens' DescribeTimeBasedAutoScalingResponse [TimeBasedAutoScalingConfiguration]

An array of TimeBasedAutoScalingConfiguration objects that describe the configuration for the specified instances.

dtbasrsResponseStatus :: Lens' DescribeTimeBasedAutoScalingResponse Int

- | The response status code.