Network.AWS.OpsWorks.DescribeTimeBasedAutoScaling

Description

Describes time-based auto scaling configurations for specified instances.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_DescribeTimeBasedAutoScaling.html

Request

data DescribeTimeBasedAutoScaling

Request constructor

describeTimeBasedAutoScaling :: DescribeTimeBasedAutoScaling

DescribeTimeBasedAutoScaling constructor.

The fields accessible through corresponding lenses are:

dtbasInstanceIds :: [Text]

Request lenses

dtbasInstanceIds :: Lens' DescribeTimeBasedAutoScaling [Text]

An array of instance IDs.

Response

data DescribeTimeBasedAutoScalingResponse

Response constructor

describeTimeBasedAutoScalingResponse :: DescribeTimeBasedAutoScalingResponse

DescribeTimeBasedAutoScalingResponse constructor.

The fields accessible through corresponding lenses are:

dtbasrTimeBasedAutoScalingConfigurations :: [TimeBasedAutoScalingConfiguration]

Response lenses

dtbasrTimeBasedAutoScalingConfigurations :: Lens' DescribeTimeBasedAutoScalingResponse [TimeBasedAutoScalingConfiguration]

An array of TimeBasedAutoScalingConfiguration objects that describe the configuration for the specified instances.