Network.AWS.OpsWorks.RegisterInstance

Description

Registers instances that were created outside of AWS OpsWorks Stacks with a specified stack.

Registered instances have the same requirements as instances that are created by using the CreateInstance API. For example, registered instances must be running a supported Linux-based operating system, and they must have a supported instance type. For more information about requirements for instances that you want to register, see Preparing the Instance .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

registerInstance

Arguments

:: Text
riStackId

-> RegisterInstance

Creates a value of RegisterInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riPrivateIP - The instance's private IP address.

riHostname - The instance's hostname.

riInstanceIdentity - An InstanceIdentity object that contains the instance's identity.

riPublicIP - The instance's public IP address.

riRsaPublicKeyFingerprint - The instances public RSA key fingerprint.

riRsaPublicKey - The instances public RSA key. This key is used to encrypt communication between the instance and the service.

riStackId - The ID of the stack that the instance is to be registered with.

data RegisterInstance

See: registerInstance smart constructor.

Request Lenses

riPrivateIP :: Lens' RegisterInstance (Maybe Text)

The instance's private IP address.

riHostname :: Lens' RegisterInstance (Maybe Text)

The instance's hostname.

riInstanceIdentity :: Lens' RegisterInstance (Maybe InstanceIdentity)

An InstanceIdentity object that contains the instance's identity.

riPublicIP :: Lens' RegisterInstance (Maybe Text)

The instance's public IP address.

riRsaPublicKeyFingerprint :: Lens' RegisterInstance (Maybe Text)

The instances public RSA key fingerprint.

riRsaPublicKey :: Lens' RegisterInstance (Maybe Text)

The instances public RSA key. This key is used to encrypt communication between the instance and the service.

riStackId :: Lens' RegisterInstance Text

The ID of the stack that the instance is to be registered with.

Destructuring the Response

registerInstanceResponse

Arguments

:: Int
rirsResponseStatus

-> RegisterInstanceResponse

Creates a value of RegisterInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rirsInstanceId - The registered instance's AWS OpsWorks Stacks ID.

rirsResponseStatus - -- | The response status code.

data RegisterInstanceResponse

Contains the response to a RegisterInstanceResult request.

See: registerInstanceResponse smart constructor.

Response Lenses

rirsInstanceId :: Lens' RegisterInstanceResponse (Maybe Text)

The registered instance's AWS OpsWorks Stacks ID.

rirsResponseStatus :: Lens' RegisterInstanceResponse Int

- | The response status code.