Network.AWS.OpsWorks.SetTimeBasedAutoScaling

Description

Specify the time-based auto scaling configuration for a specified instance. For more information, see Managing Load with Time-based and Load-based Instances .

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

setTimeBasedAutoScaling

Arguments

:: Text
stbasInstanceId

-> SetTimeBasedAutoScaling

Creates a value of SetTimeBasedAutoScaling with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stbasAutoScalingSchedule - An AutoScalingSchedule with the instance schedule.

stbasInstanceId - The instance ID.

data SetTimeBasedAutoScaling

See: setTimeBasedAutoScaling smart constructor.

Request Lenses

stbasAutoScalingSchedule :: Lens' SetTimeBasedAutoScaling (Maybe WeeklyAutoScalingSchedule)

An AutoScalingSchedule with the instance schedule.

stbasInstanceId :: Lens' SetTimeBasedAutoScaling Text

The instance ID.

Destructuring the Response

setTimeBasedAutoScalingResponse :: SetTimeBasedAutoScalingResponse

Creates a value of SetTimeBasedAutoScalingResponse with the minimum fields required to make a request.

data SetTimeBasedAutoScalingResponse

See: setTimeBasedAutoScalingResponse smart constructor.