Network.AWS.OpsWorks.StartStack

Description

Starts a stack's instances.

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

startStack

Arguments

:: Text
staStackId

-> StartStack

Creates a value of StartStack with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

staStackId - The stack ID.

data StartStack

See: startStack smart constructor.

Request Lenses

staStackId :: Lens' StartStack Text

The stack ID.

Destructuring the Response

startStackResponse :: StartStackResponse

Creates a value of StartStackResponse with the minimum fields required to make a request.

data StartStackResponse

See: startStackResponse smart constructor.