Network.AWS.OpsWorks.Types

Service

data OpsWorks

Version 2013-02-18 of the Amazon OpsWorks service.

Error

data JSONError :: *

SslConfiguration

data SslConfiguration

sslConfiguration

Arguments

:: Text
scCertificate

-> Text
scPrivateKey

-> SslConfiguration

SslConfiguration constructor.

The fields accessible through corresponding lenses are:

scCertificate :: Text

scChain :: Maybe Text

scPrivateKey :: Text

scCertificate :: Lens' SslConfiguration Text

The contents of the certificate's domain.crt file.

scChain :: Lens' SslConfiguration (Maybe Text)

Optional. Can be used to specify an intermediate certificate authority key or client authentication.

scPrivateKey :: Lens' SslConfiguration Text

The private key; the contents of the certificate's domain.kex file.

VirtualizationType

data VirtualizationType

Constructors

Hvm
hvm

Paravirtual
paravirtual

Command

data Command

command :: Command

Command constructor.

The fields accessible through corresponding lenses are:

cAcknowledgedAt :: Maybe Text

cCommandId :: Maybe Text

cCompletedAt :: Maybe Text

cCreatedAt :: Maybe Text

cDeploymentId :: Maybe Text

cExitCode :: Maybe Int

cInstanceId :: Maybe Text

cLogUrl :: Maybe Text

cStatus :: Maybe Text

cType :: Maybe Text

cAcknowledgedAt :: Lens' Command (Maybe Text)

Date and time when the command was acknowledged.

cCommandId :: Lens' Command (Maybe Text)

The command ID.

cCompletedAt :: Lens' Command (Maybe Text)

Date when the command completed.

cCreatedAt :: Lens' Command (Maybe Text)

Date and time when the command was run.

cDeploymentId :: Lens' Command (Maybe Text)

The command deployment ID.

cExitCode :: Lens' Command (Maybe Int)

The command exit code.

cInstanceId :: Lens' Command (Maybe Text)

The ID of the instance where the command was executed.

cLogUrl :: Lens' Command (Maybe Text)

The URL of the command log.

cStatus :: Lens' Command (Maybe Text)

The command status:

failed successful skipped pending

cType :: Lens' Command (Maybe Text)

The command type:

deploy rollback start stop restart undeploy update_dependencies install_dependencies update_custom_cookbooks execute_recipes

RaidArray

data RaidArray

raidArray :: RaidArray

RaidArray constructor.

The fields accessible through corresponding lenses are:

raAvailabilityZone :: Maybe Text

raCreatedAt :: Maybe Text

raDevice :: Maybe Text

raInstanceId :: Maybe Text

raIops :: Maybe Int

raMountPoint :: Maybe Text

raName :: Maybe Text

raNumberOfDisks :: Maybe Int

raRaidArrayId :: Maybe Text

raRaidLevel :: Maybe Int

raSize :: Maybe Int

raStackId :: Maybe Text

raVolumeType :: Maybe Text

raAvailabilityZone :: Lens' RaidArray (Maybe Text)

The array's Availability Zone. For more information, see Regions and Endpoints .

raCreatedAt :: Lens' RaidArray (Maybe Text)

When the RAID array was created.

raDevice :: Lens' RaidArray (Maybe Text)

The array's Linux device. For example devmdadm0.

raInstanceId :: Lens' RaidArray (Maybe Text)

The instance ID.

raIops :: Lens' RaidArray (Maybe Int)

For PIOPS volumes, the IOPS per disk.

raMountPoint :: Lens' RaidArray (Maybe Text)

The array's mount point.

raName :: Lens' RaidArray (Maybe Text)

The array name.

raNumberOfDisks :: Lens' RaidArray (Maybe Int)

The number of disks in the array.

raRaidArrayId :: Lens' RaidArray (Maybe Text)

The array ID.

raRaidLevel :: Lens' RaidArray (Maybe Int)

The RAID level.

raSize :: Lens' RaidArray (Maybe Int)

The array's size.

raStackId :: Lens' RaidArray (Maybe Text)

The stack ID.

raVolumeType :: Lens' RaidArray (Maybe Text)

The volume type, standard or PIOPS.

ElasticLoadBalancer

data ElasticLoadBalancer

elasticLoadBalancer :: ElasticLoadBalancer

ElasticLoadBalancer constructor.

The fields accessible through corresponding lenses are:

elbAvailabilityZones :: [Text]

elbDnsName :: Maybe Text

elbEc2InstanceIds :: [Text]

elbElasticLoadBalancerName :: Maybe Text

elbLayerId :: Maybe Text

elbRegion :: Maybe Text

elbStackId :: Maybe Text

elbSubnetIds :: [Text]

elbVpcId :: Maybe Text

elbAvailabilityZones :: Lens' ElasticLoadBalancer [Text]

A list of Availability Zones.

elbDnsName :: Lens' ElasticLoadBalancer (Maybe Text)

The instance's public DNS name.

elbEc2InstanceIds :: Lens' ElasticLoadBalancer [Text]

A list of the EC2 instances that the Elastic Load Balancing instance is managing traffic for.

elbElasticLoadBalancerName :: Lens' ElasticLoadBalancer (Maybe Text)

The Elastic Load Balancing instance's name.

elbLayerId :: Lens' ElasticLoadBalancer (Maybe Text)

The ID of the layer that the instance is attached to.

elbRegion :: Lens' ElasticLoadBalancer (Maybe Text)

The instance's AWS region.

elbStackId :: Lens' ElasticLoadBalancer (Maybe Text)

The ID of the stack that the instance is associated with.

elbSubnetIds :: Lens' ElasticLoadBalancer [Text]

A list of subnet IDs, if the stack is running in a VPC.

elbVpcId :: Lens' ElasticLoadBalancer (Maybe Text)

The VPC ID.

LifecycleEventConfiguration

data LifecycleEventConfiguration

lifecycleEventConfiguration :: LifecycleEventConfiguration

LifecycleEventConfiguration constructor.

The fields accessible through corresponding lenses are:

lecShutdown :: Maybe ShutdownEventConfiguration

lecShutdown :: Lens' LifecycleEventConfiguration (Maybe ShutdownEventConfiguration)

A ShutdownEventConfiguration object that specifies the Shutdown event configuration.

RdsDbInstance

data RdsDbInstance

rdsDbInstance :: RdsDbInstance

RdsDbInstance constructor.

The fields accessible through corresponding lenses are:

rdiAddress :: Maybe Text

rdiDbInstanceIdentifier :: Maybe Text

rdiDbPassword :: Maybe Text

rdiDbUser :: Maybe Text

rdiEngine :: Maybe Text

rdiMissingOnRds :: Maybe Bool

rdiRdsDbInstanceArn :: Maybe Text

rdiRegion :: Maybe Text

rdiStackId :: Maybe Text

rdiAddress :: Lens' RdsDbInstance (Maybe Text)

The instance's address.

rdiDbInstanceIdentifier :: Lens' RdsDbInstance (Maybe Text)

The DB instance identifier.

rdiDbPassword :: Lens' RdsDbInstance (Maybe Text)

The database password.

rdiDbUser :: Lens' RdsDbInstance (Maybe Text)

The master user name.

rdiEngine :: Lens' RdsDbInstance (Maybe Text)

The instance's database engine.

rdiMissingOnRds :: Lens' RdsDbInstance (Maybe Bool)

Set to true if AWS OpsWorks was unable to discover the Amazon RDS instance. AWS OpsWorks attempts to discover the instance only once. If this value is set to true, you must deregister the instance and then register it again.

rdiRdsDbInstanceArn :: Lens' RdsDbInstance (Maybe Text)

The instance's ARN.

rdiRegion :: Lens' RdsDbInstance (Maybe Text)

The instance's AWS region.

rdiStackId :: Lens' RdsDbInstance (Maybe Text)

The ID of the stack that the instance is registered with.

AppAttributesKeys

data AppAttributesKeys

Constructors

AutoBundleOnDeploy
AutoBundleOnDeploy

DocumentRoot
DocumentRoot

RailsEnv
RailsEnv

ServiceError'

data ServiceError'

serviceError' :: ServiceError'

ServiceError' constructor.

The fields accessible through corresponding lenses are:

seCreatedAt :: Maybe Text

seInstanceId :: Maybe Text

seMessage :: Maybe Text

seServiceErrorId :: Maybe Text

seStackId :: Maybe Text

seType :: Maybe Text

seCreatedAt :: Lens' ServiceError' (Maybe Text)

When the error occurred.

seInstanceId :: Lens' ServiceError' (Maybe Text)

The instance ID.

seMessage :: Lens' ServiceError' (Maybe Text)

A message that describes the error.

seServiceErrorId :: Lens' ServiceError' (Maybe Text)

The error ID.

seStackId :: Lens' ServiceError' (Maybe Text)

The stack ID.

seType :: Lens' ServiceError' (Maybe Text)

The error type.

StackSummary

data StackSummary

stackSummary :: StackSummary

StackSummary constructor.

The fields accessible through corresponding lenses are:

ssAppsCount :: Maybe Int

ssArn :: Maybe Text

ssInstancesCount :: Maybe InstancesCount

ssLayersCount :: Maybe Int

ssName :: Maybe Text

ssStackId :: Maybe Text

ssAppsCount :: Lens' StackSummary (Maybe Int)

The number of apps.

ssArn :: Lens' StackSummary (Maybe Text)

The stack's ARN.

ssInstancesCount :: Lens' StackSummary (Maybe InstancesCount)

An InstancesCount object with the number of instances in each status.

ssLayersCount :: Lens' StackSummary (Maybe Int)

The number of layers.

ssName :: Lens' StackSummary (Maybe Text)

The stack name.

ssStackId :: Lens' StackSummary (Maybe Text)

The stack ID.

BlockDeviceMapping

data BlockDeviceMapping

blockDeviceMapping :: BlockDeviceMapping

BlockDeviceMapping constructor.

The fields accessible through corresponding lenses are:

bdmDeviceName :: Maybe Text

bdmEbs :: Maybe EbsBlockDevice

bdmNoDevice :: Maybe Text

bdmVirtualName :: Maybe Text

bdmDeviceName :: Lens' BlockDeviceMapping (Maybe Text)

The device name that is exposed to the instance, such as '/dev/sdh'. For the root device, you can use the explicit device name or you can set this parameter to ROOT_DEVICE and AWS OpsWorks will provide the correct device name.

bdmEbs :: Lens' BlockDeviceMapping (Maybe EbsBlockDevice)

An EBSBlockDevice that defines how to configure an Amazon EBS volume when the instance is launched.

bdmNoDevice :: Lens' BlockDeviceMapping (Maybe Text)

Suppresses the specified device included in the AMI's block device mapping.

bdmVirtualName :: Lens' BlockDeviceMapping (Maybe Text)

The virtual device name. For more information, see BlockDeviceMapping.

StackAttributesKeys

data StackAttributesKeys

Constructors

Color
Color

LoadBasedAutoScalingConfiguration

data LoadBasedAutoScalingConfiguration

loadBasedAutoScalingConfiguration :: LoadBasedAutoScalingConfiguration

LoadBasedAutoScalingConfiguration constructor.

The fields accessible through corresponding lenses are:

lbascDownScaling :: Maybe AutoScalingThresholds

lbascEnable :: Maybe Bool

lbascLayerId :: Maybe Text

lbascUpScaling :: Maybe AutoScalingThresholds

lbascDownScaling :: Lens' LoadBasedAutoScalingConfiguration (Maybe AutoScalingThresholds)

An AutoScalingThresholds object that describes the downscaling configuration, which defines how and when AWS OpsWorks reduces the number of instances.

lbascEnable :: Lens' LoadBasedAutoScalingConfiguration (Maybe Bool)

Whether load-based auto scaling is enabled for the layer.

lbascLayerId :: Lens' LoadBasedAutoScalingConfiguration (Maybe Text)

The layer ID.

lbascUpScaling :: Lens' LoadBasedAutoScalingConfiguration (Maybe AutoScalingThresholds)

An AutoScalingThresholds object that describes the upscaling configuration, which defines how and when AWS OpsWorks increases the number of instances.

SourceType

data SourceType

Constructors

Archive
archive

Git
git

S3
s3

Svn
svn

Volume

data Volume

volume :: Volume

Volume constructor.

The fields accessible through corresponding lenses are:

vAvailabilityZone :: Maybe Text

vDevice :: Maybe Text

vEc2VolumeId :: Maybe Text

vInstanceId :: Maybe Text

vIops :: Maybe Int

vMountPoint :: Maybe Text

vName :: Maybe Text

vRaidArrayId :: Maybe Text

vRegion :: Maybe Text

vSize :: Maybe Int

vStatus :: Maybe Text

vVolumeId :: Maybe Text

vVolumeType :: Maybe Text

vAvailabilityZone :: Lens' Volume (Maybe Text)

The volume Availability Zone. For more information, see Regions and Endpoints.

vDevice :: Lens' Volume (Maybe Text)

The device name.

vEc2VolumeId :: Lens' Volume (Maybe Text)

The Amazon EC2 volume ID.

vInstanceId :: Lens' Volume (Maybe Text)

The instance ID.

vIops :: Lens' Volume (Maybe Int)

For PIOPS volumes, the IOPS per disk.

vMountPoint :: Lens' Volume (Maybe Text)

The volume mount point. For example "devsdh".

vName :: Lens' Volume (Maybe Text)

The volume name.

vRaidArrayId :: Lens' Volume (Maybe Text)

The RAID array ID.

vRegion :: Lens' Volume (Maybe Text)

The AWS region. For more information about AWS regions, see Regions andEndpoints.

vSize :: Lens' Volume (Maybe Int)

The volume size.

vStatus :: Lens' Volume (Maybe Text)

The value returned by DescribeVolumes.

vVolumeId :: Lens' Volume (Maybe Text)

The volume ID.

vVolumeType :: Lens' Volume (Maybe Text)

The volume type, standard or PIOPS.

ChefConfiguration

data ChefConfiguration

chefConfiguration :: ChefConfiguration

ChefConfiguration constructor.

The fields accessible through corresponding lenses are:

ccBerkshelfVersion :: Maybe Text

ccManageBerkshelf :: Maybe Bool

ccBerkshelfVersion :: Lens' ChefConfiguration (Maybe Text)

The Berkshelf version.

ccManageBerkshelf :: Lens' ChefConfiguration (Maybe Bool)

Whether to enable Berkshelf.

LayerType

data LayerType

Constructors

Custom
custom

DbMaster
db-master

JavaApp
java-app

Lb
lb

Memcached
memcached

MonitoringMaster
monitoring-master

NodejsApp
nodejs-app

PhpApp
php-app

RailsApp
rails-app

Web
web

AutoScalingThresholds

data AutoScalingThresholds

autoScalingThresholds :: AutoScalingThresholds

AutoScalingThresholds constructor.

The fields accessible through corresponding lenses are:

astCpuThreshold :: Maybe Double

astIgnoreMetricsTime :: Maybe Natural

astInstanceCount :: Maybe Int

astLoadThreshold :: Maybe Double

astMemoryThreshold :: Maybe Double

astThresholdsWaitTime :: Maybe Natural

astCpuThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The CPU utilization threshold, as a percent of the available CPU.

astIgnoreMetricsTime :: Lens' AutoScalingThresholds (Maybe Natural)

The amount of time (in minutes) after a scaling event occurs that AWS OpsWorks should ignore metrics and not raise any additional scaling events. For example, AWS OpsWorks adds new instances following an upscaling event but the instances won't start reducing the load until they have been booted and configured. There is no point in raising additional scaling events during that operation, which typically takes several minutes. IgnoreMetricsTime allows you to direct AWS OpsWorks to not raise any scaling events long enough to get the new instances online.

astInstanceCount :: Lens' AutoScalingThresholds (Maybe Int)

The number of instances to add or remove when the load exceeds a threshold.

astLoadThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The load threshold. For more information about how load is computed, see Load(computing).

astMemoryThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The memory utilization threshold, as a percent of the available memory.

astThresholdsWaitTime :: Lens' AutoScalingThresholds (Maybe Natural)

The amount of time, in minutes, that the load must exceed a threshold before more instances are added or removed.

App

data App

app :: App

App constructor.

The fields accessible through corresponding lenses are:

appAppId :: Maybe Text

appAppSource :: Maybe Source

appAttributes :: HashMap AppAttributesKeys Text

appCreatedAt :: Maybe Text

appDataSources :: [DataSource]

appDescription :: Maybe Text

appDomains :: [Text]

appEnableSsl :: Maybe Bool

appEnvironment :: [EnvironmentVariable]

appName :: Maybe Text

appShortname :: Maybe Text

appSslConfiguration :: Maybe SslConfiguration

appStackId :: Maybe Text

appType :: Maybe AppType

appAppId :: Lens' App (Maybe Text)

The app ID.

appAppSource :: Lens' App (Maybe Source)

A Source object that describes the app repository.

appAttributes :: Lens' App (HashMap AppAttributesKeys Text)

The stack attributes.

appCreatedAt :: Lens' App (Maybe Text)

When the app was created.

appDataSources :: Lens' App [DataSource]

The app's data sources.

appDescription :: Lens' App (Maybe Text)

A description of the app.

appDomains :: Lens' App [Text]

The app vhost settings with multiple domains separated by commas. For example: ''www.example.com, example.com''

appEnableSsl :: Lens' App (Maybe Bool)

Whether to enable SSL for the app.

appEnvironment :: Lens' App [EnvironmentVariable]

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances. For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 10 KB (10240 Bytes). This limit should accommodate most if not all use cases, but if you do exceed it, you will cause an exception (API) with an "Environment: is too large (maximum is 10KB)" message.

appName :: Lens' App (Maybe Text)

The app name.

appShortname :: Lens' App (Maybe Text)

The app's short name.

appSslConfiguration :: Lens' App (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

appStackId :: Lens' App (Maybe Text)

The app stack ID.

appType :: Lens' App (Maybe AppType)

The app type.

ElasticIp

data ElasticIp

elasticIp :: ElasticIp

ElasticIp constructor.

The fields accessible through corresponding lenses are:

eiDomain :: Maybe Text

eiInstanceId :: Maybe Text

eiIp :: Maybe Text

eiName :: Maybe Text

eiRegion :: Maybe Text

eiDomain :: Lens' ElasticIp (Maybe Text)

The domain.

eiInstanceId :: Lens' ElasticIp (Maybe Text)

The ID of the instance that the address is attached to.

eiIp :: Lens' ElasticIp (Maybe Text)

The IP address.

eiName :: Lens' ElasticIp (Maybe Text)

The name.

eiRegion :: Lens' ElasticIp (Maybe Text)

The AWS region. For more information, see Regions and Endpoints.

ShutdownEventConfiguration

data ShutdownEventConfiguration

shutdownEventConfiguration :: ShutdownEventConfiguration

ShutdownEventConfiguration constructor.

The fields accessible through corresponding lenses are:

secDelayUntilElbConnectionsDrained :: Maybe Bool

secExecutionTimeout :: Maybe Int

secDelayUntilElbConnectionsDrained :: Lens' ShutdownEventConfiguration (Maybe Bool)

Whether to enable Elastic Load Balancing connection draining. For more information, see Connection Draining

secExecutionTimeout :: Lens' ShutdownEventConfiguration (Maybe Int)

The time, in seconds, that AWS OpsWorks will wait after triggering a Shutdown event before shutting down an instance.

InstanceIdentity

data InstanceIdentity

instanceIdentity :: InstanceIdentity

InstanceIdentity constructor.

The fields accessible through corresponding lenses are:

iiDocument :: Maybe Text

iiSignature :: Maybe Text

iiDocument :: Lens' InstanceIdentity (Maybe Text)

A JSON document that contains the metadata.

iiSignature :: Lens' InstanceIdentity (Maybe Text)

A signature that can be used to verify the document's accuracy and authenticity.

UserProfile

data UserProfile

userProfile :: UserProfile

UserProfile constructor.

The fields accessible through corresponding lenses are:

upAllowSelfManagement :: Maybe Bool

upIamUserArn :: Maybe Text

upName :: Maybe Text

upSshPublicKey :: Maybe Text

upSshUsername :: Maybe Text

upAllowSelfManagement :: Lens' UserProfile (Maybe Bool)

Whether users can specify their own SSH public key through the My Settings page. For more information, see Managing User Permissions.

upIamUserArn :: Lens' UserProfile (Maybe Text)

The user's IAM ARN.

upName :: Lens' UserProfile (Maybe Text)

The user's name.

upSshPublicKey :: Lens' UserProfile (Maybe Text)

The user's SSH public key.

upSshUsername :: Lens' UserProfile (Maybe Text)

The user's SSH user name.

AutoScalingType

data AutoScalingType

Constructors

Load
load

Timer
timer

Source

data Source

source :: Source

Source constructor.

The fields accessible through corresponding lenses are:

sPassword :: Maybe Text

sRevision :: Maybe Text

sSshKey :: Maybe Text

sType :: Maybe SourceType

sUrl :: Maybe Text

sUsername :: Maybe Text

sPassword :: Lens' Source (Maybe Text)

This parameter depends on the repository type.

For Amazon S3 bundles, set Password to the appropriate IAM secret access key. For HTTP bundles and Subversion repositories, set Password to the password. For more information on how to safely handle IAM credentials, see http://docs.aws.amazon.com/general/latest/gr/aws-access-keys-best-practices.html.

sRevision :: Lens' Source (Maybe Text)

The application's version. AWS OpsWorks enables you to easily deploy new versions of an application. One of the simplest approaches is to have branches or revisions in your repository that represent different versions that can potentially be deployed.

sSshKey :: Lens' Source (Maybe Text)

The repository's SSH key.

sType :: Lens' Source (Maybe SourceType)

The repository type.

sUrl :: Lens' Source (Maybe Text)

The source URL.

sUsername :: Lens' Source (Maybe Text)

This parameter depends on the repository type.

For Amazon S3 bundles, set Username to the appropriate IAM access key ID. For HTTP bundles, Git repositories, and Subversion repositories, set Username to the user name.

DataSource

data DataSource

dataSource :: DataSource

DataSource constructor.

The fields accessible through corresponding lenses are:

dsArn :: Maybe Text

dsDatabaseName :: Maybe Text

dsType :: Maybe Text

dsArn :: Lens' DataSource (Maybe Text)

The data source's ARN.

dsDatabaseName :: Lens' DataSource (Maybe Text)

The database name.

dsType :: Lens' DataSource (Maybe Text)

The data source's type, AutoSelectOpsworksMysqlInstance, OpsworksMysqlInstance , or RdsDbInstance.

Architecture

data Architecture

Constructors

I386
i386

X8664
x86_64

StackConfigurationManager

data StackConfigurationManager

stackConfigurationManager :: StackConfigurationManager

StackConfigurationManager constructor.

The fields accessible through corresponding lenses are:

scmName :: Maybe Text

scmVersion :: Maybe Text

scmName :: Lens' StackConfigurationManager (Maybe Text)

The name. This parameter must be set to Chef.

scmVersion :: Lens' StackConfigurationManager (Maybe Text)

The Chef version. This parameter must be set to 0.9, 11.4, or 11.10. The default value is 11.4.

EbsBlockDevice

data EbsBlockDevice

ebsBlockDevice :: EbsBlockDevice

EbsBlockDevice constructor.

The fields accessible through corresponding lenses are:

ebdDeleteOnTermination :: Maybe Bool

ebdIops :: Maybe Int

ebdSnapshotId :: Maybe Text

ebdVolumeSize :: Maybe Int

ebdVolumeType :: Maybe VolumeType

ebdDeleteOnTermination :: Lens' EbsBlockDevice (Maybe Bool)

Whether the volume is deleted on instance termination.

ebdIops :: Lens' EbsBlockDevice (Maybe Int)

The number of I/O operations per second (IOPS) that the volume supports. For more information, see EbsBlockDevice.

ebdSnapshotId :: Lens' EbsBlockDevice (Maybe Text)

The snapshot ID.

ebdVolumeSize :: Lens' EbsBlockDevice (Maybe Int)

The volume size, in GiB. For more information, see EbsBlockDevice.

ebdVolumeType :: Lens' EbsBlockDevice (Maybe VolumeType)

The volume type. gp2 for General Purpose (SSD) volumes, io1 for Provisioned IOPS (SSD) volumes, and standard for Magnetic volumes.

LayerAttributesKeys

data LayerAttributesKeys

Constructors

BundlerVersion
BundlerVersion

EnableHaproxyStats
EnableHaproxyStats

GangliaPassword
GangliaPassword

GangliaUrl
GangliaUrl

GangliaUser
GangliaUser

HaproxyHealthCheckMethod
HaproxyHealthCheckMethod

HaproxyHealthCheckUrl
HaproxyHealthCheckUrl

HaproxyStatsPassword
HaproxyStatsPassword

HaproxyStatsUrl
HaproxyStatsUrl

HaproxyStatsUser
HaproxyStatsUser

JavaAppServer
JavaAppServer

JavaAppServerVersion
JavaAppServerVersion

Jvm
Jvm

JvmOptions
JvmOptions

JvmVersion
JvmVersion

ManageBundler
ManageBundler

MemcachedMemory
MemcachedMemory

MysqlRootPassword
MysqlRootPassword

MysqlRootPasswordUbiquitous
MysqlRootPasswordUbiquitous

NodejsVersion
NodejsVersion

PassengerVersion
PassengerVersion

RailsStack
RailsStack

RubyVersion
RubyVersion

RubygemsVersion
RubygemsVersion

VolumeConfiguration

data VolumeConfiguration

volumeConfiguration

Arguments

:: Text
vcMountPoint

-> Int
vcNumberOfDisks

-> Int
vcSize

-> VolumeConfiguration

VolumeConfiguration constructor.

The fields accessible through corresponding lenses are:

vcIops :: Maybe Int

vcMountPoint :: Text

vcNumberOfDisks :: Int

vcRaidLevel :: Maybe Int

vcSize :: Int

vcVolumeType :: Maybe Text

vcIops :: Lens' VolumeConfiguration (Maybe Int)

For PIOPS volumes, the IOPS per disk.

vcMountPoint :: Lens' VolumeConfiguration Text

The volume mount point. For example "devsdh".

vcNumberOfDisks :: Lens' VolumeConfiguration Int

The number of disks in the volume.

vcRaidLevel :: Lens' VolumeConfiguration (Maybe Int)

The volume RAID level.

vcSize :: Lens' VolumeConfiguration Int

The volume size.

vcVolumeType :: Lens' VolumeConfiguration (Maybe Text)

The volume type:

standard - Magnetic io1 - Provisioned IOPS (SSD) gp2 - General Purpose (SSD)

VolumeType

data VolumeType

Constructors

Gp2
gp2

Io1
io1

Standard
standard

ReportedOs

data ReportedOs

reportedOs :: ReportedOs

ReportedOs constructor.

The fields accessible through corresponding lenses are:

roFamily :: Maybe Text

roName :: Maybe Text

roVersion :: Maybe Text

roFamily :: Lens' ReportedOs (Maybe Text)

The operating system family.

roName :: Lens' ReportedOs (Maybe Text)

The operating system name.

roVersion :: Lens' ReportedOs (Maybe Text)

The operating system version.

Permission

data Permission

permission :: Permission

Permission constructor.

The fields accessible through corresponding lenses are:

pAllowSsh :: Maybe Bool

pAllowSudo :: Maybe Bool

pIamUserArn :: Maybe Text

pLevel :: Maybe Text

pStackId :: Maybe Text

pAllowSsh :: Lens' Permission (Maybe Bool)

Whether the user can use SSH.

pAllowSudo :: Lens' Permission (Maybe Bool)

Whether the user can use sudo.

pIamUserArn :: Lens' Permission (Maybe Text)

The Amazon Resource Name (ARN) for an AWS Identity and Access Management (IAM) role. For more information about IAM ARNs, see Using Identifiers.

pLevel :: Lens' Permission (Maybe Text)

The user's permission level, which must be the following:

deny show deploy manage iam_only For more information on the permissions associated with these levels, see Managing User Permissions

pStackId :: Lens' Permission (Maybe Text)

A stack ID.

EnvironmentVariable

data EnvironmentVariable

environmentVariable

Arguments

:: Text
evKey

-> Text
evValue

-> EnvironmentVariable

EnvironmentVariable constructor.

The fields accessible through corresponding lenses are:

evKey :: Text

evSecure :: Maybe Bool

evValue :: Text

evKey :: Lens' EnvironmentVariable Text

(Required) The environment variable's name, which can consist of up to 64 characters and must be specified. The name can contain upper- and lowercase letters, numbers, and underscores (_), but it must start with a letter or underscore.

evSecure :: Lens' EnvironmentVariable (Maybe Bool)

(Optional) Whether the variable's value will be returned by the DescribeApps action. To conceal an environment variable's value, set Secure to true. DescribeApps then returns '**Filtered**' instead of the actual value. The default value for Secure is false.

evValue :: Lens' EnvironmentVariable Text

(Optional) The environment variable's value, which can be left empty. If you specify a value, it can contain up to 256 characters, which must all be printable.

Layer

data Layer

layer :: Layer

Layer constructor.

The fields accessible through corresponding lenses are:

lAttributes :: HashMap LayerAttributesKeys Text

lAutoAssignElasticIps :: Maybe Bool

lAutoAssignPublicIps :: Maybe Bool

lCreatedAt :: Maybe Text

lCustomInstanceProfileArn :: Maybe Text

lCustomRecipes :: Maybe Recipes

lCustomSecurityGroupIds :: [Text]

lDefaultRecipes :: Maybe Recipes

lDefaultSecurityGroupNames :: [Text]

lEnableAutoHealing :: Maybe Bool

lInstallUpdatesOnBoot :: Maybe Bool

lLayerId :: Maybe Text

lLifecycleEventConfiguration :: Maybe LifecycleEventConfiguration

lName :: Maybe Text

lPackages :: [Text]

lShortname :: Maybe Text

lStackId :: Maybe Text

lType :: Maybe LayerType

lUseEbsOptimizedInstances :: Maybe Bool

lVolumeConfigurations :: [VolumeConfiguration]

lAttributes :: Lens' Layer (HashMap LayerAttributesKeys Text)

The layer attributes.

lAutoAssignElasticIps :: Lens' Layer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

lAutoAssignPublicIps :: Lens' Layer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How toEdit a Layer.

lCreatedAt :: Lens' Layer (Maybe Text)

Date when the layer was created.

lCustomInstanceProfileArn :: Lens' Layer (Maybe Text)

The ARN of the default IAM profile to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

lCustomRecipes :: Lens' Layer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer's custom recipes.

lCustomSecurityGroupIds :: Lens' Layer [Text]

An array containing the layer's custom security group IDs.

lDefaultRecipes :: Lens' Layer (Maybe Recipes)

lDefaultSecurityGroupNames :: Lens' Layer [Text]

An array containing the layer's security group names.

lEnableAutoHealing :: Lens' Layer (Maybe Bool)

Whether auto healing is disabled for the layer.

lInstallUpdatesOnBoot :: Lens' Layer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or 'apt-get' (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

lLayerId :: Lens' Layer (Maybe Text)

The layer ID.

lLifecycleEventConfiguration :: Lens' Layer (Maybe LifecycleEventConfiguration)

A LifeCycleEventConfiguration object that specifies the Shutdown event configuration.

lName :: Lens' Layer (Maybe Text)

The layer name.

lPackages :: Lens' Layer [Text]

An array of Package objects that describe the layer's packages.

lShortname :: Lens' Layer (Maybe Text)

The layer short name.

lStackId :: Lens' Layer (Maybe Text)

The layer stack ID.

lType :: Lens' Layer (Maybe LayerType)

The layer type.

lUseEbsOptimizedInstances :: Lens' Layer (Maybe Bool)

Whether the layer uses Amazon EBS-optimized instances.

lVolumeConfigurations :: Lens' Layer [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

Recipes

data Recipes

recipes :: Recipes

Recipes constructor.

The fields accessible through corresponding lenses are:

rConfigure :: [Text]

rDeploy :: [Text]

rSetup :: [Text]

rShutdown :: [Text]

rUndeploy :: [Text]

rConfigure :: Lens' Recipes [Text]

An array of custom recipe names to be run following a configure event.

rDeploy :: Lens' Recipes [Text]

An array of custom recipe names to be run following a deploy event.

rSetup :: Lens' Recipes [Text]

An array of custom recipe names to be run following a setup event.

rShutdown :: Lens' Recipes [Text]

An array of custom recipe names to be run following a shutdown event.

rUndeploy :: Lens' Recipes [Text]

An array of custom recipe names to be run following a undeploy event.

TimeBasedAutoScalingConfiguration

data TimeBasedAutoScalingConfiguration

timeBasedAutoScalingConfiguration :: TimeBasedAutoScalingConfiguration

TimeBasedAutoScalingConfiguration constructor.

The fields accessible through corresponding lenses are:

tbascAutoScalingSchedule :: Maybe WeeklyAutoScalingSchedule

tbascInstanceId :: Maybe Text

tbascAutoScalingSchedule :: Lens' TimeBasedAutoScalingConfiguration (Maybe WeeklyAutoScalingSchedule)

A WeeklyAutoScalingSchedule object with the instance schedule.

tbascInstanceId :: Lens' TimeBasedAutoScalingConfiguration (Maybe Text)

The instance ID.

SelfUserProfile

data SelfUserProfile

selfUserProfile :: SelfUserProfile

SelfUserProfile constructor.

The fields accessible through corresponding lenses are:

supIamUserArn :: Maybe Text

supName :: Maybe Text

supSshPublicKey :: Maybe Text

supSshUsername :: Maybe Text

supIamUserArn :: Lens' SelfUserProfile (Maybe Text)

The user's IAM ARN.

supName :: Lens' SelfUserProfile (Maybe Text)

The user's name.

supSshPublicKey :: Lens' SelfUserProfile (Maybe Text)

The user's SSH public key.

supSshUsername :: Lens' SelfUserProfile (Maybe Text)

The user's SSH user name.

RootDeviceType

data RootDeviceType

Constructors

Ebs
ebs

InstanceStore
instance-store

Stack

data Stack

stack :: Stack

Stack constructor.

The fields accessible through corresponding lenses are:

sArn :: Maybe Text

sAttributes :: HashMap StackAttributesKeys Text

sChefConfiguration :: Maybe ChefConfiguration

sConfigurationManager :: Maybe StackConfigurationManager

sCreatedAt :: Maybe Text

sCustomCookbooksSource :: Maybe Source

sCustomJson :: Maybe Text

sDefaultAvailabilityZone :: Maybe Text

sDefaultInstanceProfileArn :: Maybe Text

sDefaultOs :: Maybe Text

sDefaultRootDeviceType :: Maybe RootDeviceType

sDefaultSshKeyName :: Maybe Text

sDefaultSubnetId :: Maybe Text

sHostnameTheme :: Maybe Text

sName :: Maybe Text

sRegion :: Maybe Text

sServiceRoleArn :: Maybe Text

sStackId :: Maybe Text

sUseCustomCookbooks :: Maybe Bool

sUseOpsworksSecurityGroups :: Maybe Bool

sVpcId :: Maybe Text

sArn :: Lens' Stack (Maybe Text)

The stack's ARN.

sAttributes :: Lens' Stack (HashMap StackAttributesKeys Text)

The stack's attributes.

sChefConfiguration :: Lens' Stack (Maybe ChefConfiguration)

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version. For more information, see Create a New Stack.

sConfigurationManager :: Lens' Stack (Maybe StackConfigurationManager)

The configuration manager.

sCreatedAt :: Lens' Stack (Maybe Text)

Date when the stack was created.

sCustomCookbooksSource :: Lens' Stack (Maybe Source)

sCustomJson :: Lens' Stack (Maybe Text)

A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format and must escape characters such as '"'.:

'"{"key1": "value1", "key2": "value2",...}"'

For more information on custom JSON, see Use Custom JSON to Modify the StackConfiguration JSON.

sDefaultAvailabilityZone :: Lens' Stack (Maybe Text)

The stack's default Availability Zone. For more information, see Regions andEndpoints.

sDefaultInstanceProfileArn :: Lens' Stack (Maybe Text)

The ARN of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

sDefaultOs :: Lens' Stack (Maybe Text)

The stack's default operating system.

sDefaultRootDeviceType :: Lens' Stack (Maybe RootDeviceType)

The default root device type. This value is used by default for all instances in the stack, but you can override it when you create an instance. For more information, see Storage for the Root Device.

sDefaultSshKeyName :: Lens' Stack (Maybe Text)

A default Amazon EC2 key pair for the stack's instances. You can override this value when you create or update an instance.

sDefaultSubnetId :: Lens' Stack (Maybe Text)

The default subnet ID, if the stack is running in a VPC.

sHostnameTheme :: Lens' Stack (Maybe Text)

The stack host name theme, with spaces replaced by underscores.

sName :: Lens' Stack (Maybe Text)

The stack name.

sRegion :: Lens' Stack (Maybe Text)

The stack AWS region, such as "us-east-1". For more information about AWS regions, see Regions and Endpoints.

sServiceRoleArn :: Lens' Stack (Maybe Text)

The stack AWS Identity and Access Management (IAM) role.

sStackId :: Lens' Stack (Maybe Text)

The stack ID.

sUseCustomCookbooks :: Lens' Stack (Maybe Bool)

Whether the stack uses custom cookbooks.

sUseOpsworksSecurityGroups :: Lens' Stack (Maybe Bool)

Whether the stack automatically associates the AWS OpsWorks built-in security groups with the stack's layers.

sVpcId :: Lens' Stack (Maybe Text)

The VPC ID, if the stack is running in a VPC.

DeploymentCommand

data DeploymentCommand

deploymentCommand

Arguments

:: DeploymentCommandName
dcName

-> DeploymentCommand

DeploymentCommand constructor.

The fields accessible through corresponding lenses are:

dcArgs :: HashMap Text [Text]

dcName :: DeploymentCommandName

dcArgs :: Lens' DeploymentCommand (HashMap Text [Text])

The arguments of those commands that take arguments. It should be set to a JSON object with the following format:

'{"arg_name1" : ["value1", "value2", ...], "arg_name2" : ["value1", "value2",...], ...}'

The update_dependencies command takes two arguments:

upgrade_os_to - Specifies the desired Amazon Linux version for instances whose OS you want to upgrade, such as 'Amazon Linux 2014.09'. You must also set the allow_reboot argument to true. allow_reboot - Specifies whether to allow AWS OpsWorks to reboot the instances if necessary, after installing the updates. This argument can be set to either true or false. The default value is false. For example, to upgrade an instance to Amazon Linux 2014.09, set Args to the following.

' { "upgrade_os_to":["Amazon Linux 2014.09"], "allow_reboot":["true"] } '

dcName :: Lens' DeploymentCommand DeploymentCommandName

Specifies the operation. You can specify only one command.

For stacks, the following commands are available:

execute_recipes: Execute one or more recipes. To specify the recipes, set an Args parameter named recipes to the list of recipes to be executed. For example, to execute 'phpapp::appsetup', set Args to '{"recipes":["phpapp::appsetup"]}'. install_dependencies: Install the stack's dependencies. update_custom_cookbooks: Update the stack's custom cookbooks. update_dependencies: Update the stack's dependencies. For apps, the following commands are available:

deploy: Deploy an app. Rails apps have an optional Args parameter named migrate. Set Args to {"migrate":["true"]} to migrate the database. The default setting is {"migrate":["false"]}. rollback Roll the app back to the previous version. When you update an app, AWS OpsWorks stores the previous version, up to a maximum of five versions. You can use this command to roll an app back as many as four versions. start: Start the app's web or application server. stop: Stop the app's web or application server. restart: Restart the app's web or application server. undeploy: Undeploy the app.

WeeklyAutoScalingSchedule

data WeeklyAutoScalingSchedule

weeklyAutoScalingSchedule :: WeeklyAutoScalingSchedule

WeeklyAutoScalingSchedule constructor.

The fields accessible through corresponding lenses are:

wassFriday :: HashMap Text Text

wassMonday :: HashMap Text Text

wassSaturday :: HashMap Text Text

wassSunday :: HashMap Text Text

wassThursday :: HashMap Text Text

wassTuesday :: HashMap Text Text

wassWednesday :: HashMap Text Text

wassFriday :: Lens' WeeklyAutoScalingSchedule (HashMap Text Text)

The schedule for Friday.

wassMonday :: Lens' WeeklyAutoScalingSchedule (HashMap Text Text)

The schedule for Monday.

wassSaturday :: Lens' WeeklyAutoScalingSchedule (HashMap Text Text)

The schedule for Saturday.

wassSunday :: Lens' WeeklyAutoScalingSchedule (HashMap Text Text)

The schedule for Sunday.

wassThursday :: Lens' WeeklyAutoScalingSchedule (HashMap Text Text)

The schedule for Thursday.

wassTuesday :: Lens' WeeklyAutoScalingSchedule (HashMap Text Text)

The schedule for Tuesday.

wassWednesday :: Lens' WeeklyAutoScalingSchedule (HashMap Text Text)

The schedule for Wednesday.

DeploymentCommandName

data DeploymentCommandName

Constructors

Deploy
deploy

ExecuteRecipes
execute_recipes

InstallDependencies
install_dependencies

Restart
restart

Rollback
rollback

Start
start

Stop
stop

Undeploy
undeploy

UpdateCustomCookbooks
update_custom_cookbooks

UpdateDependencies
update_dependencies

Instance

data Instance

instance' :: Instance

Instance constructor.

The fields accessible through corresponding lenses are:

iAmiId :: Maybe Text

iArchitecture :: Maybe Architecture

iAutoScalingType :: Maybe AutoScalingType

iAvailabilityZone :: Maybe Text

iBlockDeviceMappings :: [BlockDeviceMapping]

iCreatedAt :: Maybe Text

iEbsOptimized :: Maybe Bool

iEc2InstanceId :: Maybe Text

iElasticIp :: Maybe Text

iHostname :: Maybe Text

iInfrastructureClass :: Maybe Text

iInstallUpdatesOnBoot :: Maybe Bool

iInstanceId :: Maybe Text

iInstanceProfileArn :: Maybe Text

iInstanceType :: Maybe Text

iLastServiceErrorId :: Maybe Text

iLayerIds :: [Text]

iOs :: Maybe Text

iPrivateDns :: Maybe Text

iPrivateIp :: Maybe Text

iPublicDns :: Maybe Text

iPublicIp :: Maybe Text

iRegisteredBy :: Maybe Text

iReportedAgentVersion :: Maybe Text

iReportedOs :: Maybe ReportedOs

iRootDeviceType :: Maybe RootDeviceType

iRootDeviceVolumeId :: Maybe Text

iSecurityGroupIds :: [Text]

iSshHostDsaKeyFingerprint :: Maybe Text

iSshHostRsaKeyFingerprint :: Maybe Text

iSshKeyName :: Maybe Text

iStackId :: Maybe Text

iStatus :: Maybe Text

iSubnetId :: Maybe Text

iVirtualizationType :: Maybe VirtualizationType

iAmiId :: Lens' Instance (Maybe Text)

A custom AMI ID to be used to create the instance. The AMI should be based on one of the standard AWS OpsWorks APIs: Amazon Linux, Ubuntu 12.04 LTS, or Ubuntu 14.04 LTS. For more information, see Instances

iArchitecture :: Lens' Instance (Maybe Architecture)

The instance architecture, "i386" or "x86_64".

iAutoScalingType :: Lens' Instance (Maybe AutoScalingType)

For load-based or time-based instances, the type.

iAvailabilityZone :: Lens' Instance (Maybe Text)

The instance Availability Zone. For more information, see Regions andEndpoints.

iBlockDeviceMappings :: Lens' Instance [BlockDeviceMapping]

An array of BlockDeviceMapping objects that specify the instance's block device mappings.

iCreatedAt :: Lens' Instance (Maybe Text)

The time that the instance was created.

iEbsOptimized :: Lens' Instance (Maybe Bool)

Whether this is an Amazon EBS-optimized instance.

iEc2InstanceId :: Lens' Instance (Maybe Text)

The ID of the associated Amazon EC2 instance.

iElasticIp :: Lens' Instance (Maybe Text)

The instance Elastic IP address.

iHostname :: Lens' Instance (Maybe Text)

The instance host name.

iInfrastructureClass :: Lens' Instance (Maybe Text)

For registered instances, the infrastructure class: ec2 or 'on-premises'

iInstallUpdatesOnBoot :: Lens' Instance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or 'apt-get' (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

iInstanceId :: Lens' Instance (Maybe Text)

The instance ID.

iInstanceProfileArn :: Lens' Instance (Maybe Text)

The ARN of the instance's IAM profile. For more information about IAM ARNs, see Using Identifiers.

iInstanceType :: Lens' Instance (Maybe Text)

The instance type. AWS OpsWorks supports all instance types except Cluster Compute, Cluster GPU, and High Memory Cluster. For more information, see Instance Families and Types. The parameter values that specify the various types are in the API Name column of the Available Instance Types table.

iLastServiceErrorId :: Lens' Instance (Maybe Text)

The ID of the last service error. For more information, call DescribeServiceErrors.

iLayerIds :: Lens' Instance [Text]

An array containing the instance layer IDs.

iOs :: Lens' Instance (Maybe Text)

The instance's operating system.

iPrivateDns :: Lens' Instance (Maybe Text)

The instance private DNS name.

iPrivateIp :: Lens' Instance (Maybe Text)

The instance private IP address.

iPublicDns :: Lens' Instance (Maybe Text)

The instance public DNS name.

iPublicIp :: Lens' Instance (Maybe Text)

The instance public IP address.

iRegisteredBy :: Lens' Instance (Maybe Text)

For registered instances, who performed the registration.

iReportedAgentVersion :: Lens' Instance (Maybe Text)

The instance's reported AWS OpsWorks agent version.

iReportedOs :: Lens' Instance (Maybe ReportedOs)

For registered instances, the reported operating system.

iRootDeviceType :: Lens' Instance (Maybe RootDeviceType)

The instance's root device type. For more information, see Storage for theRoot Device.

iRootDeviceVolumeId :: Lens' Instance (Maybe Text)

The root device volume ID.

iSecurityGroupIds :: Lens' Instance [Text]

An array containing the instance security group IDs.

iSshHostDsaKeyFingerprint :: Lens' Instance (Maybe Text)

The SSH key's DSA fingerprint.

iSshHostRsaKeyFingerprint :: Lens' Instance (Maybe Text)

The SSH key's RSA fingerprint.

iSshKeyName :: Lens' Instance (Maybe Text)

The instance's Amazon EC2 key pair name.

iStackId :: Lens' Instance (Maybe Text)

The stack ID.

iStatus :: Lens' Instance (Maybe Text)

The instance status:

booting connection_lost online pending rebooting requested running_setup setup_failed shutting_down start_failed stopped stopping terminated terminating

iSubnetId :: Lens' Instance (Maybe Text)

The instance's subnet ID, if the stack is running in a VPC.

iVirtualizationType :: Lens' Instance (Maybe VirtualizationType)

The instance's virtualization type, paravirtual or hvm.

Deployment

data Deployment

deployment :: Deployment

Deployment constructor.

The fields accessible through corresponding lenses are:

dAppId :: Maybe Text

dCommand :: Maybe DeploymentCommand

dComment :: Maybe Text

dCompletedAt :: Maybe Text

dCreatedAt :: Maybe Text

dCustomJson :: Maybe Text

dDeploymentId :: Maybe Text

dDuration :: Maybe Int

dIamUserArn :: Maybe Text

dInstanceIds :: [Text]

dStackId :: Maybe Text

dStatus :: Maybe Text

dAppId :: Lens' Deployment (Maybe Text)

The app ID.

dCommand :: Lens' Deployment (Maybe DeploymentCommand)

dComment :: Lens' Deployment (Maybe Text)

A user-defined comment.

dCompletedAt :: Lens' Deployment (Maybe Text)

Date when the deployment completed.

dCreatedAt :: Lens' Deployment (Maybe Text)

Date when the deployment was created.

dCustomJson :: Lens' Deployment (Maybe Text)

A string that contains user-defined custom JSON. It is used to override the corresponding default stack configuration JSON values for stack. The string should be in the following format and must escape characters such as '"'.:

'"{"key1": "value1", "key2": "value2",...}"'

For more information on custom JSON, see Use Custom JSON to Modify the StackConfiguration JSON.

dDeploymentId :: Lens' Deployment (Maybe Text)

The deployment ID.

dDuration :: Lens' Deployment (Maybe Int)

The deployment duration.

dIamUserArn :: Lens' Deployment (Maybe Text)

The user's IAM ARN.

dInstanceIds :: Lens' Deployment [Text]

The IDs of the target instances.

dStackId :: Lens' Deployment (Maybe Text)

The stack ID.

dStatus :: Lens' Deployment (Maybe Text)

The deployment status:

running successful failed

InstancesCount

data InstancesCount

instancesCount :: InstancesCount

InstancesCount constructor.

The fields accessible through corresponding lenses are:

icAssigning :: Maybe Int

icBooting :: Maybe Int

icConnectionLost :: Maybe Int

icDeregistering :: Maybe Int

icOnline :: Maybe Int

icPending :: Maybe Int

icRebooting :: Maybe Int

icRegistered :: Maybe Int

icRegistering :: Maybe Int

icRequested :: Maybe Int

icRunningSetup :: Maybe Int

icSetupFailed :: Maybe Int

icShuttingDown :: Maybe Int

icStartFailed :: Maybe Int

icStopped :: Maybe Int

icStopping :: Maybe Int

icTerminated :: Maybe Int

icTerminating :: Maybe Int

icUnassigning :: Maybe Int

icAssigning :: Lens' InstancesCount (Maybe Int)

The number of instances in the Assigning state.

icBooting :: Lens' InstancesCount (Maybe Int)

The number of instances with booting status.

icConnectionLost :: Lens' InstancesCount (Maybe Int)

The number of instances with connection_lost status.

icDeregistering :: Lens' InstancesCount (Maybe Int)

The number of instances in the Deregistering state.

icOnline :: Lens' InstancesCount (Maybe Int)

The number of instances with online status.

icPending :: Lens' InstancesCount (Maybe Int)

The number of instances with pending status.

icRebooting :: Lens' InstancesCount (Maybe Int)

The number of instances with rebooting status.

icRegistered :: Lens' InstancesCount (Maybe Int)

The number of instances in the Registered state.

icRegistering :: Lens' InstancesCount (Maybe Int)

The number of instances in the Registering state.

icRequested :: Lens' InstancesCount (Maybe Int)

The number of instances with requested status.

icRunningSetup :: Lens' InstancesCount (Maybe Int)

The number of instances with running_setup status.

icSetupFailed :: Lens' InstancesCount (Maybe Int)

The number of instances with setup_failed status.

icShuttingDown :: Lens' InstancesCount (Maybe Int)

The number of instances with shutting_down status.

icStartFailed :: Lens' InstancesCount (Maybe Int)

The number of instances with start_failed status.

icStopped :: Lens' InstancesCount (Maybe Int)

The number of instances with stopped status.

icStopping :: Lens' InstancesCount (Maybe Int)

The number of instances with stopping status.

icTerminated :: Lens' InstancesCount (Maybe Int)

The number of instances with terminated status.

icTerminating :: Lens' InstancesCount (Maybe Int)

The number of instances with terminating status.

icUnassigning :: Lens' InstancesCount (Maybe Int)

The number of instances in the Unassigning state.

AppType

data AppType

Constructors

Java
java

Nodejs
nodejs

Other
other

Php
php

Rails
rails

Static
static