Network.AWS.OpsWorks.UpdateApp

Description

Updates a specified app.

Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see ManagingUser Permissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_UpdateApp.html

Request

data UpdateApp

Request constructor

updateApp

Arguments

:: Text
uaAppId

-> UpdateApp

UpdateApp constructor.

The fields accessible through corresponding lenses are:

uaAppId :: Text

uaAppSource :: Maybe Source

uaAttributes :: HashMap AppAttributesKeys Text

uaDataSources :: [DataSource]

uaDescription :: Maybe Text

uaDomains :: [Text]

uaEnableSsl :: Maybe Bool

uaEnvironment :: [EnvironmentVariable]

uaName :: Maybe Text

uaSslConfiguration :: Maybe SslConfiguration

uaType :: Maybe AppType

Request lenses

uaAppId :: Lens' UpdateApp Text

The app ID.

uaAppSource :: Lens' UpdateApp (Maybe Source)

A Source object that specifies the app repository.

uaAttributes :: Lens' UpdateApp (HashMap AppAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

uaDataSources :: Lens' UpdateApp [DataSource]

The app's data sources.

uaDescription :: Lens' UpdateApp (Maybe Text)

A description of the app.

uaDomains :: Lens' UpdateApp [Text]

The app's virtual host settings, with multiple domains separated by commas. For example: ''www.example.com, example.com''

uaEnableSsl :: Lens' UpdateApp (Maybe Bool)

Whether SSL is enabled for the app.

uaEnvironment :: Lens' UpdateApp [EnvironmentVariable]

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances.For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 10 KB (10240 Bytes). This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 10KB)."

This parameter is supported only by Chef 11.10 stacks. If you have specified one or more environment variables, you cannot modify the stack's Chef version.

uaName :: Lens' UpdateApp (Maybe Text)

The app name.

uaSslConfiguration :: Lens' UpdateApp (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

uaType :: Lens' UpdateApp (Maybe AppType)

The app type.

Response

data UpdateAppResponse

Response constructor

updateAppResponse :: UpdateAppResponse

UpdateAppResponse constructor.