Network.AWS.OpsWorks.UpdateInstance

Description

Updates a specified instance.

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

updateInstance

Arguments

:: Text
uiInstanceId

-> UpdateInstance

Creates a value of UpdateInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uiInstallUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true . To control when updates are installed, set this value to false . You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

uiHostname - The instance host name.

uiSSHKeyName - The instance's Amazon EC2 key name.

uiAgentVersion - The default AWS OpsWorks Stacks agent version. You have the following options: * INHERIT - Use the stack's default agent version setting. * version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, you must edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance. The default setting is INHERIT . To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions . AgentVersion cannot be set to Chef 12.2.

uiInstanceType - The instance type, such as t2.micro . For a list of supported instance types, open the stack in the console, choose Instances , and choose + Instance . The Size list contains the currently supported types. For more information, see Instance Families and Types . The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

uiEBSOptimized - This property cannot be updated.

uiOS - The instance's operating system, which must be set to one of the following. You cannot update an instance that is using a custom AMI. * A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2017.03 , Amazon Linux 2016.09 , Amazon Linux 2016.03 , Amazon Linux 2015.09 , or Amazon Linux 2015.03 . * A supported Ubuntu operating system, such as Ubuntu 16.04 LTS , Ubuntu 14.04 LTS , or Ubuntu 12.04 LTS . * CentOS Linux 7 * Red Hat Enterprise Linux 7 * A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base , Microsoft Windows Server 2012 R2 with SQL Server Express , Microsoft Windows Server 2012 R2 with SQL Server Standard , or Microsoft Windows Server 2012 R2 with SQL Server Web . For more information on the supported operating systems, see AWS OpsWorks Stacks Operating Systems . The default option is the current Amazon Linux version. If you set this parameter to Custom , you must use the AmiId parameter to specify the custom AMI that you want to use. For more information on the supported operating systems, see Operating Systems . For more information on how to use custom AMIs with OpsWorks, see Using Custom AMIs .

uiAutoScalingType - For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

uiLayerIds - The instance's layer IDs.

uiArchitecture - The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types .

uiAMIId - The ID of the AMI that was used to create the instance. The value of this parameter must be the same AMI ID that the instance is already using. You cannot apply a new AMI to an instance by running UpdateInstance. UpdateInstance does not work on instances that are using custom AMIs.

uiInstanceId - The instance ID.

data UpdateInstance

See: updateInstance smart constructor.

Request Lenses

uiInstallUpdatesOnBoot :: Lens' UpdateInstance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true . To control when updates are installed, set this value to false . You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

uiHostname :: Lens' UpdateInstance (Maybe Text)

The instance host name.

uiSSHKeyName :: Lens' UpdateInstance (Maybe Text)

The instance's Amazon EC2 key name.

uiAgentVersion :: Lens' UpdateInstance (Maybe Text)

The default AWS OpsWorks Stacks agent version. You have the following options: * INHERIT - Use the stack's default agent version setting. * version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, you must edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance. The default setting is INHERIT . To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions . AgentVersion cannot be set to Chef 12.2.

uiInstanceType :: Lens' UpdateInstance (Maybe Text)

The instance type, such as t2.micro . For a list of supported instance types, open the stack in the console, choose Instances , and choose + Instance . The Size list contains the currently supported types. For more information, see Instance Families and Types . The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

uiEBSOptimized :: Lens' UpdateInstance (Maybe Bool)

This property cannot be updated.

uiOS :: Lens' UpdateInstance (Maybe Text)

The instance's operating system, which must be set to one of the following. You cannot update an instance that is using a custom AMI. * A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2017.03 , Amazon Linux 2016.09 , Amazon Linux 2016.03 , Amazon Linux 2015.09 , or Amazon Linux 2015.03 . * A supported Ubuntu operating system, such as Ubuntu 16.04 LTS , Ubuntu 14.04 LTS , or Ubuntu 12.04 LTS . * CentOS Linux 7 * Red Hat Enterprise Linux 7 * A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base , Microsoft Windows Server 2012 R2 with SQL Server Express , Microsoft Windows Server 2012 R2 with SQL Server Standard , or Microsoft Windows Server 2012 R2 with SQL Server Web . For more information on the supported operating systems, see AWS OpsWorks Stacks Operating Systems . The default option is the current Amazon Linux version. If you set this parameter to Custom , you must use the AmiId parameter to specify the custom AMI that you want to use. For more information on the supported operating systems, see Operating Systems . For more information on how to use custom AMIs with OpsWorks, see Using Custom AMIs .

uiAutoScalingType :: Lens' UpdateInstance (Maybe AutoScalingType)

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

uiLayerIds :: Lens' UpdateInstance [Text]

The instance's layer IDs.

uiArchitecture :: Lens' UpdateInstance (Maybe Architecture)

The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types .

uiAMIId :: Lens' UpdateInstance (Maybe Text)

The ID of the AMI that was used to create the instance. The value of this parameter must be the same AMI ID that the instance is already using. You cannot apply a new AMI to an instance by running UpdateInstance. UpdateInstance does not work on instances that are using custom AMIs.

uiInstanceId :: Lens' UpdateInstance Text

The instance ID.

Destructuring the Response

updateInstanceResponse :: UpdateInstanceResponse

Creates a value of UpdateInstanceResponse with the minimum fields required to make a request.

data UpdateInstanceResponse

See: updateInstanceResponse smart constructor.