Network.AWS.OpsWorks.UpdateLayer

Description

Updates a specified layer.

Required Permissions : To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions .

Creating a Request

updateLayer

Arguments

:: Text
ulLayerId

-> UpdateLayer

Creates a value of UpdateLayer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ulCustomInstanceProfileARN - The ARN of an IAM profile to be used for all of the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers .

ulCustomSecurityGroupIds - An array containing the layer's custom security group IDs.

ulInstallUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true . To control when updates are installed, set this value to false . You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

ulCloudWatchLogsConfiguration - Specifies CloudWatch Logs configuration options for the layer. For more information, see CloudWatchLogsLogStream .

ulLifecycleEventConfiguration -

ulShortname - For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorks Stacks and by Chef. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters and must be in the following format: A[a-z0-9-_.]+Z. The built-in layers' short names are defined by AWS OpsWorks Stacks. For more information, see the Layer Reference

ulCustomRecipes - A LayerCustomRecipes object that specifies the layer's custom recipes.

ulCustomJSON - A JSON-formatted string containing custom stack configuration and deployment attributes to be installed on the layer's instances. For more information, see Using Custom JSON .

ulVolumeConfigurations - A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

ulEnableAutoHealing - Whether to disable auto healing for the layer.

ulPackages - An array of Package objects that describe the layer's packages.

ulAttributes - One or more user-defined key/value pairs to be added to the stack attributes.

ulName - The layer name, which is used by the console.

ulAutoAssignPublicIPs - For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer .

ulUseEBSOptimizedInstances - Whether to use Amazon EBS-optimized instances.

ulAutoAssignElasticIPs - Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer .

ulLayerId - The layer ID.

data UpdateLayer

See: updateLayer smart constructor.

Request Lenses

ulCustomInstanceProfileARN :: Lens' UpdateLayer (Maybe Text)

The ARN of an IAM profile to be used for all of the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers .

ulCustomSecurityGroupIds :: Lens' UpdateLayer [Text]

An array containing the layer's custom security group IDs.

ulInstallUpdatesOnBoot :: Lens' UpdateLayer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true . To control when updates are installed, set this value to false . You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

ulCloudWatchLogsConfiguration :: Lens' UpdateLayer (Maybe CloudWatchLogsConfiguration)

Specifies CloudWatch Logs configuration options for the layer. For more information, see CloudWatchLogsLogStream .

ulLifecycleEventConfiguration :: Lens' UpdateLayer (Maybe LifecycleEventConfiguration)

ulShortname :: Lens' UpdateLayer (Maybe Text)

For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorks Stacks and by Chef. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters and must be in the following format: A[a-z0-9-_.]+Z. The built-in layers' short names are defined by AWS OpsWorks Stacks. For more information, see the Layer Reference

ulCustomRecipes :: Lens' UpdateLayer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer's custom recipes.

ulCustomJSON :: Lens' UpdateLayer (Maybe Text)

A JSON-formatted string containing custom stack configuration and deployment attributes to be installed on the layer's instances. For more information, see Using Custom JSON .

ulVolumeConfigurations :: Lens' UpdateLayer [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

ulEnableAutoHealing :: Lens' UpdateLayer (Maybe Bool)

Whether to disable auto healing for the layer.

ulPackages :: Lens' UpdateLayer [Text]

An array of Package objects that describe the layer's packages.

ulAttributes :: Lens' UpdateLayer (HashMap LayerAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

ulName :: Lens' UpdateLayer (Maybe Text)

The layer name, which is used by the console.

ulAutoAssignPublicIPs :: Lens' UpdateLayer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer .

ulUseEBSOptimizedInstances :: Lens' UpdateLayer (Maybe Bool)

Whether to use Amazon EBS-optimized instances.

ulAutoAssignElasticIPs :: Lens' UpdateLayer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer .

ulLayerId :: Lens' UpdateLayer Text

The layer ID.

Destructuring the Response

updateLayerResponse :: UpdateLayerResponse

Creates a value of UpdateLayerResponse with the minimum fields required to make a request.

data UpdateLayerResponse

See: updateLayerResponse smart constructor.