Network.AWS.OpsWorks.UpdateLayer

Description

Updates a specified layer.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_UpdateLayer.html

Request

data UpdateLayer

Request constructor

updateLayer

Arguments

:: Text
ulLayerId

-> UpdateLayer

UpdateLayer constructor.

The fields accessible through corresponding lenses are:

ulAttributes :: HashMap LayerAttributesKeys Text

ulAutoAssignElasticIps :: Maybe Bool

ulAutoAssignPublicIps :: Maybe Bool

ulCustomInstanceProfileArn :: Maybe Text

ulCustomRecipes :: Maybe Recipes

ulCustomSecurityGroupIds :: [Text]

ulEnableAutoHealing :: Maybe Bool

ulInstallUpdatesOnBoot :: Maybe Bool

ulLayerId :: Text

ulLifecycleEventConfiguration :: Maybe LifecycleEventConfiguration

ulName :: Maybe Text

ulPackages :: [Text]

ulShortname :: Maybe Text

ulUseEbsOptimizedInstances :: Maybe Bool

ulVolumeConfigurations :: [VolumeConfiguration]

Request lenses

ulAttributes :: Lens' UpdateLayer (HashMap LayerAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

ulAutoAssignElasticIps :: Lens' UpdateLayer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

ulAutoAssignPublicIps :: Lens' UpdateLayer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How toEdit a Layer.

ulCustomInstanceProfileArn :: Lens' UpdateLayer (Maybe Text)

The ARN of an IAM profile to be used for all of the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

ulCustomRecipes :: Lens' UpdateLayer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer's custom recipes.

ulCustomSecurityGroupIds :: Lens' UpdateLayer [Text]

An array containing the layer's custom security group IDs.

ulEnableAutoHealing :: Lens' UpdateLayer (Maybe Bool)

Whether to disable auto healing for the layer.

ulInstallUpdatesOnBoot :: Lens' UpdateLayer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or 'apt-get' (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

ulLayerId :: Lens' UpdateLayer Text

The layer ID.

ulLifecycleEventConfiguration :: Lens' UpdateLayer (Maybe LifecycleEventConfiguration)

ulName :: Lens' UpdateLayer (Maybe Text)

The layer name, which is used by the console.

ulPackages :: Lens' UpdateLayer [Text]

An array of Package objects that describe the layer's packages.

ulShortname :: Lens' UpdateLayer (Maybe Text)

For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorksand by Chef. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters and must be in the following format: A[a-z0-9-_.]+Z.

The built-in layers' short names are defined by AWS OpsWorks. For more information, see the Layer Reference

ulUseEbsOptimizedInstances :: Lens' UpdateLayer (Maybe Bool)

Whether to use Amazon EBS-optimized instances.

ulVolumeConfigurations :: Lens' UpdateLayer [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

Response

data UpdateLayerResponse

Response constructor

updateLayerResponse :: UpdateLayerResponse

UpdateLayerResponse constructor.