Network.AWS.OpsWorks.UpdateStack

Description

Updates a specified stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing UserPermissions.

http://docs.aws.amazon.com/opsworks/latest/APIReference/API_UpdateStack.html

Request

data UpdateStack

Request constructor

updateStack

Arguments

:: Text
usStackId

-> UpdateStack

UpdateStack constructor.

The fields accessible through corresponding lenses are:

usAttributes :: HashMap StackAttributesKeys Text

usChefConfiguration :: Maybe ChefConfiguration

usConfigurationManager :: Maybe StackConfigurationManager

usCustomCookbooksSource :: Maybe Source

usCustomJson :: Maybe Text

usDefaultAvailabilityZone :: Maybe Text

usDefaultInstanceProfileArn :: Maybe Text

usDefaultOs :: Maybe Text

usDefaultRootDeviceType :: Maybe RootDeviceType

usDefaultSshKeyName :: Maybe Text

usDefaultSubnetId :: Maybe Text

usHostnameTheme :: Maybe Text

usName :: Maybe Text

usServiceRoleArn :: Maybe Text

usStackId :: Text

usUseCustomCookbooks :: Maybe Bool

usUseOpsworksSecurityGroups :: Maybe Bool

Request lenses

usAttributes :: Lens' UpdateStack (HashMap StackAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

usChefConfiguration :: Lens' UpdateStack (Maybe ChefConfiguration)

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version on Chef 11.10 stacks. For more information, see Create aNew Stack.

usConfigurationManager :: Lens' UpdateStack (Maybe StackConfigurationManager)

The configuration manager. When you clone a stack we recommend that you use the configuration manager to specify the Chef version, 0.9, 11.4, or 11.10. The default value is currently 11.4.

usCustomCookbooksSource :: Lens' UpdateStack (Maybe Source)

usCustomJson :: Lens' UpdateStack (Maybe Text)

A string that contains user-defined, custom JSON. It is used to override the corresponding default stack configuration JSON values. The string should be in the following format and must escape characters such as '"'.:

'"{"key1": "value1", "key2": "value2",...}"'

For more information on custom JSON, see Use Custom JSON to Modify the StackConfiguration JSON.

usDefaultAvailabilityZone :: Lens' UpdateStack (Maybe Text)

The stack's default Availability Zone, which must be in the specified region. For more information, see Regions and Endpoints. If you also specify a value for DefaultSubnetId, the subnet must be in the same zone. For more information, see CreateStack.

usDefaultInstanceProfileArn :: Lens' UpdateStack (Maybe Text)

The ARN of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

usDefaultOs :: Lens' UpdateStack (Maybe Text)

The stack's operating system, which must be set to one of the following.

Standard operating systems: an Amazon Linux version such as 'Amazon Linux2014.09', 'Ubuntu 12.04 LTS', or 'Ubuntu 14.04 LTS'. Custom AMIs: Custom. You specify the custom AMI you want to use when you create instances. The default option is the current Amazon Linux version.

usDefaultRootDeviceType :: Lens' UpdateStack (Maybe RootDeviceType)

The default root device type. This value is used by default for all instances in the stack, but you can override it when you create an instance. For more information, see Storage for the Root Device.

usDefaultSshKeyName :: Lens' UpdateStack (Maybe Text)

A default Amazon EC2 key pair name. The default value is none. If you specify a key pair name, AWS OpsWorks installs the public key on the instance and you can use the private key with an SSH client to log in to the instance. For more information, see Using SSH to Communicate with an Instance and Managing SSH Access. You can override this setting by specifying a different key pair, or no key pair, when you create an instance.

usDefaultSubnetId :: Lens' UpdateStack (Maybe Text)

The stack's default VPC subnet ID. This parameter is required if you specify a value for the VpcId parameter. All instances are launched into this subnet unless you specify otherwise when you create the instance. If you also specify a value for DefaultAvailabilityZone, the subnet must be in that zone. For information on default values and when this parameter is required, see the VpcId parameter description.

usHostnameTheme :: Lens' UpdateStack (Maybe Text)

The stack's new host name theme, with spaces are replaced by underscores. The theme is used to generate host names for the stack's instances. By default, HostnameTheme is set to Layer_Dependent, which creates host names by appending integers to the layer's short name. The other themes are:

Baked_Goods Clouds Europe_Cities Fruits Greek_Deities Legendary_creatures_from_Japan Planets_and_Moons Roman_Deities Scottish_Islands US_Cities Wild_Cats To obtain a generated host name, call GetHostNameSuggestion, which returns a host name based on the current theme.

usName :: Lens' UpdateStack (Maybe Text)

The stack's new name.

usServiceRoleArn :: Lens' UpdateStack (Maybe Text)

The stack AWS Identity and Access Management (IAM) role, which allows AWS OpsWorks to work with AWS resources on your behalf. You must set this parameter to the Amazon Resource Name (ARN) for an existing IAM role. For more information about IAM ARNs, see Using Identifiers.

You must set this parameter to a valid service role ARN or the action will fail; there is no default value. You can specify the stack's current service role ARN, if you prefer, but you must do so explicitly.

usStackId :: Lens' UpdateStack Text

The stack ID.

usUseCustomCookbooks :: Lens' UpdateStack (Maybe Bool)

Whether the stack uses custom cookbooks.

usUseOpsworksSecurityGroups :: Lens' UpdateStack (Maybe Bool)

Whether to associate the AWS OpsWorks built-in security groups with the stack's layers.

AWS OpsWorks provides a standard set of built-in security groups, one for each layer, which are associated with layers by default. UseOpsworksSecurityGroups allows you to instead provide your own custom security groups. UseOpsworksSecurityGroups has the following settings:

True - AWS OpsWorks automatically associates the appropriate built-in security group with each layer (default setting). You can associate additional security groups with a layer after you create it but you cannot delete the built-in security group. False - AWS OpsWorks does not associate built-in security groups with layers. You must create appropriate EC2 security groups and associate a security group with each layer that you create. However, you can still manually associate a built-in security group with a layer on creation; custom security groups are required only for those layers that need custom settings. For more information, see Create a NewStack.

Response

data UpdateStackResponse

Response constructor

updateStackResponse :: UpdateStackResponse

UpdateStackResponse constructor.