Network.AWS.OpsWorks.V2013_02_18.CreateLayer

Description

Creates a layer. For more information, see How to Create a Layer. You should use CreateLayer for noncustom layer types such as PHP App Server only if the stack does not have an existing layer of that type. A stack can have at most one instance of each noncustom layer; if you attempt to create a second instance, CreateLayer fails. A stack can have an arbitrary number of custom layers, so you can call CreateLayer as many times as you like for that layer type. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data CreateLayer

Request constructor

mkCreateLayer

Arguments

:: Text
clStackId

-> LayerType
clType

-> Text
clName

-> Text
clShortname

-> CreateLayer

Smart constructor for the minimum required parameters to construct a valid CreateLayer request.

Request lenses

clStackId :: Lens' CreateLayer Text

The layer stack ID.

clType :: Lens' CreateLayer LayerType

The layer type. A stack cannot have more than one built-in layer of the same type. It can have any number of custom layers. This parameter must be set to one of the following: custom: A custom layer db-master: A MySQL layer java-app: A Java App Server layer rails-app: A Rails App Server layer lb: An HAProxy layer memcached: A Memcached layer monitoring-master: A Ganglia layer nodejs-app: A Node.js App Server layer php-app: A PHP App Server layer web: A Static Web Server layer.

clName :: Lens' CreateLayer Text

The layer name, which is used by the console.

clShortname :: Lens' CreateLayer Text

The layer short name, which is used internally by AWS OpsWorks and by Chef recipes. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters, which are limited to the alphanumeric characters, -, '_', and ..

clAttributes :: Lens' CreateLayer (Map LayerAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

clCustomInstanceProfileArn :: Lens' CreateLayer (Maybe Text)

The ARN of an IAM profile that to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

clCustomSecurityGroupIds :: Lens' CreateLayer [Text]

An array containing the layer custom security group IDs.

clPackages :: Lens' CreateLayer [Text]

An array of Package objects that describe the layer packages.

clVolumeConfigurations :: Lens' CreateLayer [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

clEnableAutoHealing :: Lens' CreateLayer (Maybe Bool)

Whether to disable auto healing for the layer.

clAutoAssignElasticIps :: Lens' CreateLayer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

clAutoAssignPublicIps :: Lens' CreateLayer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

clCustomRecipes :: Lens' CreateLayer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer custom recipes.

clInstallUpdatesOnBoot :: Lens' CreateLayer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances. We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

clUseEbsOptimizedInstances :: Lens' CreateLayer (Maybe Bool)

Whether to use Amazon EBS-optimized instances.

Response

data CreateLayerResponse

Contains the response to a CreateLayer request.

Response constructor

mkCreateLayerResponse :: CreateLayerResponse

Smart constructor for the minimum required parameters to construct a valid CreateLayerResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

clrLayerId :: Lens' CreateLayerResponse (Maybe Text)

The layer ID.