Network.AWS.OpsWorks.V2013_02_18.CreateUserProfile

Description

Creates a new user profile. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data CreateUserProfile

Request constructor

mkCreateUserProfile

Arguments

:: Text
cupIamUserArn

-> CreateUserProfile

Smart constructor for the minimum required parameters to construct a valid CreateUserProfile request.

Request lenses

cupIamUserArn :: Lens' CreateUserProfile Text

The user's IAM ARN.

cupSshUsername :: Lens' CreateUserProfile (Maybe Text)

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], -, and '_'. If the specified name includes other punctuation marks, AWS OpsWorks removes them. For example, my.name will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks generates one from the IAM user name.

cupSshPublicKey :: Lens' CreateUserProfile (Maybe Text)

The user's public SSH key.

cupAllowSelfManagement :: Lens' CreateUserProfile (Maybe Bool)

Whether users can specify their own SSH public key through the My Settings page. For more information, see Setting an IAM User's Public SSH Key.

Response

data CreateUserProfileResponse

Contains the response to a CreateUserProfile request.

Response constructor

mkCreateUserProfileResponse :: CreateUserProfileResponse

Smart constructor for the minimum required parameters to construct a valid CreateUserProfileResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

cuprIamUserArn :: Lens' CreateUserProfileResponse (Maybe Text)

The user's IAM ARN.