Network.AWS.OpsWorks.V2013_02_18.DescribeApps

Description

Requests a description of a specified set of apps. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeApps :: DescribeApps

Minimum specification for a DescribeApps request.

data DescribeApps

Constructors

DescribeApps

Fields

_dasStackId :: Maybe Text

The app stack ID. If you use this parameter, DescribeApps returns a description of the apps in the specified stack.

_dasAppIds :: [Text]

An array of app IDs for the apps to be described. If you use this parameter, DescribeApps returns a description of the specified apps. Otherwise, it returns a description of every app.

dasStackId :: Lens' DescribeApps (Maybe Text)

dasAppIds :: Lens' DescribeApps [Text]

data DescribeAppsResponse

Constructors

DescribeAppsResponse

Fields

_datApps :: [App]

An array of App objects that describe the specified apps.

datApps :: Iso' DescribeAppsResponse [App]