Network.AWS.OpsWorks.V2013_02_18.DescribePermissions

Description

Describes the permissions for a specified stack. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describePermissions :: DescribePermissions

Minimum specification for a DescribePermissions request.

data DescribePermissions

Constructors

DescribePermissions

Fields

_dprIamUserArn :: Maybe Text

The user's IAM ARN. For more information about IAM ARNs, see Using Identifiers.

_dprStackId :: Maybe Text

The stack ID.

dprStackId :: Lens' DescribePermissions (Maybe Text)

dprIamUserArn :: Lens' DescribePermissions (Maybe Text)

data DescribePermissionsResponse

Constructors

DescribePermissionsResponse

Fields

_dpsPermissions :: [Permission]

An array of Permission objects that describe the stack permissions. If the request object contains only a stack ID, the array contains a Permission object with permissions for each of the stack IAM ARNs. If the request object contains only an IAM ARN, the array contains a Permission object with permissions for each of the user's stack IDs. If the request contains a stack ID and an IAM ARN, the array contains a single Permission object with permissions for the specified stack and IAM ARN.

dpsPermissions :: Iso' DescribePermissionsResponse [Permission]