Network.AWS.OpsWorks.V2013_02_18.DescribeUserProfiles

Description

Describe specified users. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data DescribeUserProfiles

Request constructor

mkDescribeUserProfiles :: DescribeUserProfiles

Smart constructor for the minimum required parameters to construct a valid DescribeUserProfiles request.

Request lenses

dup1IamUserArns :: Lens' DescribeUserProfiles [Text]

An array of IAM user ARNs that identify the users to be described.

Response

data DescribeUserProfilesResponse

Contains the response to a DescribeUserProfiles request.

Response constructor

mkDescribeUserProfilesResponse :: DescribeUserProfilesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeUserProfilesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

duprUserProfiles :: Lens' DescribeUserProfilesResponse [UserProfile]

A Users object that describes the specified users.