Network.AWS.OpsWorks.V2013_02_18.RegisterElasticIp

Description

Registers an Elastic IP address with a specified stack. An address can be registered with only one stack at a time. If the address is already registered, you must first deregister it by calling DeregisterElasticIp. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data RegisterElasticIp

Request constructor

mkRegisterElasticIp

Arguments

:: Text
reiElasticIp

-> Text
reiStackId

-> RegisterElasticIp

Smart constructor for the minimum required parameters to construct a valid RegisterElasticIp request.

The fields accessible through corresponding lenses are:

ElasticIp :: Text

StackId :: Text

Request lenses

reiElasticIp :: Lens' RegisterElasticIp Text

The Elastic IP address.

reiStackId :: Lens' RegisterElasticIp Text

The stack ID.

Response

data RegisterElasticIpResponse

Contains the response to a RegisterElasticIp request.

Response constructor

mkRegisterElasticIpResponse :: RegisterElasticIpResponse

Smart constructor for the minimum required parameters to construct a valid RegisterElasticIpResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ElasticIp :: Maybe Text

Response lenses

reirElasticIp :: Lens' RegisterElasticIpResponse (Maybe Text)

The Elastic IP address.