Network.AWS.OpsWorks.V2013_02_18.RegisterVolume

Description

Registers an Amazon EBS volume with a specified stack. A volume can be registered with only one stack at a time. If the volume is already registered, you must first deregister it by calling DeregisterVolume. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data RegisterVolume

Request constructor

mkRegisterVolume

Arguments

:: Text
rvStackId

-> RegisterVolume

Smart constructor for the minimum required parameters to construct a valid RegisterVolume request.

The fields accessible through corresponding lenses are:

Ec2VolumeId :: Maybe Text

StackId :: Text

Request lenses

rvEc2VolumeId :: Lens' RegisterVolume (Maybe Text)

The Amazon EBS volume ID.

rvStackId :: Lens' RegisterVolume Text

The stack ID.

Response

data RegisterVolumeResponse

Contains the response to a RegisterVolume request.

Response constructor

mkRegisterVolumeResponse :: RegisterVolumeResponse

Smart constructor for the minimum required parameters to construct a valid RegisterVolumeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VolumeId :: Maybe Text

Response lenses

rvrVolumeId :: Lens' RegisterVolumeResponse (Maybe Text)

The volume ID.