Network.AWS.OpsWorks.V2013_02_18.UpdateUserProfile

Description

Updates a specified user profile. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Request

data UpdateUserProfile

Request constructor

mkUpdateUserProfile

Arguments

:: Text
uupIamUserArn

-> UpdateUserProfile

Smart constructor for the minimum required parameters to construct a valid UpdateUserProfile request.

Request lenses

uupIamUserArn :: Lens' UpdateUserProfile Text

The user IAM ARN.

uupSshUsername :: Lens' UpdateUserProfile (Maybe Text)

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], -, and '_'. If the specified name includes other punctuation marks, AWS OpsWorks removes them. For example, my.name will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks generates one from the IAM user name.

uupSshPublicKey :: Lens' UpdateUserProfile (Maybe Text)

The user's new SSH public key.

uupAllowSelfManagement :: Lens' UpdateUserProfile (Maybe Bool)

Whether users can specify their own SSH public key through the My Settings page. For more information, see Managing User Permissions.

Response

data UpdateUserProfileResponse

Response constructor

mkUpdateUserProfileResponse :: UpdateUserProfileResponse

Smart constructor for the minimum required parameters to construct a valid UpdateUserProfileResponse response.

This constructor is provided for convenience and testing purposes.